spend less time in progress bar in gradebook
the gradebook page has a progress bar which currently
waits until all data for the gradebook is loaded before
hiding the progress bar and showing the datagrid.
Seems like it makes more sense to show the gradebook
sooner and just populate it as data comes in so at
least it's partly usable.  This is a minor change right
now, but it'll make it possible to do more lazy
loading of the gradebook in the future.

fixes #3955

Change-Id: Ie4f486bdc0dd1373d4119822ec706fcdbd329b2f
Reviewed-on: https://gerrit.instructure.com/2501
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Ryan Shaw <ryan@instructure.com>