context_message_participants getting recreated
on every save all the participants get duplicated, which
is causing slow loads on messages with lots of recipients.

this commit backgrounds the create process in case
there are a lot of participants to add, and stops using
find_or_create which for some reason was failing to find
correctly.

fixes #4014

Change-Id: I3f03793994999faf2bb424581f45adeb328eb28d
Reviewed-on: https://gerrit.instructure.com/2602
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Brian Palmer <brianp@instructure.com>