finishing up respondus quiz upload support, refs #3760
I removed the assignment group listing. We don't assign a quiz to an
assignment group until the quiz is published (creating the Assignment for the
quiz), so it didn't make sense to have Respondus users select the group.

Note that attachments in the QTI uploads don't yet work. That will
happen in QTI improvements that are coming soon.

Change-Id: I2e7fd29cccbc569d413124e3d2a9a417c979a435
Reviewed-on: https://gerrit.instructure.com/2674
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Zach Wily <zach@instructure.com>