migration process for moving to new kaltura instance
This commit makes it possible to migrate media collections
from one Kaltura instance to another.  It adds an
old_media_id column to media_objects that can be used
as a fallback.  It also adds a mechanism for creating bulk
upload csv's to migrate media objects from the old
system to the new, while catching and preserving the
old id so that any links inside of content will still work
even if they have the old id.

refs #4083

Change-Id: Ifa213fd348a50ebd7e7ac7068d8eff40b23747d1
Reviewed-on: https://gerrit.instructure.com/2902
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: JT Olds <jt@instructure.com>