don't force tables to have grey borders
we had a style that would give an unstyled table
in .user_content grey borders around the table and
on every td.

that is not what default user-agent stylesheets
do, so when pages were imported from other systems,
tables that had no borders now have grey borders
in every td.

Change-Id: I3362fadcf5aaf00a0e5c0fa0e7eacb8ac1e98154
Reviewed-on: https://gerrit.instructure.com/2952
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Brian Whitmer <brian@instructure.com>