submission web snapshots with local storage, fixes #4531
the attachment has its context set to the submission, which we don't
handle as a context anywhere currently. Rather than try to shoe-horn the
submission into a context role, we allow using the /files urls without
any context at all (we still check permissions on the attachment of
course). this fixes some safefiles issues as well.

Change-Id: Iec12b02f16f01f795f70a6f4ed8a152b7661e000
Reviewed-on: https://gerrit.instructure.com/3557
Tested-by: Hudson <hudson@instructure.com>
Tested-by: Selenium <selenium@instructure.com>
Reviewed-by: Zach Wily <zach@instructure.com>