don't send cache headers when showing files; fixes #4654
If IE sees any no-cache headers in a secure request, it won't be able to
download files that would be opened in a 3rd-party application. This is
because it will refuse to store the file in a temporary location (because of
the no-cache designation). For more information, see:

http://support.microsoft.com/default.aspx?scid=KB;EN-US;q316431&

We will just not send any cache headers for the file show action.

Change-Id: Id02b9bab9c6fe46b4e8f3d9198f8bb87d4b697f8
Reviewed-on: https://gerrit.instructure.com/3845
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Brian Palmer <brianp@instructure.com>