make search box and drop down consistent in copy_course
 * updated implementation of Course.manageable_by_user
 * added method User#manageable_courses
 * renamed some "courses" methods and routes to "manageable_courses" to
   reflect the reality of the query
 * changed population of @possible_courses in
   ContentImportsController#copy_course to use User#manageable_courses
   like UsersController#courses was (now #manageable_courses)
 * fixed the "no courses" condition in copy_course.html.erb to look at
   @possible_courses for consistency

Caveat: @possible_courses in #copy_course was previously explicit about
trying to pull in courses cross-listed against the sections of the
target course... but only if they were manageable by the user. That is
no longer explicit in the new implementation, but they should still be
included implicitly, as well as other courses manageable by the user
that wouldn't have been listed before; I believe this is an enhancement,
not a regression.

Fixes #4735.

Change-Id: I38cf3e0a1b2478ad32243a0bb0d73c796fc775e0
Reviewed-on: https://gerrit.instructure.com/4117
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Jon Jensen <jon@instructure.com>