column reordering/resizing/sorting in gradebook2
if you minimize a column to it's smallest size
it will change it's formatting and show a tooltip
on hovering.  it will also remember you hid it
(per browser/user) and not show it in future.

click a column header to sort.

drag/drop assignment column headers to reorder
(will not be remembered for now)

Change-Id: Ia9d1d8e1bb82ebde85d87fedae80c1d7feb46b6f
Reviewed-on: https://gerrit.instructure.com/4642
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Jon Jensen <jon@instructure.com>