optimize FoldersController#show
Don't make thumbnail_url automatically included with every
serialization of Attachment, since it requires loading Thumbnail, and
Context.  Instead specifically include it where it is actually needed.

Then be very specific about what fields show up in
FoldersController#show, since we're returning a bunch of items, this
improves both render time, transfer time, and is also less data for
the javascript to manage.

Also fix FoldersController#index.  It obviously wasn't used, since it
was very broken.  I just fixed it because it was the easiest way for
me to find a folder id for a course to test #show.

With this and prior optimization, FoldersController#show for a a
folder with ~3000 items went from 10 minutes (!) to < 10 seconds.
Not perfect, but a huge improvemnt, and that many items in a single
folder is definitely an outlier.  "Normal" folders are now quite
snappy.

Refs #3975

Change-Id: I2952ba886e2bc8ac7d8b230105caa508084841f1
Reviewed-on: https://gerrit.instructure.com/4646
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Brian Palmer <brianp@instructure.com>