hack around periodic disappearing clicks in selenium
It looks like clicks on # links right after a page load are getting lost
periodically. I suspect it's due to the earlyClick system, but haven't been
able to verify. When I add logging to debug, the problem does not manifest
itself. This is a hack until we can figure out what's really happening.

Change-Id: I7b783fab32004982664eef78138058517c140bb9
Reviewed-on: https://gerrit.instructure.com/4782
Reviewed-by: Brian Palmer <brianp@instructure.com>
Tested-by: Hudson <hudson@instructure.com>
Tested-by: Selenium <selenium@instructure.com>