create a unique temp directory for each job, underneath the master tmpdir
This way it's easy to tell what job created a file, and it makes cleanup
more reliable, since Dir.mktmpdir will remove the whole directory when
the job ends and the block exits.

Closes #5093

I was a bit worried about the performance impact of making a new
directory for every single job, but on an active job server of ours, it
only seems to add about 160 microseconds per job.

Change-Id: Id4318da20cac75e6ba232ec2250b78c38485f9e4
Reviewed-on: https://gerrit.instructure.com/4811
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Cody Cutrer <cody@instructure.com>