fix thumbnail url for attachments that have a root_attachment_id
thumbnails are shared with the root attachment, so re-assigning the
thumbnail's parent to the child attachment is changing the s3 url to
have the wrong attachment id in it.

(again, temporary fix because this code is gone on master. but i'm looking
into whether we have this problem on master as well.)

Change-Id: I5830df8a2592e9e95963cc3a9a6ecf6cae9ec14f
Reviewed-on: https://gerrit.instructure.com/4823
Reviewed-by: Cody Cutrer <cody@instructure.com>
Tested-by: Hudson <hudson@instructure.com>