fix pageless plugin with scrape method
When including a scrape function to pageless, the data after scraping was
being dumped into the DOM incorrectly. This fixes that, and includes a test
for the announcements page.

(The bug was originally seen on the inbox, but rather than write a spec for
code that won't last much longer, I wrote the test for somewhere else.)

Change-Id: If4ca0cbed89dcf6e6e24e814f75824266422cbfa
Reviewed-on: https://gerrit.instructure.com/4992
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Brian Palmer <brianp@instructure.com>
Tested-by: Selenium <selenium@instructure.com>