conversations teaser
Ribbon shows up on /inbox.  You can click to dismiss the note,
it uses local storage to remember if it should be collapsed.
Also checks for a Setting value on 'conversations_rollout_date'
for the date of the deploy, and falls back to Aug 20 (so
we can share this with uen as well)

Change-Id: I4a44467548eef8611239be2938cf7d496178e89f