always immediately update enrollments when the restrictions change
Previously we were trying to be too smart about when we'd do the updates, and
we were missing some cases - like when a start_date is moved to the future.
Now we'll just run the update any time the dates change, which should catch
those edge cases. The update was also put in a strand to avoid any race
conditions caused by someone doing a lot of close-together updates of those
dates.

Change-Id: I655ac530ee04725ff1aad2118b7a35649759571f
Reviewed-on: https://gerrit.instructure.com/5118
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Brian Whitmer <brian@instructure.com>