external tool resource selector in WYSIWYG editor
Configured external tools can show up as buttons in the
rich editor. If you click the button then it'll pop up
a dialog that uses basic lti to load a page for the
external tool.  The user can pick a resource in that
tool and the tool will then redirect back to a Canvas
url that will fire a callback and embed the content
into the editor.

See spec/selenium/external_tool_buttons_sel.rb for an
example of how to manually configure one of these tools.
Adding configuration to the UI will come in another
commit.

testing notes:
to configure an external tool to show up in the editor:
tool.settings = {
  :editor_button => {
    :url => "http://<canvas_domain>/selection_test",
    :text => <button_label>,
    :selection_width => <width_of_iframe>,
    :selection_height => <height_of_iframe>,
    :icon_url => <url_of_button_icon>
  }
}

test plan:
- configure an external tool in the course with an editor button
- check to see if the button shows up in the rich editor in that course

- configure an external tool in the account with an editor button
- check to see if the button shows up for courses in that account

- configure more than 3 external tools
- check to see if the "more tools" dropdown works correctly

Change-Id: I681db0af578df6a9c7a2c840d293703937d81c46
Reviewed-on: https://gerrit.instructure.com/5429
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Bracken Mosbacker <bracken@instructure.com>