display "no messages" message on empty filter. fixes #5493
Message that displays when a filter matches no messages only
displays when there are no <li> elements in the $conversation_list
element. The ajax spinner is stored in an <li> though, so
this condition is never met.

Fix is to add a :not filter on the spinner <li>'s ID to the
test.

Change-Id: I97cd0d9a7e707117489be837d8e3c449e76db153
Reviewed-on: https://gerrit.instructure.com/5656
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Jon Jensen <jon@instructure.com>