change identity header cache key when masquerading. fixes #5995
User and admin masquerading were sharing the same cache key for
the identity fragment in the header, causing the "[Stop
Masquerading]" link to appear for users who were logged in while
the admin was masquerading.

Now the identity fragment key includes an "a" or "m" to indicate
if the user is actual or masquerading. This prevents an actual
user and a masquerading admin from sharing the same fragment.

Change-Id: I7d649ac57315b0fa8f80c3b41f4119e9efd04fd7
Reviewed-on: https://gerrit.instructure.com/6321
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Jon Jensen <jon@instructure.com>