remove the new wiki page notification
closes #4692

This email was sent as soon as the record was created, which wasn't at
all helpful since at that point it has no content. It also caused issues
with teachers who wanted to create hidden wiki pages, since it is sent
before the teacher has a chance to hide the page.

updated wiki page notifications will still be sent out.

Change-Id: Ie6762eafd67d683dbf942a5fb5c443dd1b2288e6
Reviewed-on: https://gerrit.instructure.com/6786
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Cody Cutrer <cody@instructure.com>