fix error in speed_grader selenium spec
When using the remote selenium driver, element.click evaluates to false on
success. That means keep_trying_until { something.click } doesn't work, even
though it will locally. You need to either wait until the element is
displayed, or return true in your block (expecting that an exception will be
raised before then if the condition is not met.)

test plan: n/a

Change-Id: I4222cb9e03e78673d20690737421900822f9a2c8
Reviewed-on: https://gerrit.instructure.com/7269
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Brian Palmer <brianp@instructure.com>