fix duplicated notification policies showing for new accounts; fixes #6247
The duplicate notification policies were being generated in ProfileController
when a user didn't have any policies. It would build a policy through the
User#notification_policies relation, then append it to the same array that
backs that relation. So the policy would show up twice.

The selenium test was failing because we were displaying duplicate policies to users
on postgres systems. Mysql test instances were not seeing the duplicates because we
were calling notification_policy_model, which stupidly creates a
NotificationPolicy for user id 1. On mysql databases, that usually was the id
of the user being tested, but not on postgres. Since user id 1 would have a
policy, the duplicate was not triggered. In the test we were assuming the duplicates
were right and testing with that assumption.

test plan: Create a new user on a postgres system. Go to communication
preferences, and verify that the email address doesn't show up doubled for
every notification type.

Change-Id: I911d400eced78304ffb1aed61a5618f2facb9b07
Reviewed-on: https://gerrit.instructure.com/7342
Reviewed-by: Brian Palmer <brianp@instructure.com>
Tested-by: Hudson <hudson@instructure.com>