don't use the alternate CAS login url during api oauth flow
Otherwise the alternate login screen may not get the user back to the
oauth2_token page in a state that the application recognizes.

test plan: Execute the oauth flow against an CAS account with a log_in_url
configured (e.g., with the iOS app). The flow should not redirect to the
alternate url, instead using the primary CAS url.  Also confirm that
normal non-oauth logins still use the alternate url.

Change-Id: Iba39b719a5f80727801880660d19afa128b898ce
Reviewed-on: https://gerrit.instructure.com/7368
Reviewed-by: Zach Wily <zach@instructure.com>
Tested-by: Hudson <hudson@instructure.com>