update total and group percentages on cell update
when you edit a grade cell in gradebook2, this change
will make it so the assignment group percentage
and the total grade percentage will also update.
for now it waits till ajax returns with the 'score'
property updated (the UI updates the 'grade' property)
like gradebook1 does. would be nice to have it be
real-time (and not wait for ajax to return), but
would need to reproduce grade_to_score functionality
in js.

test plan:
 * in gradebook2 update a grade cell
 * wait a sec for the server to respond
 * you should see the percentages for the assignment
   group and final score update to reflect the grade
   you typed
 * verify that that percentage is the same as
   gradebook1 shows
 * refresh page, make sure it is the same.

Change-Id: I174d48ffebaa264a013c4ff4be3d9a52b4586fec
Reviewed-on: https://gerrit.instructure.com/7439
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Jon Jensen <jon@instructure.com>