use defined discussion formats in assignment and submission apis
fixes #6560

The assignment and submission apis were written before the discussion
api, and the response formats diverged. This merges the two formats so
that all apis return the same data structures.

Also, removed the attachments relation from DiscussionEntry, it's cruft.
We have to continue returning the attachments array in the json for now,
for backwards compatibility, since it includes the .attachment field as
well.

test plan: API calls to read discussion assignments and their
submissions should now include the full json response data, see the api
documentation for the discussions api.

Change-Id: Ib394d7e5c3fcf261adb8663470c6822fe2476b03
Reviewed-on: https://gerrit.instructure.com/7395
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Bracken Mosbacker <bracken@instructure.com>