don't delete enrollments for students with multiple in the same course
refs #6440

we know for a fact that we don't fully support multiple student
enrollments in different sections of a course, but deleting duplicate
enrollments on gradebook load is a solution that is causing more trouble
than the original problem, especially because it's such a random event.

this exposes a known issue, where if you filter the gradebook by
section, the student will appear in one of their sections but not the
other.

test plan: enroll a student in multiple sections, either through SIS or
the console. going to the gradebook should not delete any of the
enrollments, but downloading the csv should still only list the student
once. and re-uploading the csv with a changed grade for that student
should work. you'll see other weirdness, like the student will have the
course listed twice in their dropdown. these are known issues discussed
in the the ticket.

Change-Id: I78c3ed513ce3986e90c65d51562bbc4b24439661
Reviewed-on: https://gerrit.instructure.com/7540
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Zach Wily <zach@instructure.com>