replace usage of Date.today
Date.today does not accomodate Time.zone. This is a problem when the
time at the time of the call is between Time.zone's midnight and the
system time zone's midnight; Date.today will think it's a day
earlier/later (depending on direction of offset) than Time.zone.today in
that case.

Change-Id: I54083d26dcf704ac00d812597ee6aec95ffa0685
Reviewed-on: https://gerrit.instructure.com/7584
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Cody Cutrer <cody@instructure.com>