fix truncate_html for nokogiri 1.5.0
A bug in the upgraded nokogiri library was causing nil access errors
when trying to truncate html.

See https://github.com/tenderlove/nokogiri/issues/595

test plan: Send a discussion topic message to the dashboard that is more
than 50 words long. verify that it is truncated with ..., and no error
is generated.

Change-Id: I35523d125e50beac82ed8cf7214a64d6ce0d7001
Reviewed-on: https://gerrit.instructure.com/7989
Reviewed-by: Cody Cutrer <cody@instructure.com>
Tested-by: Hudson <hudson@instructure.com>