unpublished quiz preview fixes #6528
previewing the quiz is posted, and then redirected
with a get request but the preview parameter was
not passed through to the redirect, now it is.

test plan:
1. start a new quiz
2. add and save a question
3. click 'preview quiz'

you ought to be able to preview the quiz

also fixed the spec to check for the presence of
the questions container, rather than the title
since the title is present whether you're able to
preview the quiz or not.

Change-Id: I87a3a179ed79f464b7c765c9f23ba48226bb86ca
Reviewed-on: https://gerrit.instructure.com/8099
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Ryan Florence <ryanf@instructure.com>