fix include_ungraded_assignments behavior of gradebook2
When I made gradebook2 I thought "include ungraded
assignments" was to show/hide "not-graded" assignments
it is actually there to adjust how final scores are
displayed. So that was bug 1

bug 2 was that toggling that setting did not
actually update final grade percentages.

also, reword the setting in the header to say:
"Treat ungraded as 0's"
instead of:
"Include Ungraded Assignments"
because it is less confusing

test plan

 * create an assignment of type 'not-graded'
 * make sure it does not show up in gradebook2
 * in gradebook2 click the "Treat ungraded as 0's"
   option in the cog menu in the header
 * make sure that total columns change to reflect
   treating ungraded assignments as a score of 0

Change-Id: I276d02dfcf13acafd49ce2697a8f4b4bca6bef64
fixes: #6765
Reviewed-on: https://gerrit.instructure.com/8107
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Jake Sorce <jake@instructure.com>
Tested-by: Jake Sorce <jake@instructure.com>