handle zero point multiple choice questions in qti
if a multiple choice question is worth zero points, some
qti vendors label increment the score by zero which doesn't
allow us to determine the correct answer. This commit special
cases this problem to grab the answer that was set instead
if added to. (that is the convention used by the vendor)

Test Plan:
 * Import the .zip fixture that was updated
 * The First multiple choice question should have a correct answer

refs #7219

Change-Id: If5b64d21dea509a49d4635b68242b31633b2ac6d
Reviewed-on: https://gerrit.instructure.com/8610
Reviewed-by: Brian Palmer <brianp@instructure.com>
Tested-by: Hudson <hudson@instructure.com>