don't fail when copying/import invalid urls
the urls for content tags didn't use to be validated so there
can be invalid urls in olds data. If you try to copy one of
those courses it fails. It also failed for importing content
packages.

This commit makes it so that importing will skip those items
and add an item about it to the migration warning. For course
copy it attempts to escape the url and continue copying.

Test Plan:
 * try to import a course from a package that has an invalid url
 * try to copy a course that has an invalid url. (You'll have to manually set the invalid url on the ContentTag)

closes #7262

Change-Id: I008b786919bb2b66e0ff546667e7c0832b52e192
Reviewed-on: https://gerrit.instructure.com/8710
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Brian Palmer <brianp@instructure.com>