return description in json for locked assignments. fixes #5853.
when updating an assignment, assignment.to_json wasn't including
the description for any users if the assignment was locked. this
was causing tinymce to error (because it expected at least a blank
string), and prevented the load spinner from disappearing and the
textarea from properly updating when editing an assignment.

test plan:
  * create a new assignment;
  * set a Locked Until date;
  * change the assignment description;
  * update the assignment;
  * verify that the load spinner disappears;
  * verify that the assignment description properly updates;
  * edit the newly created assignment;
  * verify that the description appears properly in the tinymce
    editor.

Change-Id: I7c0a3e944ba3c8fb19bc49b8d63f436621f8a9ff
Reviewed-on: https://gerrit.instructure.com/8818
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Zach Wily <zach@instructure.com>