escape filenames more aggressively for content-disposition header; fixes #7299
Chrome was getting upset by commas not being escaped in the UTF-8 portion of
the content-disposition header for S3 urls. It saw the comma and thought it
was ending the header, and complained of duplicate headers and a header
splitting attack. This will escape all non-[A-Za-z0-9.] characters, which
should be aggressive enough. This was tested in IE, Safari, Chrome, and
Firefox and appears to not cause problems.

test plan:
 * Enable S3 for your dev environment
 * Upload a file with a comma in the name
 * Ensure that you can download that file using at least Chrome 17.

Change-Id: Id229dedc9acc78bdbfe74ddc001fc31a305af0ea
Reviewed-on: https://gerrit.instructure.com/8833
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Brian Palmer <brianp@instructure.com>