introduce preferred user list search mode
closes #7059, #7411

preferred is a compromise between open registration and closed
registration. if a single user is found, that user is found without
introducing any temporary user to allow user disambiguation.  if no
users are found, or multiple users are found, it's the same behavior
as open registration.

as part of this, the "only search existing users" checkbox has been
removed. instead, preferred searching will be used in the following
cases:
 * adding an admin, and open registration is enabled
 * adding an admin, open registration is disabled, and the user
   has permission to create new users
 * adding a user to a course, open registration is enabled, and
   delegated authentication is in use
 * adding a user to a course, open registration is disabled, and
   the user has permission to create new users

the notable case that preferred searching is not used, and the
user would want to use search existing users is open registration
enabled, and adding a user to a course.  in this case, a temporary
user will be created instead of sending the invitation to the
existing user.  however, the end user experience is identical
(invitation sent to e-mail, invitation visible in the same locations
in the UI), with the small exception that the end user gets the
opportunity to create a separate account if desired (but not
preferred).

test plan:
 * go through each of the cases above for a user that doesn't exist
   yet, a single matching user exists, or multiple users exist.
   it should behave as described above

Change-Id: Idbc7fe23bfc7430b4cd25f7981f5dc08b9e4ffda
Reviewed-on: https://gerrit.instructure.com/8966
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Brian Palmer <brianp@instructure.com>