add a hidden_dialog helper to handle dialogs that we only want one of on a page
There are several places where we want to add a hidden <div> that will later
be used as a dialog. Normally that div would have an id, and would be
displayed using the id. Where to place the div is a problem in cases where the
dialog might be invoked from several different places - we only want one and
only want copy of the div on the page.

This helper will let you specify content for a hidden div and an id, and will
only place one instance on the page. That means you can specify the div close
to a help button that might invoke it, and you don't have to worry about if
it's added twice.

The helper will raise a warning if you try to add a div with an id matching an
existing one, but with different content.

The other change is fixing some bad indentation in _full_assignment.html.erb
which led to somebody adding an extra </div> while trying to add another
instance of a dialog div.

Change-Id: If91428db89f92f890ea5917028331040d12741a6
Reviewed-on: https://gerrit.instructure.com/9138
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Brian Palmer <brianp@instructure.com>