support for deferred delayed job creation, refs #7277
adds Delayed::Batch.serial_batch which, when given a block, will delay
the creation of (most) new jobs triggered within that block. a single
job will be created (per priority and queue) containing all the info for
those jobs. when it runs, it will create those jobs and run them
synchronously. if any of those jobs fail, they will be handled (and
possibly retried) according to the default logic.

note that jobs with a strand or a run at will not be deferred and will
be created as usual.

this is particularly helpful for places like conversations, where a
bulk private message can trigger thousands of jobs (2 per private
conversation). even though the jobs are not run synchronously with the
request, their mere creation is a significant performance bottleneck.

test plan:
1. do some stuff that creates jobs
2. ensure the jobs are created normally
3. do some stuff that creates jobs in a Delayed::Batch.serial_batch
   block
4. ensure a batch job is created for each unique priority+queue
5. ensure no jobs with a strand or run_at get put in batches
4. ensure that invoking each batch job creates all its component
   deferred jobs and runs them synchronously

Change-Id: I36a56fa0c7ee2ffcd19e82dd5733496d4872f51a
Reviewed-on: https://gerrit.instructure.com/9169
Reviewed-by: Jon Jensen <jon@instructure.com>
Tested-by: Hudson <hudson@instructure.com>