hide content until scripts are loaded
Because RequireJS loads scripts asynchronously the
page renders but the JS hasn't been evaluated. In this
commit we hide all <a> and <button> tags until all
resources are loaded.

Previously the script tags blocked the page from
rendering while they loaded and executed.

When we have a single top-level require we can
remove this logic from the layout, and do it there
without all the polling fanfare.

Change-Id: I13771caf588508bacc6a70e4609c4af03bf522cd
Reviewed-on: https://gerrit.instructure.com/9191
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Ryan Florence <ryanf@instructure.com>