prevent group discussion assignments from spawning jobs recursively
discussion topics touch their root topics when they are saved so that post
counts get updated correctly. However, when the discussion is part of a group
assignment, part of the post save logic is to update the child topics to keep
the topics for each group in sync with the root topic. This was done in
a delayed job, and was kicking off a recursive job.

test-plan:
- with delayed jobs running
- create a group discussion assignment
- make sure delayed jobs for #refresh_subtopics don't start growing out of
  control

Change-Id: I592275db18239ba49eaed766dfab5ff1439e3024
Reviewed-on: https://gerrit.instructure.com/9431
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Zach Wily <zach@instructure.com>