retry updating discussion participant on duplicate key error
A race condition could cause this method to be called twice nearly in
parallel, meaning both processes would try to create the participant. The
second would fail due to the unique constraint. This catches the unique
constraint violation and retries, which means it will find the already created
one and do an update. (This is what we're opting to do instead of some global
locking for inserting discussion participants - it's optimistic locking.)

test plan:
 * this will only work in an environment where multiple rails processes are
   running. meaning: not development mode with webrick. :)
 * create a discussion topic and add a couple short replies.
 * make sure your window is tall enough to show both replies at the same time
 * log in as another user in that course and go to that discussion
 * verify that you don't get an ajax error, and unread counts get updated
   correctly

Change-Id: I35f6119cabaef890a228c570799b7b3c88bf3f74
Reviewed-on: https://gerrit.instructure.com/9467
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Zach Wily <zach@instructure.com>