user avatar api; refs #7763
this commit provides a new api endpoint that exposes the possible profile
pictures of a given user that can be set as her avatar. it also modifies the
existing user update api to allow the avatar image to be set. this permission
is given to the user herself and any account admin that has manage student
permissions

test-plan:
- make sure editing a user from the /users/x page and the /profile page still
  works.
- with the api, try querying the users profile pictures, and setting a new
  avatar.
- make sure to exercise different permissions:
  * look up and change your own avatar => should succeed
  * as an admin, look up and change a student's avatar => should succeed
  * as a student, try to change another student's avatar => should return 401

Change-Id: Idefe8549a92e109a248936970f9e3ef2bc414a59
Reviewed-on: https://gerrit.instructure.com/9645
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Zach Pendleton <zachp@instructure.com>