enable forgery protection for all selenium specs
The Rails default of not enabling forgery protection in specs means that
a whole class of bugs isn't tested for. Unfortunately enabling this for
non-selenium specs will take a significant amount of work on the specs,
but since selenium specs, make requests using the browser, enabling it
will be easy.

This would've caught cc06aed9351

Change-Id: I0f095a05170f44993b0c15ca1df51271d4459318
Reviewed-on: https://gerrit.instructure.com/9880
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Jon Jensen <jon@instructure.com>