disallow object_snippet calls on non-safefiles domains
fixes #8079

test plan: with safefiles configured, embed an object tag for a swf file
in a user content field (like a discussion topic), and then view html
source to pull out the form that posts to /object_snippet, including the
canvas verified signature. make that same post to the main non-files
domain, and verify it's denied. verify that object_snippet is still
allowed if safefiles is not configured -- though someday that may
change.

Change-Id: I7e6ae37482ff0c022bd4797d6589ba9a433a2626
Reviewed-on: https://gerrit.instructure.com/9902
Reviewed-by: Cody Cutrer <cody@instructure.com>
Tested-by: Hudson <hudson@instructure.com>