change tool load order to let courses override accounts
This is a somewhat significant change. Before, domain-matching
took precedence over url-matching, and account-level tools
took precedence over course-level tools. This didn't make sense
to any of us, but was based on what we thought was part of the
LTI spec.

Turns out it isn't part of the spec, so we're going to take
the opportunity to change the order or precedence to something
that makes sense to us, which is url before domain matching,
and course before account matching. The new code will start at
the course and work up looking for an exact URL match. If it
can't find one then it will start at the course and work up
looking for a domain match, and will use the first one it
comes across.

test plan:
- verify that the following precedence holds:
  (you can use the fake lti tool included in canvas to
  run these tests, "<canvasdomain>/selection_test" is the
  URL to launch. The easiest way to test would be to set
  a different consumer key for each tool).
1. course-level url matching
2. sub-account-level url matching
3. root account-level url matching
4. course-level domain matching
5. sub-account-level domain matching
6. root account-level domain matching

Change-Id: If58e055c605524f7a8ea9941c1b71dbcbd5ab816
Reviewed-on: https://gerrit.instructure.com/9913
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Brian Palmer <brianp@instructure.com>