correctly track dropdown answers, fixes #8156
fixes the behavior so that matching and multiple-dropdown questions only
show up as answered in the right side when *all* have been specified.
previously it would mark as answered whenever just one was specified (and
clear it when any dropdown was cleared).

also fixes multiple-dropdown questions so that your answers are remembered
in the form if you leave the quiz and resume it.

test plan:
1. create a quiz with a matching question and a multiple dropdown question
2. take the quiz
3. ensure that the questions are only flagged as answered on the right
   side when they have been fully answered
4. leave the quiz and resume it
5. confirm that your previous answers are there, and the right side
   correctly reflects what you have answered

Change-Id: Icdba25c70bbbfdc98dffa77ccd779190d7b672da
Reviewed-on: https://gerrit.instructure.com/10143
Tested-by: Hudson <hudson@instructure.com>
Reviewed-by: Joe Tanner <joe@instructure.com>