refactor and bug fixes for QuizQuestion.score_question
fixes #8267

This refactors the large switch into an individual class for each
question type. The hope is that we can start using these classes for
other type-specific code as well, such as the giant switch in the
parse_question method.

While refactoring, I fixed all the bugs that had pending specs in
quiz_question_spec.rb

test plan: verify that quizzes are still scored correctly. the only
visible changes should be the bug fixes for the pending specs:
  * there should be no floating point errors in scoring numerical questions
  * there should be no issues with floating point errors causing some
    question types to give 0 instead of full credit in edge cases
  * a blank answer shouldn't be treated as 0.0 in numerical questions
  * no answer shouldn't be treated as a blank answer for the purposes of
    undefined_if_blank (not currently exposed in the UI)

Change-Id: I74d2efcf6b087247adeafc5d4685cb3ea2bba9c2
Reviewed-on: https://gerrit.instructure.com/10263
Reviewed-by: Cody Cutrer <cody@instructure.com>
Tested-by: Hudson <hudson@instructure.com>