remove invalid quotes from speedgrader url link; fixes #8575
in gradebook1, we were using a json string to generate the speedgrader url to
a specific student, but using the json string inserted an extra set of double
quotes around the value that the url wasn't expecting and didn't handle well.
so we switching to just hardcode the format that we need.

also, this removes one block of dead code from speedgrader.js

test plan:
- go to gradebook1
- click on a cell for a submitted online text entry assignment
- make sure the "submission details" and the "view submission" links both load
  the speed grader with the correct student

Change-Id: I3023a1398ed1059ee00e6daa2655ce91226bc8e8
Reviewed-on: https://gerrit.instructure.com/10706
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Jon Jensen <jon@instructure.com>