upgrade ruby-saml gem
This version fixes the case where a SAML response arrives that doesn't have a
signature around the whole thing, but has a signature in the encrypted
portion.

test plan: Set up SAML and configure it to not sign the entire response, and
to encrypt assertions. Make sure you can still log in.

Change-Id: I3306b5595b82750c344a3c4a229aff1d86bbdde6
Reviewed-on: https://gerrit.instructure.com/10735
Reviewed-by: Simon Williams <simon@instructure.com>
Tested-by: Zach Wily <zach@instructure.com>