1.9 - fix discussion topic pagination
In 1.8, Enumerable methods that return a new object seem to dup the receiver
and then call the modify-in-place method on that dupe. That meant that you
could chain calls to a WillPaginate::Collection object.

In 1.9, the object is no longer duplicated first, so we need to edit something
like WillPaginate::Collection in place, or it will lose it's special
pagination attributes.

Change-Id: I3bd043984d49145df9d1ea4e4931f698de3e1336
Reviewed-on: https://gerrit.instructure.com/10860
Tested-by: Zach Wily <zach@instructure.com>
Reviewed-by: Zach Wily <zach@instructure.com>