make the locked_* dj columns match what is in the db
Because of how we query the function in a subselect, the returned data
doesn't include the changes to these two columns that happen in the
UPDATE in the function. so we update the columns client-side to match
what's in the database.

test plan: schedule a delayed job and then delete the record it points
to before it runs. then allow the job to run, and verify that it gets
properly unlocked after it fails.

Change-Id: I08807c88adb2a3ae4f901535f8d0e220962b887a
Reviewed-on: https://gerrit.instructure.com/10895
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Zach Wily <zach@instructure.com>