ensure course export includes all content; fixes #8693
a course export by default includes all content, but it uses the same path as
a selective export which ensures that referenced content is included even if
it's not explicitely requested. this is done by adding referenced content to
a hash, which caused the full export to be confused and think it too was
selective. this change makes explicit that the course export is to include
everything, and also prevents adding referenced content if the export already
includes everything.

test plan:
- in a course with content in at least two modules
- export the course as an .imscc package
- import the package, and make sure all content is available for import

Change-Id: I3dc7643b8941a1fb97102de2429a3d4f28b01795
Reviewed-on: https://gerrit.instructure.com/11007
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Bracken Mosbacker <bracken@instructure.com>