fix race condition in psql jobs popping optimization
Because a volatile function gets its own mvcc view of the data, the
subselect could end up finding and locking a row that doesn't exist for
the outer select, ending up with a row that's locked but not being
worked on.

We've switched gears to return the actual job row from the function,
removing the need for the subselect.

test plan: this race condition is rare, and impossible to reproduce
without dozens of jobs processes hitting the queue, along with many
processes inserting jobs. a regressing test is worthwhile though, making
sure delayed jobs function normally.

Change-Id: I49a8eb062d4e8ceee12283510d62c0ce0087c13c
Reviewed-on: https://gerrit.instructure.com/11069
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Ben Chobot <bench@instructure.com>
Reviewed-by: Bracken Mosbacker <bracken@instructure.com>