allow multiple scribd strands
we don't want scribd jobs to run wild, but we don't necessarily want to
limit them to 1 at a time either.

test plan: upload attachments that are scribdable, and by default
they'll all be created on the strand "scribd". Increase the
"scribd_num_strands" Setting and they'll also be created on multiple
numbered strands like "scribd:2".

Change-Id: Id06400d903a7badf6def5d46c61505e4f5e50fcb
Reviewed-on: https://gerrit.instructure.com/11338
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Cody Cutrer <cody@instructure.com>