optimize a few potentially expensive checks in the course wizard
Doing something like:

  @context.attachments.active.empty?

can actually be expensive if there are a lot of attachments, since it does a
count and then compares that to 0. In this case, we don't actually need to do
a count, we just need to know if there's at least 1 of the item:

  @context.attachments.active.first.nil?

We may want to monkey patch the named_scope implementation of empty? to do
this.

Also, we probably shouldn't even bother rendering the course wizard content on
the page until it's requested, but that will be a more involved change.

test plan:
 * Make sure you can still load your course home page
 * Create a new course, and make sure the wizard's steps highlight properly,
   and then dim properly after you've completed the steps.

Change-Id: I0850f26d26f24cd39862877842f54b71fc9dcf47
Reviewed-on: https://gerrit.instructure.com/11699
Reviewed-by: Jon Jensen <jon@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>