export the course's grading_standard too
The course.grading_standard isn't necessarily included in
course.grading_standards -- I think because in the old course copy code,
we just copied the id over and the standard was still owned by the old
course. This change makes it so we export+import that standard, as well.

Note there is a difference still here, the old course copy was a copy by
ref rather than copy by value, so updating the standard in one place
updated it everywhere. That's not true of the new export/import
workflow.

closes #9197

test plan: make a course that has a grading_standard_id pointing to a
standard owned by another course (or find one already in the db). copy
that course, and verify that the new course correctly has the standard
as well.

Change-Id: I74272db46f511cc329c68493962ef35aedfeba5d
Reviewed-on: https://gerrit.instructure.com/12000
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Bracken Mosbacker <bracken@instructure.com>