allow resetting student view; closes #9229
this creates a fixed bottom bar that shows up on every page when you are in
student view, serving as both a reminder and a place for the buttons to leave
student view or to reset the test student.  resetting the test student is a new
piece of functionality which removes all course artifacts created by the test
student, letting a teacher "start over" the test student.

test plan:
- enter student view
- there should be a permanant bar at the bottom of the page, with buttons
- there should no longer be a [Leave Student View] link in the top right corner
- the leave button should work
- as the test student, interact with the course (submit assignments, take
  quizzes, comment in discussions, etc)
- click the reset student link in the bottom bar
- all of the interactions above should be gone

Change-Id: If73be7a28f6af5378da26f92e82a69a72352d831
Reviewed-on: https://gerrit.instructure.com/12057
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Simon Williams <simon@instructure.com>