doesn't show concluded users in collaborations and conferences
fixes #8845

originally, this issue was that if a student was concluded from a course
they would still show up in the list of students to be added to both
collaborations and conferences. this commit fixes that and also prevents
the student view student from showing up in the list of students when
creating a new conference.

test plan
  - be a teacher in a course with a few students and collaborations and
    conferences enabled
  - go to a collaboration and a conference and see the students listed
    there (when starting a new collaboration/conference)
  - conclude a student from the course
  - click student student view, then leave student view
  - back in the collaborations and conferences, the concluded student
    should no longer be listed, and test student should not be listed

Change-Id: I08da39a79ef1508e53064adf54bfacac0c527457
Reviewed-on: https://gerrit.instructure.com/12100
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Sam Olds <sam@instructure.com>