masquerading bar; closes #9477
when masquerading, the only visual cue was the presence of the [Stop
Masquerading] link in the identity bar in the header.  this change moves that
link to a button in a fixed footer bar that is always present at the bottom of
the window and also contains a warning message as a reminder that you are
masquerading. Additionally, if avatars are enabled, the avatar of the user
being masqueraded as will be shown.

test plan:
- masquerade as someone
- you should get a fixed bottom bar
- the top right bar should say "As: other_person" and the other person's name
  should be highlighted
- the stop bottom should work

Change-Id: Ifb92b8a7d4b4dcada30ae505d99f9f01c59293ed
Reviewed-on: https://gerrit.instructure.com/12260
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Jon Jensen <jon@instructure.com>