fix todo api for teachers who are also students
The student needs submitting query was omitting needs_grading_count which is
returned in the assignment json if the user is a teacher in the course.

fixes #9621

test plan: enroll as both a teacher and student in the same course.
create an assignment due in the next couple days, with online turn in.
hit the /api/v1/users/self/todo endpoint, and verify that you get back
that assignment as a needs submitting todo, without any error.

Change-Id: Ic049dab07be5ab8863d22e1f2df98cd6873ea594
Reviewed-on: https://gerrit.instructure.com/12505
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Simon Williams <simon@instructure.com>