translate submission online_text_entry in api responses
fixes #9468

test plan: submit an assignment with online text entry, using the API or
the RTE in the UI. Include a media comment or a link to an attachment in
the course. Retrieve the submission information in the API as
either the student or a teacher, and verify that the content is
translated so that media comments become <video> tags and attachment
links become absolute urls that don't require a canvas session to
download.

Change-Id: If0c652334151f16eafb19525a173ca66ae7e4c36
Reviewed-on: https://gerrit.instructure.com/12641
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Simon Williams <simon@instructure.com>