only assign a group_comment_id for group submission comments
The logic here was causing inappropriate deletions to happen.

Also scope the SubmissionComment#destroy to only look in the current
assignment's submissions.

fixes #10280

test plan: as a teacher leave a comment on a group assignment
submission. also leave the same comment on another submission. then
delete the first comment, verify all copies are deleted for that group
but not outside that group.

Change-Id: I413c19becfbbd855c64fe162cb698c4df14aee2e
Reviewed-on: https://gerrit.instructure.com/13358
Reviewed-by: Simon Williams <simon@instructure.com>
Tested-by: Brian Palmer <brianp@instructure.com>