correct constraints on observer permissions
only grant observer permissions if the user is tied to the student as an
observer, and in the same course. do not grant observer permissions if
the user is tied to the student via a StudentEnrollment (how does this
tie get created?) or if the user is tied to the student as an observer
in a different course.

fixes #10508

test-plan:
 - enroll students John and Jane in course A and course B
 - set Jane as the associated_user_id on John's StudentEnrollment
   - John should not be able to read Jane's grades
 - add Bob as an observer of Jane in course A and as an observer of John
   in course B.
   - Bob should not be able to read Jane's grades in course B nor John's
     grades in course A.

Change-Id: I292e518fe9fbebeaf9e36ee0c85d97279049ae55
Reviewed-on: https://gerrit.instructure.com/13659
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Simon Williams <simon@instructure.com>