fix wiki links to use the correct page url on export
if you click "Link to New Page" in the UI, it will create a link
that may not match the eventual URL of the page.  the web
interface doesn't break because WikiPagesController tries
matching by urlified title or id.  this commit makes the
exporter fix wiki links to use the actual page urls.

fixes #9945 #9292

test plan:
 * create a new course
 * edit the main wiki page
 * click "Link to New Page", and enter something with colons
   or periods in the name
 * save the page
 * click on the link you made to create the new page
 * put some content in and save the new page
 * copy the course
 * the copied course should have a working wiki link from
   the main page

Change-Id: Ib944cd55679d0fce5bf87616564ff9c953e9d6c2
Reviewed-on: https://gerrit.instructure.com/13903
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Bracken Mosbacker <bracken@instructure.com>