fix sending bounce emails
The incoming message processor wasn't actually sending bounce
emails. The Message record was being created, but because it
wasn't assigned a user or communication channel, the message
never got sent.

Did some refactoring so user and communication channel is
initialized for the bounce message when we know them. If we
don't have enough information to identify the communication
channel, then we will just try to send the bounce message
directly without using the usual mechanisms. This should be
good enough for now.

refs #11068

test plan:
 - create a course with a teacher and a student
 - set student's notification for annoucements to asap
 - create an announcement as the teacher and then lock the
 announcement topic
 - find the annoucement message for the student and reply to it
 - the student should receive a bounce message for trying to
 reply to a locked topic

Change-Id: I7c1a80dd72ba25931ac3bad77d4d972108064e5b
Reviewed-on: https://gerrit.instructure.com/14688
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Mark Ericksen <marke@instructure.com>
Reviewed-by: Zach Pendleton <zachp@instructure.com>
Reviewed-by: Marc LeGendre <marc@instructure.com>