fix error when viewing a video file that has an unknown source type
There are some videos that have valid sources that don't have an extension for
some reason. This fixes handling of those correctly.

There was also a bug that would sort by bitrate as strings, instead of as
integers. The sort function was refactored out and tests written to ensure
that's fixed.

test plan:
 * find a video in kaltura that has an asset with no extension, and make
   sure you can view that video in a wiki page

Change-Id: Ie1bdabd042e30805711a6af59ed6e3d5527ebdba
Reviewed-on: https://gerrit.instructure.com/14699
Reviewed-by: Ryan Shaw <ryan@instructure.com>
Tested-by: Zach Wily <zach@instructure.com>