make media_object info action not an api, fixes: #11464
in order to view videos as an un-logged-in user
in a public course we need to make the endpoint
used to get a media_object's tracks and sources
not be an API action.  right now the API does not
support making requests as an un-logged-in user.

test plan:

try to view a kaltura video as an un-logged-in
user in a public course, it should work.

Change-Id: Ic08dc1f4402e3f4a9dfdc787346259eeec587620
Reviewed-on: https://gerrit.instructure.com/14715
Reviewed-by: Brian Palmer <brianp@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>

Conflicts:

	config/routes.rb