refactor tabs api to handle groups too, closes: #11474
we just added a tabs API to handle showing the
available tabs in a course, this refactors it to
handle groups too. I am using it in the wikiSidebar
so you can insert links to the tabs in the context.

test plan:

go to /api/v1/groups/x/tabs and make sure it shows
what would be shown on the left side nav for that
group.

(or if the wikiSidebar stuff is committed by the
time you test this, make sure that you can see the
links to all the parts of your course in the wikiSidebar
when you are editing something in that group/course)

Change-Id: I442d62955f3e9642ee123b8f220db4721f5f3b05
Reviewed-on: https://gerrit.instructure.com/14752
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Joel Hough <joel@instructure.com>