fix quiz form translation key error
fixes #11569

we changed a translation key in a quiz form from having two interpolated values
to having four, without changing the name of the key.  this caused the page to
break when a locale was used that hadn't updated their translations.

this fix changes the key, and uses a better style of html interpolation for the
strings.

test plan:
- set your course to russian
- create a new quiz
- it should work
- in any language numerical answer quiz questions should still function exactly
  the same as before (specifically providing answer possibilities either exact
  with error margin or in a range)

Change-Id: Ia9754c797a894569eae5bccc6e534b3459dce683
Reviewed-on: https://gerrit.instructure.com/14879
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Stanley Stuart <stanley+gerrit@instructure.com>