untaint DJ handlers
because it comes from the database, it's marked as tainted, but when
it's deserialized, it taints actual classes that are used. This is
bad for specs, cause when you stub or unstub something, new methods
are defined, and the tainting spreads to them, and the original
method is called via a call(), which if tainted, raises a security
exception regardless of $SAFE level.

this isn't a *huge* deal, cause canvas runs at $SAFE=0, so we're
not relying on tainting anyway.

Change-Id: Ic7ed5642de7ebfba7d4930bb6bca428e24f621ff
Reviewed-on: https://gerrit.instructure.com/15065
Reviewed-by: Brian Palmer <brianp@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>