refs #5773 fix plugin asset loading for 1.9.3
Ruby 1.9 uses a different YAML engine (although
we reset it to the 1.8 engine in the delayed
job plugin).  This resulted in the yaml
dump to the assets.yml file (for plugin
asset inclusion) not producing
quite the expected output format at server 
startup (before the delayed job plugin had
loaded), which caused
selenium tests to fail in the canvalytics 
plugin because they ONLY loaded the assets.yml
file at startup rather than before each
request like in development.

A small regex change in the ruby block
in assets.yml fixed the actual problem.  While
I was in here, I moved the ruby block into
it's own class so that all the processing
can be tested.  the "subdoc" stuff in bundle_yml
could probably still stand to be slimmed down
a bit but this is a decent intermediate step.

Change-Id: I9300c167130fb3305c7f37710b0e3cfcfda19f48
Reviewed-on: https://gerrit.instructure.com/15111
Reviewed-by: Jacob Fugal <jacob@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>