handle more device types on mobile login page
fixes: #11476

restyle mobile login page to handle different
screen dimensions and resolution (retina/non-retina)
so that it works on new iPhone, android browsers, etc.

use mobile_login for iPad iOS app and android app too
the iOS and android apps are now going to pass a
query string param of mobile=1 when it loads the
webview for login. it will set a cookie to
remember that it prefers the mobile optimized
login screen so that if it goes through any
redirection (bad user/pass) it still looks right.
fixes: #11686

Change-Id: Iafdc7522189dc44be5c5bc4592c5cb194a188612
Reviewed-on: https://gerrit.instructure.com/15283
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Chris Hart <chris@12spokes.com>
Reviewed-on: https://gerrit.instructure.com/15668
Reviewed-by: Cody Cutrer <cody@instructure.com>
Tested-by: Cody Cutrer <cody@instructure.com>