remove circular dependency in ajaxJSON/instructure_forms
circular requirement dependencies break qunit for some reason now
(recently started). they appear to be fine in production, but we lose
test coverage. this circular dependency was one of the worst offenders.

each defined methods that made more sense in the other, and depended on
things from the other. since only one of the methods moved actually
depended on anything from the current host, this resolves the
dependency.

most of the files that needed one or more of these three methods already
included both, so it was just a comment change in which file is
providing which method. a handful added a new require. one file dropped
a dependency (yay).

jquery.ajaxJSON no longer pulls in jquery.instructure_forms, so if
something was depending on instructure_forms implicitly through
ajaxJSON, that may be broken (and should be fixed by making the
requirement explicit).

meanwhile, take advantage of touching the zip_file_import_form partial
to extract its js into a bundle.

Change-Id: I3b85ab267748a7492662e4e3281820d749d3fe08
test-plan:
 - primarily a refactor. no new behavior, just check for regressions,
   particularly around zip_file_import_form
 - rake js:test should include most if not all javascript specs now (a
   large percentage didn't run before)
Reviewed-on: https://gerrit.instructure.com/15411
Reviewed-by: Ethan Vizitei <ethan@12spokes.com>
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Clare Hetherington <clare@instructure.com>
Reviewed-by: Brian Palmer <brianp@instructure.com>