fix submitted assignments randomly showing as overdue
fixes #11953

This fixes a typo in sorts_assignments which got the wrong key out of
the passed hash in SortsAssignments.by_varied_due_date.

The spec has been updated as well to take a list of submissions. It
now also checks that #by_varied_due_date will raise an error if any
fields it requires are missing.

test plan:
  create an assignment with an overdue due date ( due_at < Time.now )
  create an assignment with a non-overdue due date ( due_at > Time.now )
  as a student in the course, visit the Course Asssignment page
  make sure that the overdue assignment appears under "Overdue
  Assignments"
  make sure that the non-overdue assignment is not in the "Overdue
  Assignments" section on the page.

Change-Id: I74bc99521e4675b5f4f92bfdfa7c175daa99b3aa
Reviewed-on: https://gerrit.instructure.com/15586
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Simon Williams <simon@instructure.com>