Prevent malicious zip uploads
refs #11857

don't allow zip files with more than
100000 files to get attached

don't allow zip files that add enough
data (with a minimum file size of 512)
to go over the context quota

also refactor unzip_attachment.rb a
bit as long as we're in there because
it's a bit dense and monolithic.  There
is still much that could be done there,
but it's not one enormous method anymore.

TEST PLAN:
  1) login as any user with permissions
  to upload files.

  2) attempt to upload a zip archive with
  more than the configured max number of
  files inside of it (currently 100,000)

  3) you should get an error preventing
  this action from occuring and no
  new attachments should have been made
  to that context

  4) navigate to some context where the
  attachment quota is known.

  5) attempt to upload a zip archive to
  that context with enough size to exceed
  that quota.

  6) you should get an error preventing
  any of the new attachments from being
  attached to that context.

Change-Id: I35ce673e96fb98728cb36cfd27538fd802a249f2
Reviewed-on: https://gerrit.instructure.com/15671
Reviewed-by: Cody Cutrer <cody@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Clare Hetherington <clare@instructure.com>