spec: reload before comparing times from DB
some DBs truncate times to second precision (instead of microsecond). if
the original in-memory object has microseconds, but the object compared
against has the same value loaded from the DB, they might not match when
expected to. so make sure the in-memory object also has the value that
has been round-tripped through the DB.

Change-Id: I531d7160604291d2ae0364e1b125b1b1edaaceee
Reviewed-on: https://gerrit.instructure.com/15763
Reviewed-by: Brian Palmer <brianp@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>