dont hit scribd from the files_controller
refs #CNVS-1267

this reworks the thumbnail_url method
on the attachment model to just use
the cached scribd thumbnail if available
and to return nil if not (in the case
of a scribd document).

also did a bit of refactoring and removed
some trailing whitespace.

TEST PLAN:

create a scribd doc attachment to an assignment
submission.  It should not hit the Scribd API
at all during the course of the request.  If
the tester does not have a way to monitor network traffic
than this could still be confirmed by ensuring the
request response time isn't significantly longer
than uploading a NON-scribd attachment

Change-Id: I324a99ec94e2efdeb8d1ead80ee0cd92eec036ac
Reviewed-on: https://gerrit.instructure.com/15821
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Brian Palmer <brianp@instructure.com>
Tested-by: Brian Palmer <brianp@instructure.com>
QA-Review: Brian Palmer <brianp@instructure.com>