don't load unncessary assignment information for js grade calc
we load the assignment groups, which include assignments, which by default
includes the discussion topic info the the assignment is linked to
a discussion. this was unncessary, and was causing an error because the
discussion topic json does a select for update, and this is all happening in
a slave block.

fixes #CNVS-2430

test plan:
- in an environment with a master and slave database (user)
- try to load gradebook1; it should work
- try to load a student's grade summary page; it should work

Change-Id: I2fb9a202d1b30660871708ace16cb3be4cebb445
Reviewed-on: https://gerrit.instructure.com/16396
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Stanley Stuart <stanley+gerrit@instructure.com>
QA-Review: Stanley Stuart <stanley+gerrit@instructure.com>

Conflicts:

	app/models/assignment.rb
	spec/controllers/gradebooks_controller_spec.rb