additional spec coverage for enrollment permissions
refs #CNVS-2036

Added these specs while exploring the story
regarding teachers and the manage_sections
permission.  Per brianp and cody, this may
not really be a bug, so it may not require
any work, but these specs are still valid
for the current behavior of the system
so it seems valuable to keep them.

TEST PLAN:
No production code has changed.  If the
specs pass jenkins, we're good.

Change-Id: I6f46bf997df43149de46561fe8473faddbd8a8bc
Reviewed-on: https://gerrit.instructure.com/16397
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Jacob Fugal <jacob@instructure.com>
Reviewed-by: Cody Cutrer <cody@instructure.com>
QA-Review: Clare Hetherington <clare@instructure.com>