don't try to sanitize params on already graded quiz
in order to sanitize params on a can't go back quiz, we need to compare them
with the params in submission_data, which doesn't work if the quiz is already
graded and submission_data has become an array.

fixes #CNVS-2795

test plan: i'm not positive how to reproduce this issue, because it requires
submitting an already submitted quiz.  maybe if you have a quiz open in two
different tabs and submit first in one and then in the other, it should
redirect you to the results page (although not change your answers)

Change-Id: I7a3cee402077343ab9de85a51ec33d02217c6754
Reviewed-on: https://gerrit.instructure.com/16736
Reviewed-by: Stanley Stuart <stanley+gerrit@instructure.com>
QA-Review: Stanley Stuart <stanley+gerrit@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>
QA-Review: Myller de Araujo <myller@instructure.com>