prevent bad sis_user_id queries
fixes #CNVS-2172

found every place I could in the codebase
that passed some kind of input into a finder
that looked up a pseudonym with an sis_user_id.
Then wrote specs to pass in integers (which in 
all cases turned up the error described
in the ticket), and 
made sure to do explicit casts on the inputs
until the specs didn't break anymore.

TEST PLAN (unsure of how to trigger from UI):

1) Take a pseudonym that already exists and has
an sis_user_id that could be an integer (is
made up of all numeric characters).

2) In the console, try to create a new 
pseudonym with the same sis_user_id, but use
an integer instead of a string.

3) you SHOULD NOT get a postgres error
telling you that you need to do an
explicit type cast

4) you SHOULD get a validation error telling
you that 2 pseudonyms cannot have the same
sis_user_id.

Change-Id: Ic69da6a40609a50234a5388af02dd1fbeb021214
Reviewed-on: https://gerrit.instructure.com/16925
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Brian Palmer <brianp@instructure.com>
QA-Review: Clare Hetherington <clare@instructure.com>