stop invalid submissions queries
fixes #CNVS-2183

grep-ing the project I only found one
place where invalid queries of this type
were likely to be originating (open input
from the client side, passed to a query
that would generate SQL finding a single
submission by user_id and assignment_id).
The submissions controller was open to this in 2
places, both related to the 'turnitin'
functionality.  I added specs to prove this pattern
 was possible and put in some guard clauses to
prevent it from getting down to the database, but
the actions will still return 400s as the
input {{ user_id }} for a user_id is nonsensical.
I found several places in view code where
'{{ user_id }}' was used as a placeholder for 
javascript to later replace, and in this file
in particular:

app/views/gradebooks/speed_grader.html.erg

there are 2 anchor tags that would create urls
that could be used in this way.  ids:

a#assignment_submission_turnitin_report_url
a#assignment_submission_resubmit_to_turnitin_url

in the javascript that makes use of these a
tags, the href values get run through
$.replaceTags which should strip out
these '{{ user_id }}' spots, but if this anchor
tag were to be used on it's own for some reason
(perhaps through some automated browser tool
that pings all the hrefs of a tags on the page
regardless of their visibility, or perhaps
through somebody just messing around and exploring
the source html of the page they're looking at)
it would create exactly the error described in 
the ticket.  Given I can't find a use case
where the client side code would call these
urls as-is without processing first, I don't see
anything that needs to be fixed in the js/coffeescript.

TEST PLAN:

There is no change in behavior, this just stops an
edge error case from ever hitting the database.
Regression tests on the speed grader should 
be sufficient.

Change-Id: Ic8eb8a2b04401a077bfe5abbd905634d781e32fa
Reviewed-on: https://gerrit.instructure.com/17080
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Cody Cutrer <cody@instructure.com>
QA-Review: Clare Hetherington <clare@instructure.com>