added null check for child events in processAppointmentData
fixes #CNVS-3387. The CommonEventFactory was returning null
for some child events because it couldn't find the proper
context for them. This might be a legitimate filter, although
the api probably shouldn't be returning them. Either way, it
makes sense to put a null check on the results of the factory
since we do this elsewhere as well.

Test Plan:
 - Check for scheduler regressions

Change-Id: Idca8942ac5e138107248cda9c4dd04452823239e
Reviewed-on: https://gerrit.instructure.com/17290
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Zach Pendleton <zachp@instructure.com>
QA-Review: Marc LeGendre <marc@instructure.com>