Rails PG adapter: Properly quote Infinity and NaN
refs CNVS-1368

This backports a Rails 3.1 fix as a monkey patch to our Rails 2.3.X
installation. It should address unquoted instances of Infinity we are
seeing occasionally in web conference queries.

The test plan for this is not UI-based, since I was unable to reproduce
by poking around the BigBlueButton web conferences UI.

Test Plan:
 - Inside a ./script/console, choose a model with a column of type
   "float", like WebConference
 - Try to create it with 0.0/0 and 1.0/0 as the value of that column,
   like BigBlueButtonConference.new(:duration => 0.0/0).save(false)
 - Verify that no database exceptions are thrown

The official Rails commit with this fix is here:

https://github.com/rails/rails/commit/06c23c4c7ff842f7c6237f3ac43fc9d19509a947

Change-Id: I41141054aa06c88ed121a33e6e2da578ff9be82d
Reviewed-on: https://gerrit.instructure.com/17317
Reviewed-by: Cody Cutrer <cody@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>
QA-Review: Clare Hetherington <clare@instructure.com>