only set association in one direction
refs CNVS-2651

makes it so 1. we don't need to clear the errors and 2. it doesn't get
auto-saved along with the context if has been accessed (e.g. in g/17020,
we check Context#profile in an after_save, which makes it want to save it
if we try to save the course a second time)

test plan:
* https://gerrit.instructure.com/17020

Change-Id: I182da6c39d50bde03bbbbab9fef57826a919f462
Reviewed-on: https://gerrit.instructure.com/17426
Reviewed-by: Joe Tanner <joe@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>
QA-Review: Jon Jensen <jon@instructure.com>