spec: fix calendar2 spec on sundays
The beginning of the calendar week view is sunday, but
time.beginning_of_week returns the previous monday. So on sundays, the
event was getting created in the week before and so not getting
displayed, breaking the spec.

This should make sure that the event always shows up on today on the
calendar, and so will always be displayed.

Change-Id: I2dcc25ce8c1fd33fb4223e288668fc435fdc8c22
Reviewed-on: https://gerrit.instructure.com/17812
Reviewed-by: Jon Jensen <jon@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>
QA-Review: Brian Palmer <brianp@instructure.com>