stick mail gem to version 2.4.4, even in ruby 1.9
v2.5.3 of the gem has a regex parsing error where it performs
negative lookahead on whitespace after the blank link that
separates the header from the body. This means that it will
misparse a message when the body starts with whitespace.

There is a fix on the master branch of the gem's repository, so
the next version of the gem should work. We should stick to
2.4.4 until we drop 1.8 support and the new gem has been
released.

fixes CNVS-4026

test plan:
 - bundle update and make sure bundler installs and uses
   version 2.4.4 of the mail gem

Change-Id: I4f3d2405e742a724e217fc803945e15c906f0f37
Reviewed-on: https://gerrit.instructure.com/17947
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Zach Pendleton <zachp@instructure.com>
QA-Review: Zach Pendleton <zachp@instructure.com>