prevent error when google docs plugin not configured
refs #CNVS-3196

simon noted that if the google docs plugin was
not configured, trying to access the config hash
for the google docs library would result in
a "no such method [] on nil" error.  I placed
specs at the controller level and at the library
level to reproduce the problem and put in a fix
that checks for a google docs config and fails
early if there is not one with a nil return.

Tested manually locally to confirm this solves it.

TEST PLAN:

1) login as any student to a canvas instance
where the google docs plugin is NOT enabled.

2) navigate to the show page for any assignment

3) you should not experience any server errors

Change-Id: I25f90ded428a5a6a9c15ddd625b1a690f6d9077f
Reviewed-on: https://gerrit.instructure.com/18177
Reviewed-by: Cody Cutrer <cody@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>
QA-Review: Clare Hetherington <clare@instructure.com>