fix spec failure in ruby 1.8
Using Symbol#to_proc on the index_by here triggers what looks to be a
Mocha bug in ruby 1.8, it calls the spec helper method "course" instead
of the course method on the mock object. The easiest workaround is to
just change to an explicit block until we drop 1.8 support.

test plan: specs should pass on 1.8 and 1.9

Change-Id: I96bd3ce3fe2c8df06d69c69398891815a6fa94ad
Reviewed-on: https://gerrit.instructure.com/18465
Tested-by: Jenkins <jenkins@instructure.com>
QA-Review: Bryan Madsen <bryan@instructure.com>
Reviewed-by: Paul Hinze <paulh@instructure.com>