remove redundant someone-else-edited warning
there are two different warning messages when viewing a page
that has changed since your page load - one (with a prompt to
reload) when you are just viewing the page, and another
(with a scary icon) when you are editing a page and about
to overwrite someone else's changes.

we were actually displaying _both_ messages in the latter
case, resulting in a bit of a mess.  just display the one.

fixes CNVS-4652

Change-Id: I479510319c8fa8244191477ee36758095507cf60
Reviewed-on: https://gerrit.instructure.com/18654
Tested-by: Jenkins <jenkins@instructure.com>
QA-Review: Adam Phillipps <adam@instructure.com>
Reviewed-by: Bracken Mosbacker <bracken@instructure.com>