decrease periodic quiz submissions
We currently submit quiz submission values every 30 seconds as a
precautionary measure to prevent accidental data loss. However, when
hundreds or thousands of students take a quiz at the same time it is
unnecessary and problematic to send so much redundant data.

This patch checks the submitted data (on the client side) to see if
the data is identical to the most recent submission. If identical,
it just doesn't submit it.

fixes CNVS-2153

Change-Id: If10d48acefc29e2ebadb3ea4d1031f5ecb209a5d
Reviewed-on: https://gerrit.instructure.com/18662
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Stanley Stuart <stanley@instructure.com>
QA-Review: Clare Hetherington <clare@instructure.com>
Product-Review: Duane Johnson <duane@instructure.com>