quiz api: only index action should care about tab_enabled?
all the quiz api actions were checking tab_enabled? but it is possible in
canvas to be able to take a quiz even if the quiz tab is disabled (for example,
if it is linked to from a module), so to better mirror that behavior, only the
index action should care if the tab is enabled or not.

fixes CNVS-5473

test plan:
- create a course with some quizzes, and disable the quiz tab
- as a student, /api/v1/courses/X/quizzes should be a 404
- but /api/v1/courses/X/quizzes/Y should return the quiz object

Change-Id: Icbfb1850e1294784b2a3cf2783b5b2f5dedd491f
Reviewed-on: https://gerrit.instructure.com/19976
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Stanley Stuart <stanley@instructure.com>
QA-Review: Amber Taniuchi <amber@instructure.com>
Product-Review: Simon Williams <simon@instructure.com>