Hijack the Start New Course button (CANVAS-192)
This turned out to be a bit more difficult than I expected. This code grabs the click events bound to the document DOM node and iterates over the list, looking for one targeting .element_toggler. If it finds one, it caches the handler function, removes the click and change handlers from the document and adds a new one. Our new handler checks to see if the target (what element was clicked) is a Start New Course button and if so, redirects to the form. If not, it runs the original event handler in the context of the event.