draft state on quiz_show page
fixes CNVS-5675

This adds the "publish" button to the
quiz show page and includes an ajax callback
to talk to the server about changing the publish
state of the quiz.

TEST PLAN:
1) login as an instructor
2) navigate to the show page for any quiz
3) use the "publish" button to toggle the
state of the quiz to be published and unpublished
4) if you are on a quiz with student submissions
already in, the button should do nothing, and you
should get a rollover showing you the reason
why you can't click on it.

Change-Id: I45543ce9e62a1a4882d9a6b00a559ddd6eda9897
Reviewed-on: https://gerrit.instructure.com/21308
Reviewed-by: Stanley Stuart <stanley@instructure.com>
QA-Review: Myller de Araujo <myller@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Derek DeVries <ddevries@instructure.com>
Product-Review: Ethan Vizitei <ethan@12spokes.com>