edit page draft state
fixes #CNVS-6293

this adds a draft state indicator to the edit
page, no functionality is attached, per the mockups
mentioned in the ticket.

TEST PLAN:
1) login as an instructor
2) go to the quizzes index and edit a quiz
3) you should see an indicator that is GREEN if
the draft state is currently published.  If it's not
the indicator should be a light grey and say 
'Not Published'

Change-Id: I3247c145c8ddd9731a901090c0e1dca1a43af2fb
Reviewed-on: https://gerrit.instructure.com/21460
Reviewed-by: Ethan Vizitei <ethan@12spokes.com>
Product-Review: Ethan Vizitei <ethan@12spokes.com>
QA-Review: Ethan Vizitei <ethan@12spokes.com>
Tested-by: Jenkins <jenkins@instructure.com>