Conditionally show unenroll link/button if allowed
Only show the unenroll link/button if the enrollment is not defined by
SIS, because the operation is otherwise not allowed.

In the People roster, enrollments are grouped by user. The link is shown
only if ALL of the user's enrollments are not defined by SIS.

NOTE: This does prevent admins who have the manage_account_settings
permission from removing SIS-defined enrollments. If this is a valid use
case, this changeset should be revisited.

CANVAS-224