fix time elapsed display when taking untimed quiz
fixes CNVS-7924
This fixes a bug introduced when fixing CNVS-6858.  When an untimed quiz
was taken, the time was not being displayed correctly (NaN due to
incorrect parsing and logic). It also fixes a bug with the countdown
becoming stalled when the window is out of focus.

Test plans:
1. Take an untimed quiz
- the time elapsed should be shown correctly
- and a timed quiz should show the time left correctly

2. Take a timed quiz
- go to a different browser window for 30 seconds
- go back to the quiz
- the quiz's countdown should sync immediately

Change-Id: I402a14930af9dd79dbc8f2646cb4a72162fd1662
Reviewed-on: https://gerrit.instructure.com/24167
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Derek DeVries <ddevries@instructure.com>
QA-Review: Myller de Araujo <myller@instructure.com>
Product-Review: Josh Simpson <jsimpson@instructure.com>