Add a permission option in the api to return if the user can create topics.
closes CNVS-6824

This adds a permissions attribute to the returned json for discussion topic
contexts (Course, Group).

The permissions attribute contains an optional permission of
"create_discission_topic" which returns true or false depending on whether the
current user can create discussion topics for the course or group.  For
performance reasons this is only added to a single course/group json and not in
lists so the only call that will return it is /api/v1/<context>/<context_id>
where context is a course or group.  Since we did not want to include this on
every response its a custom permissions attribute for course and groups in the
course_json or group_json serialization methods.

Using the includes parameter for the API supplying a value of "permissions"
will include the permissions with "create_discussion_topic" for a group and
course.  When the object is serialized to json it checked to see if the model
implements a serialize_permissions method and calls that to render or override
permissions generated from the policies.

- Create a test Course. Make sure the course allows members to post topics.
- Add a student to the course.
- Make a call to /api/v1/courses/<id> where "<id>" is the id of the created
  course.
- The response should include a permissions attribute with a boolean value for
  "create_discission_topic" see the example below.

- Make a call to /api/v1/courses to return a list of course objects.
- The permissions attribute should not be included in the response.

- Create a test Group tied to the course created in the first step.
- Add a mamber to the group.
- Make a call to /api/v1/groups/<id> where "<id>" is the id of the created
  group.
- The response should include a permissions attribute with a boolean value for
  "create_discission_topic" see the example below.

- Make a call to /api/v1/groups to return a list of group objects.
- The permissions attribute should not be included in the response.

Example Response:
{
	id: 42,
	...
	permissions: {
		create_discission_topic: true
	}
}

Change-Id: Ia02d5aa67e345740a93dd0f63e357e7cb5e1efd6
Reviewed-on: https://gerrit.instructure.com/24478
Reviewed-by: Jacob Fugal <jacob@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Product-Review: Nick Cloward <ncloward@instructure.com>