[WebAIM] @title attributes for icons in quiz views
Some <a> tags included icons (marked by <i> tags with a @class of
"icon-" something) and had either a descriptive @title or @text set
but the icon elements did not and that was affecting the accessibility
of these icons. So, all icons that are parented by an anchor with a
description now have a @title that mimics that description accordingly.

Specs had to be adjusted to properly locate accessible anchors since
their text() value now returns "duplicated" values, once for the anchor
itself, and once for any embedded span.screenreader-only elements.

I've added/modified a few helpers to quizzes_common.rb to accommodate
these changes.

closes CNVS-6546

Note 1: there are a few cases that were not handled as I didn't know
what to actually do with them, more on that in the related JIRA issue.

Note 2: the number of affected files is large because the ticket
requested that not only the reported items be fixed, but any other
related items as well.

Change-Id: Ie34048d7d68b41b74461696b76751340e000344f
Reviewed-on: https://gerrit.instructure.com/24493
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Derek DeVries <ddevries@instructure.com>
QA-Review: Myller de Araujo <myller@instructure.com>
Product-Review: Derek DeVries <ddevries@instructure.com>