fix for interval saving for quizzes
fixes: CNVS-4676
because we use the same function to do both auto saving (on a 30 second
interval), and event-based saving (e.g. when the user answers a question
or otherwise interacts with the quiz form), sometimes the auto interval
wouldn't fire due to various circumstances.  we bypass this by
explicitly declaring that we're performing an interval save (vs an event
based submission save). i've also added new messaging when we try to
save, but when there's no new data to save (giving the user a time stamp for the
latest save attempt).

test plan:
(this can be confusing to test)
1. take a quiz
2. answer a question
- note that the quiz is still saving (or attempting to save, if there's
  no new data to save) at a regular interval

Change-Id: I139f94cb5b3cdae1d8b316d1c9510f1352d83ae3
Reviewed-on: https://gerrit.instructure.com/24558
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Derek DeVries <ddevries@instructure.com>
QA-Review: Myller de Araujo <myller@instructure.com>
Product-Review: Josh Simpson <jsimpson@instructure.com>