Wrap everything in a if test to only execute on a test_cluster machine
A bug where character encoding was being changed from UTF-8 to ASCII-8BIT was encountered after this plugin was introduced. The mere presence of ApplicationController.class_eval (even as a no-op) triggers it, so as a workaround we only execute on true test_cluster machines, where the bug will exist but isn't as critical.