replace one-off fudge calculation with fudge call
now the use of user_offset and tz_offset is apparent -- it's just
fudging the parsed "true" timestamp into a date object that will format
as if in the profile timezone. this is what fudgeDateForProfileTimezone
is for; use it.

Change-Id: I82a82dc3f1946cd733e399331fe9d9950fa3b2f9
Reviewed-on: https://gerrit.instructure.com/31975
Reviewed-by: Jacob Fugal <jacob@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>
Product-Review: Jacob Fugal <jacob@instructure.com>
QA-Review: Jacob Fugal <jacob@instructure.com>