Fix inability to preview deleted/re-uploaded file
The problem affects an instance using local storage and Google Docs
Preview. When previewing a file that was previously deleted, an error
message saying "unable to find the document at the original source" will
be displayed in Google Docs Preview. This is because the preview URL is
no longer valid. In fact, it redirects to the login page.

Test Plan:
* Disable Scribd but enable Google Docs Preview
* Upload a brand new file to a course
* Preview the file
* Delete it, and re-upload the same file
* Confirm that the file can still be previewed