Simplify section_csv method
Eliminate `unless cross_list` statement by enumerating where the
section_csv method was called, instead of calling it recursively.

Simplify logic that determines when a section is added to the array, by
using `next` to skip, instead of multiple if/elsif statements.

Eliminate the need to join and split section info parts using ":_:".