fix loading course calendar event edit page with section dates
there was a javascript error on this page because it was
comparing a string id to a numeric id. the splitAssetString
function was still returning the id it parses out as a number
instead of a string. most places that call this function won't
care about the change. the one place that did only cared because
it was comparing it to data coming from ENV that should have
been using string ids anyway.

this also exposed a race condition in our javascript loading,
where one bundle can start an ajax call before the common bundle
loads the ajaxPrefilter in the instructure.js file. an explicit
dependency fixes this specific issue; the more general problem
should be solved in a separate commit.

fixes CNVS-9182

test plan:
 - create a course calendar event with section specific dates.
 - navigate to one of the events in the calendar, click on it,
   click the "edit" button, then the "more options" button.
 - should take you to the event's edit page, which should load
   normally. Note that it can take a minute to load the event
   edit page the first time and it doesn't have a loading
   indicator (which is a separate issue).

Change-Id: I7c2525081973ad5327b5508bba3d544e83a34b9c
Reviewed-on: https://gerrit.instructure.com/26567
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Zach Pendleton <zachp@instructure.com>
QA-Review: Steven Shepherd <sshepherd@instructure.com>
Product-Review: Jon Willesen <jonw@instructure.com>