only include unpublishable in the json response when applicable
In other words, when the API caller can publish/unpublish the quiz

Also change the semantics of Quiz#has_student_submissions slightly, so
that we can do a simple count > 0 query, rather than loading all the
submissions just to determine it. The only difference in behavior is if
a user took the quiz and then their enrollment was dropped.

Change-Id: I2aa779d247b87a5f1775fc93baa74b6920045569