Quiz Questions API - Allow student access
Students can now get a censored version of quiz questions, with enough
data to build the quizzes they have to take, but not enough to read the
correct answers and make the entire thing moot.

closes CNVS-10222

Note to code reviewers: the spec file shows too many changes because of
a re-indentation/rescope .. the original specs are untouched, but it was
necessary to scope them under a context ("as a teacher") to escape the
wrath of the global before(:each) block that was installed for all
specs.

TEST PLAN
---- ----

Endpoints:

  1. /api/v1/courses/:course_id/quizzes/:quiz_id/questions
  2. /api/v1/courses/:course_id/quizzes/:quiz_id/questions/:question_id

- Create a quiz with a question or two
- As a student using the API, make a GET request to both endpoints
  above and verify that:
  - you can read the questions
  - you don't get to see the answer weights
- Add a "?include=assessment_question" to the URL and:
  - verify that the "answers" array entries in the "assessment_question"
    do not list the answer weight either

Change-Id: I5c32030d92a1c46993706c2aa6f0575f858b60c0
Reviewed-on: https://gerrit.instructure.com/28022
QA-Review: Myller de Araujo <myller@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Derek DeVries <ddevries@instructure.com>
Product-Review: Ahmad Amireh <ahmad@instructure.com>