Quiz Submission Questions - Index/Show
This patch makes it possible to list the "question records" for a QS
which clients can use to tell what questions are answered, flagged,
etc.

closes CNVS-10025

TEST PLAN
---- ----

Fire up some computer, launch a Canvas instance, create a quiz with a
question or two, take it as a student, answer it, then get ready for
API business:

> Index action - Listing all question records

[GET] api/v1/quiz_submissions/:quiz_submission_id/questions

  - Perform the request
  - Verify you get N entries, one for each question you answered or
    marked
  - Set an "include" parameter with "quiz_question" (ie,
    "?include=user")
    - Verify that the returned set now contains 3 entries:
      1. "quiz_submission_questions" what you saw earlier
      2. "quiz_questions" with Quiz Questions that you answered
      3. "meta" that you can safely ignore

> Show action - Fetching a single question record

[GET] /api/v1/quiz_submissions/:quiz_submission_id/questions/:id

  - Perform the request
  - Verify you get 1 entry
  - Try the "include" parameter as you did in the Index case

As a sidenote, the output of these endpoints should be the same as the
stuff you get when you answer a question (see CNVS-9844), they
basically should tell you what answer you provided if you did, and if
you marked the question or not.

Change-Id: If39b84d8fda77870c3990822ade06bd28b276c3d
Reviewed-on: https://gerrit.instructure.com/27598
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Derek DeVries <ddevries@instructure.com>
QA-Review: Myller de Araujo <myller@instructure.com>
Product-Review: Ahmad Amireh <ahmad@instructure.com>