fix rails3 call to symbolize_keys! in submission
in rails 3, symbolize_keys! is not defined for a
HashWithIndifferentAccess (it is still defined for
a Hash). symbolize_keys, however, is. The method
doesn't do much since the whole point of 
HashWithIndifferentAccess is to allow one to access
keys using strings or symbols, but we still call it
as we also accept a Hash for options.

Change-Id: I3d78e7e2779aa24c3598db6e636ee8e071ced777
Reviewed-on: https://gerrit.instructure.com/28545
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Landon Wilkins <lwilkins@instructure.com>
Product-Review: Anthus Williams <awilliams@instructure.com>
QA-Review: Anthus Williams <awilliams@instructure.com>