fix spec failures for rails3 on student analysis
things fixed:
- Calling course#enroll_user makes a call to user#touch. In Rails2, #touch runs
callbacks for a model since it actually calls #save! under the hood. In Rails3
model#touch does an #update_all and doesn't execute callbacks. I have fixed the test
to #save! after enrolling the user because I'm assuming that #touch in
course#enroll_user is actually meant to update the user's updated_at date and not
actually intended to save with callbacks (if that was the case we would have
just used #save right?)

- ActionController::TestUploadedFile is no longer available in Rails3. We
however can use the more generic fixture_file_upload() method which works in both
Rails2 and Rails3.

- Since the stats records are returned by updated_at in the code, and updated_at
is the same date for both records we're sorting, we need to explicitly sort by
the id so that we don't get fragile specs.

Change-Id: I9c318fcf4d7384d8351991f18d82b3c43bbcc2b1
Reviewed-on: https://gerrit.instructure.com/29340
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Cody Cutrer <cody@instructure.com>
Product-Review: Derek DeVries <ddevries@instructure.com>
QA-Review: Derek DeVries <ddevries@instructure.com>