Don't complain if asset view score is undefined
Closes CNVS-10914

TEST PLAN
---- ----

I'm really not sure how to reproduce it. We need a student to have
accessed a quiz (viewed it) but the view score to also be nil and not 1
(or 0), then try to view the access report for that student.

Anyway, it's just a guard against nil-values, so the spec should be
sufficient.

Change-Id: Id1fbcb3a36745fb289be73f76c78f1d52656a3b7