Fix links in RCE using IE10
Closes CNVS-11280

The patch for CNVS-5644 (https://gerrit.instructure.com/#/c/29651/) did
not work for IE10 because it relied on IE conditional compiling which is
done using comments, and our optimizer was stripping out those comments.

Sadly, I've had to resort to using eval() to actually run the CC, but I
guess in this case it isn't as evil as the hack itself, so it's okay!

TEST PLAN
---- ----

Please refer to CNVS-5644 for the test plan, verify it works for IE10.

Change-Id: Icb4881b7a8036c34564e3ee379bdbf427898e072
Reviewed-on: https://gerrit.instructure.com/30290
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Jason Madsen <jmadsen@instructure.com>
QA-Review: Caleb Guanzon <cguanzon@instructure.com>
Product-Review: Ahmad Amireh <ahmad@instructure.com>