get rid of some IE 6/7/8 code.
fixes: CNVS-12363

this is all just old crufty stuff that since we don't
support IE <= 8 we don't need.

the only noticeable difference is that if you try to
go to canvas in IE 8 it will give you that
"this browser is not supported" screen. I don't think
it did that before

Change-Id: I9e11eb0925502d1862616a971ce88126c7d5eb67
Reviewed-on: https://gerrit.instructure.com/30409
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Ryan Florence <ryanf@instructure.com>
QA-Review: Nathan Rogowski <nathan@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>