add correct html roles for wiki sidebar links tab
fixes CNVS-8999

This commit adds the correct html roles in the wiki sidebar for tabs. I
also fixed a small css issue that was bothering me. There was a dashed
boarder that looked horrible.

Test Plan
1. Go to any page with a wiki sidebar.
2. Using a screenreader, navigate to the "Links" tab and use the tab to
navigate around the wikisidebar.NOTE: If you are in safari with
voiceover, you need to hit alt+tab to go to the sub items in the links
accordion. This is normal behavior.

Change-Id: I458c7207de00363ebb4dd6d14135fc6736aec340
Reviewed-on: https://gerrit.instructure.com/30940
Tested-by: Jenkins <jenkins@instructure.com>
QA-Review: Matt Fairbourn <mfairbourn@instructure.com>
Reviewed-by: Ryan Florence <ryanf@instructure.com>
Product-Review: Sterling Cobb <sterling@instructure.com>