temporarily raise default_per_page on list courses
this endpoint was originally non-paginated. it was appropriate to make
it paginated (it was documented as paginated, and we cannot leave it
unpaginated), but we should have communicated and given a grace period
for endpoint consumers to prepare first.

temporarily raise the default per_page while they catch up.

fixes CNVS-11574

Change-Id: I853dadd9a093dac555d3a240aa64d275d040b29b
Reviewed-on: https://gerrit.instructure.com/31209
Reviewed-by: Simon Williams <simon@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Miles Wright <mwright@instructure.com>
Product-Review: Rob Orton <rob@instructure.com>
QA-Review: Rob Orton <rob@instructure.com>