fix CanvasCassandra::Database parameter handling
CassandraCQL::Database will take, for example :retries => nil as if
specified and use it to override it's default of 0. but then it will get
sad when trying to treat it as a number.

also, make sure to send the values from the config into this method.

Change-Id: I9df099635283e8455de045e6bbe736ed431f5093
Reviewed-on: https://gerrit.instructure.com/31214
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Brian Palmer <brianp@instructure.com>
Product-Review: Jacob Fugal <jacob@instructure.com>
QA-Review: Jacob Fugal <jacob@instructure.com>