fix module item footer for quizzes
fixes CNVS-11540

when a url was built with the model name it replaced the namespace model
string of "quizzes/quiz" with "quizzes:quiz" so that the slash wouldn't mess
up the url. We need to translate this back before attempting to classify
the quiz string back to the class name.

test plan:
- add a quiz to a module with something before and after it in the module
- navigate to the quiz from the module
- the footer with prev/next links should appear

Change-Id: I559b646fb799f48c6782dc554eb26dca7f6b348b
Reviewed-on: https://gerrit.instructure.com/31215
QA-Review: Myller de Araujo <myller@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Josh Simpson <jsimpson@instructure.com>
Product-Review: Derek DeVries <ddevries@instructure.com>