convert legacy/add_assignment from bundle
the only place it was used as a bundle was in a partial only required by
another partial that includes another bundle that also required this
directly (phew). then once it's not a bundle, attachAddAssignment
doesn't need to be a method on window, and the $(document).ready hook
can move into the requiring bundle.

Change-Id: I5a0059543066933b299d58311e0a7b5cb87235ee
Reviewed-on: https://gerrit.instructure.com/31807
Reviewed-by: Landon Wilkins <lwilkins@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Product-Review: Jacob Fugal <jacob@instructure.com>