A work in progress version of the new ember files app
this commit is meant to be a way to commit some of
the progress I have been making on this.  everything
new is behind a hidden feature flag so it should
not do anything any different for any existing users

if you want to test it out, open a console
and do:
Course.find(<your course id>).enable_feature! :better_file_browsing

then browse to the 'files' part of your course.

the main thing to QA right now is that we did not
break anything for people that don't have that
feature flag turned on.  I know there are a ton of
things that are either broken or still need to be
implemented on the new app.

this is a few commits squashed into one and
closes multiple tickets:

create rails routes for new files UI
closes: CNVS-12140

for the new files UI we are going to use pushState
urls so we had to find a url namespace that we
could completely take over. the thing we chose was
/(courses|groups)/<course_id>/files/folder/foo/bar/baz
(assuming you were looking at the foo/bar/baz folder)

example urls:
/course/1/files/folder/foo/bar/baz
/groups/2/files/folder/lectures/1 Cell Mitosis/handouts/takehome work/
/files
/course/1/files

acceptance criteria:
the following NEW urls SHOULD be routed to new
ember files page:

/(courses|groups|accounts|users)/:id/files/folder*full_path
/dashboard/files/folder*full_path
/files/folder*full_path

the following urls already exist and if the
:new_files feature is enabled SHOULD be routed
also to the new ember files page instead of where
they go now (but if the :new_files feature is not
enabled, they should still go where they go now):

/files
/(courses|groups|accounts|users)/:id/files
/dashboard/files

style mockups for the new files page
fixes CNVS-11216

Enable feature flag for new files page, closes #CNVS-11215
closes: CNVS-12730
closes: CNVS-12659
closes: CNVS-12663

Change-Id: I3868fcf8efd6b517f38f56e88c7bb1e56cf564e8
Reviewed-on: https://gerrit.instructure.com/36005
QA-Review: Clare Strong <clare@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>