create a page_view on a participation
fixes: CNVS-12029

This consolidates the use of log_asset_access so when this is set as
a participation and a page_view does not exist it will create one.
This will keep the asset user access in sync for participations so
analytics and the access reports are not off.

Test Plan:

Create a course with some quizes.  Take the quiz and view the quiz
to generate some activity for a student.

Open the Student Activity report along with analytics for the
student/course.

Make sure the numbers of participations match in each report.

Change-Id: If8b6e7d86e350d5b2a3dd15ca9236c14f52d300a
Reviewed-on: https://gerrit.instructure.com/32382
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Brian Palmer <brianp@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Product-Review: Nick Cloward <ncloward@instructure.com>