add new item link to the top of the list
fixes CNVS-10771

When you add a new module item you, you get this super long list that
has a "New Assignment" etc link in it. That link was previously at the
bottom of the page but it's now at the top.

1. Create a new module item.
2. When making the module item, make sure that the link to create a new
item is at the top of the multiple select section.
3. Check the following:
- Assignments
- Quizzes
- Files
- Content Page
- Discussion

Change-Id: I8e6c9d31cbeb0c87838ad1fe8e0b837151134b9d
Reviewed-on: https://gerrit.instructure.com/32653
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>
QA-Review: Matt Fairbourn <mfairbourn@instructure.com>
Product-Review: Sterling Cobb <sterling@instructure.com>