default subaccount time zone to root account tz
refs CNVS-9312

we don't ever look at the default time zone on a subaccount (or
at least shouldn't). but if we happen to, default to the root account's
default time zone, rather than going straight to Mountain. but if we
happen to look at this field for a subaccount that has set it (rare, can
only be achieved through API, UI disallows it, and it's probably wrong
to do), go ahead and honor it. we may remove that in the future

test-plan:
 - have a root account with Eastern as default time zone
 - create a sub account under that root account
 - in script/console, check value of subaccount.default_time_zone; it
   should match be Eastern
 - set subaccount.default_time_zone to 'America/New_York'
 - reload and check value; should still be 'America/New_York'
 - clear default_time_zone on both subaccount and root account (e.g.
   assign nil to the field)
 - check subaccount.default_time_zone; should be Mountain

Change-Id: Ib32932159db926a9a92dd889ae60c9c12e0a94d9
Reviewed-on: https://gerrit.instructure.com/32789
Reviewed-by: Rob Orton <rob@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>
QA-Review: Nathan Rogowski <nathan@instructure.com>
Product-Review: Jacob Fugal <jacob@instructure.com>