truncate module item title text and add title attr
When working in draft state if you go to the modules page module items
are now truncated to 98 characters. If you hover over the title of a
module item you'll see the default title attribute displaying the title
in the hover box.

fixes CNVS-11999

Test Plan
-----------
As any type of user
With draft state enabled
When I go to the modules page
And I have created a module item with a text header over 98 characters long
I should see that the module item's text header is truncated to 98
characters
And it should have an elipsise (...)
And when I hover over it with my mouse it should show the full text
And it should have the title attribute on it.

Change-Id: I8cfeaac7e0e91290f7788af8b8fced20d52d98a9
Reviewed-on: https://gerrit.instructure.com/32828
Tested-by: Jenkins <jenkins@instructure.com>
Product-Review: Sterling Cobb <sterling@instructure.com>
Reviewed-by: Bracken Mosbacker <bracken@instructure.com>
QA-Review: Nathan Rogowski <nathan@instructure.com>