Add time_zone to courses
closes CNVS-10395

This provides the ability to set the timezone
at the course level in the Course Details
area of the Settings tab.  If there
is no time_zone set, it defaults to the
default_time_zone of the root account. For
this ticket, this value is used nowhere, it just
allows it to happen (and tests are added to
verify that it does happen).  Permissions
are currently granted in tandem with
the ability to rename a course.  If you can rename
it, you can set the timezone. Making
use of this new value throughout the rest of
the application is taken care of in: CNVS-9814

TEST PLAN

1) Log into a course as an instructor.
2) Click on the settings tab.  You should see
the timezone for the associated root account
(if the course timezone won't be set yet).
3) Edit the course to change the timezone and save.
4) the timezone should now be the one selected.

Change-Id: I50e1ed1da283689511795bf018adaa021f42d6f5
Reviewed-on: https://gerrit.instructure.com/34779
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Drew Bowman <dbowman@instructure.com>
QA-Review: Steven Shepherd <sshepherd@instructure.com>
Product-Review: Ethan Vizitei <evizitei@instructure.com>