fix embedded youtube videos in the quiz description
the javascript in instructure.js does a bunch of stuff that we should eventually
move to ember so that it's not this global javascript that magically happens if
a div happens to have a certain class name. Until then, we need to make sure
that the logic in there happens so things don't break.

fixes CNVS-12853

test plan
- as a teacher
  - create a quiz with a link to a youtube video
  - save the quiz
  - enable fabulous quizzes
  - view the quiz
  - the clickable thumbnail that embeds the video should be present

Change-Id: I75e1db2318112c687680c715dbffc6506feefea5
Reviewed-on: https://gerrit.instructure.com/34863
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Jason Madsen <jmadsen@instructure.com>
QA-Review: Caleb Guanzon <cguanzon@instructure.com>
Product-Review: Derek DeVries <ddevries@instructure.com>