fix poll visibility issue for students
fixes CNVS-13374
This commit addresses a mistaken assumption about poll visibility for
students.  Previously, there had to be opened poll sessions for a
student to have read access to the poll. This has been changed to be
based around a students enrollments - if the poll has a poll session
associated with a course or course_section the student is enrolled in,
then they are able to read the poll.

Test plan
- Create a poll as a teacher
- Create a poll session associated with a particular course
  - A student enrolled in the associated course should be able to access
   the poll via #show, regardless of whether or not the session is opened.

  - A student who isn't enrolled in the associated course should not be
   able to access the poll via #show, regardless of whether or not the
   session is opened.

Change-Id: Idf7d4ebeac27d15ce8d37ef498ebc3ab871a6d7e
Reviewed-on: https://gerrit.instructure.com/35722
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Jason Madsen <jmadsen@instructure.com>
QA-Review: Caleb Guanzon <cguanzon@instructure.com>
Product-Review: Josh Simpson <jsimpson@instructure.com>