fix style issues with course title's formatting
fixes #CNVS-12785

The course title was wrapping weird when you were on a wiki front page.
This commit fixes that.

Test Plan ----------------
As a teacher
Given you have a wiki page
And that wiki page is set up as the front page
And the homepage is set to the wiki front page
And you have a very long course title
When you navigate to the course homepage
Then you should see the course title is ellipsed
And the edit and admin gear should be formatted nicely to the right of
it
And you should be able to change the screen size without this breaking

Change-Id: Ie5e267afba57b31c06142d5b93f1da81270725e2
Reviewed-on: https://gerrit.instructure.com/36178
Product-Review: Hilary Scharton <hilary@instructure.com>
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>
QA-Review: Clare Strong <clare@instructure.com>