turn add module item link to a button
fixes #CNVS-13548

When adding a module item the + link should have been a button with a
label. It now is a button with a label and works in JAWs and voice over.

Test Plan ------------
As a user that can create a module item
Given I'm using jaws and/or voice over
When I navigate to the modules page
I should be able to click on the add module items button
And it should be a button
And it should have a label
And it should work in Jaws
And it should work in voice over

Change-Id: I617dc3e37a9aeb2e54ee17a273f8284dbc7787f4
Reviewed-on: https://gerrit.instructure.com/36203
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>
QA-Review: Clare Strong <clare@instructure.com>
Product-Review: Hilary Scharton <hilary@instructure.com>