Make sure no one has the new styles feature enabled
Since the work we are doing here is still very much
WIP, this makes sure no one has the new styles feature
flag turned on

this is a different approach than previous patchsets
since, because of caching or whatever, it was not
working correctly on test portals.

this just renames the feature so if anything was cached
for the previous feature, it won't matter.

and finally, just to clean up, we clear out the db
of any people that had set the old, now meaningless,
feature flag of 'new_styles'

fixes: CNVS-13571 CNVS-14903

test plan:
a) enable the feature flag before checking out the
   patch set
b) check out the patch set and run migrations
c) verify that the feature is no longer set and
   is hidden from non-site-admins

Change-Id: I6230064fe965c80128b047786c393043292114d5
Reviewed-on: https://gerrit.instructure.com/36225
Reviewed-by: Colleen Palmer <colleen@instructure.com>
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
Product-Review: Matt Fairbourn <mfairbourn@instructure.com>
QA-Review: Matt Fairbourn <mfairbourn@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>