enable Audit log search
refs: CNVS-14114
refs: CNVS-13987

Test Plan:

Reference: g/37503

Message in admin tools should be gone and drop down should be visible.
The funcionality should behave as it was before it was disabled.

Hit an Audit log api call and it should work as it was before we
disabled it.

Change-Id: I13eac39aece25dff28b403eb733060e8b3a7f338
Reviewed-on: https://gerrit.instructure.com/37541
Reviewed-by: Jacob Fugal <jacob@instructure.com>
QA-Review: Jeremy Putnam <jeremyp@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>
Product-Review: Nick Cloward <ncloward@instructure.com>