Fix copyright compliance notice compatibility
In the 7/12/14 release, a new Publish button was added to the right
sidebar. This allows users to publish a course directly, therefore
bypassing the original copyright compliance notice placed under the old
Publish Course button.

With this fix, when users click the new Publish button, a confirmation
dialog containing the notice will display.