Send global root_account_id to Kaltura in partner_data
Will allow proper differentiation of media in Kaltura in the future.

There are two ways that partner_data makes it up to Kaltura:

 (1) via a bulk_upload of attachments in the backend
 (2) via the rich text editor upload/record video button

For (1), we globalize the id in KalturaMediaFileHandler. For (2) we
move the domain_root_account_id to js_env as we globalize it.

fixes CNVS-14870

Test Plan

(requires a dev with db access to kaltura to check partner_data)

Kaltura should receive partner_data that includes a global
root_account_id in all of the following contexts:

 - upload a video via RTE
 - record a video via RTE
 - upload a file and add it to a document

Change-Id: Ia4e96ee555465d447015e4cf7a7508e093e51ec2
Reviewed-on: https://gerrit.instructure.com/39496
Reviewed-by: Zach Pendleton <zachp@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>
QA-Review: Steven Shepherd <sshepherd@instructure.com>
Product-Review: Paul Hinze <paulh@instructure.com>