allow saml_logout for SAMLRequest parameter
fixes: CNVS-14869

Adds a method for checking if the request is a SAMLRequest or
SAMLResponse type and will logout a user if its either when saml_logout
is called.

Test Plan:

Ensure that the request to /saml_logout does not get
a 400 when a SAMLRequest parameter or SAMLResponse parameter is supplied
for a SAML configured account.

Change-Id: I6859859cd6518e429e58164aa6076a6a1fdecd51
Reviewed-on: https://gerrit.instructure.com/39474
Reviewed-by: Jacob Fugal <jacob@instructure.com>
QA-Review: Jacob Fugal <jacob@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>
Product-Review: Nick Cloward <ncloward@instructure.com>