Added to code to check for tutorials and labs for non-enrollment sections
Since any sectionCode can be used for enrollment and non-enrollment, we have to create our own priority setting list. Currently we only know of tutorials and labs.

So code now checks to see if an enrollment section has any tutorials and labs. If it has both, only get the the tutorials, else take all of the non-enrollment sections.