let accounts that don't use cas, saml, or ldap set custom login_handle_name
fixes: CNVS-13196 CNVS-14408 CNVS-14960

accounts that have an AccountAutorizationCofig
(those that use cas ldap or saml) are able to set
a login_handle_name (so on the login screen they
could have it say "USU ID #" instead of "Email").
but there are a lot of customers that don't use
cas, saml or ldap that need to set it too.

for example, a school that uses a csv upload for users.

This allows these accounts to set a custom
login_handle_name too.

Test plan:

* for an account that does not use cas, saml, or ldap
* go to the account settings page, you should see
  http://cl.ly/image/1W3B3K392p2B
* set something for "Login Label:"
* go to the login screen
* it should use that new value you set

Change-Id: I5a7fd025229a792db4bf898015752024dafa173d
Reviewed-on: https://gerrit.instructure.com/39628
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Clay Diffrient <cdiffrient@instructure.com>
QA-Review: Clare Strong <clare@instructure.com>
Product-Review: Hilary Scharton <hilary@instructure.com>