Increase link and text color contrast
Fixes: CNVS-14594

Note: The only changes here are to the Canvas
high-contrast UI, not the regular UI.

To enable the high-contrast UI, go to Settings
from the top-right user menu, then scroll down
and turn Use High Contrast Styles to ON. Then
refresh the browser.

Per Dana, the solution to the issue of text color
and link color contrast is to make all text links
underlined by default when Canvas is in high-contrast
mode. When the links are hovered, the underline
should go away -- this is the opposite of the default UI.

**Exceptions are links in the main Canvas menu and
sidebar menu, as well as buttons.**

QA Test Plan:
- Enable high-contrast mode and refresh the browser
- Check to make sure text links in Canvas are now
  underlined by default, like in this example:
  http://cl.ly/image/411a1n082X0b/link-underline-example.gif
- Turn high-contrast mode OFF, and click around to
  make sure links in Canvas are no longer underlined
  by default.

Change-Id: I36dae3536dec288a42cb1912078c5275e761074e
Reviewed-on: https://gerrit.instructure.com/39768
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Colleen Palmer <colleen@instructure.com>
QA-Review: Anna Koalenz <akoalenz@instructure.com>
Product-Review: Chris Hart <chart@instructure.com>