Adds accessibility warning to groups page
Fixes: CNVS-14749

This adds a warning message to the groups page
when first tabbing into the main content. It
informs the users if they are keyboard users, that
there is a MoveTo button in the menu for them to use
to move the students around.

Test-Plan:
1.) Navigate to the groups page
2.) Tab into the main content
3.) Verify that a message appears to inform of MoveTo option

Change-Id: I3e8d860fc3a3a22d9426ff9e856ea87a767a2e12
Reviewed-on: https://gerrit.instructure.com/40510
Reviewed-by: Matthew Wheeler <mwheeler@instructure.com>
Product-Review: Matthew Wheeler <mwheeler@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>
QA-Review: Nathan Rogowski <nathan@instructure.com>