return to an onclick handler for logout
fixes CNVS-15295

In rails2, the `:method => :delete` parameter added an onclick handler
that built up and submitted a form. In rails3, this transparently
changed to using UJS and even hooks, which has been causing problems
with people clicking the link (a) when a js error prevents UJS from
getting setup or (b) before the UJS has a chance to get set up via AMD.

It's important that logout works even in the face of these errors, so
I'm bringing back the onclick handler, but with a form always present
for simplicity.

test plan: The logout link in the top right should still function as
usual. Especially test in IE10 and IE11, where problems with it have
been most prevelant.

(note that there are existing selenium tests for the logout link)

Change-Id: I612884ef261628dace4a24407fce7b84bbd6dd4e
Reviewed-on: https://gerrit.instructure.com/40833
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Cody Cutrer <cody@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Product-Review: Brian Palmer <brianp@instructure.com>