Speed up friendlyDatetime component in react files
the friendlyDatetime component was taking almost
all of the time in rendering newFiles.
I guess some of the tz, $.fudgeDate,
$.datetimestring, etc. functions inside of it were
super slow.

for the folder i was testing it went from ~8000ms
to ~200ms

test plan:

in new files,
in a folder with tons of files/folders load the page
it should be a lot faster than without this commit

another way to tell is to shift-click to select
a bunch of things, it should be fast. before it was
slow.

Change-Id: If8a0b99d0f11a1d2bf3a4e93303b9f510e8d23ef
Reviewed-on: https://gerrit.instructure.com/41020
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Jason Madsen <jmadsen@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>
QA-Review: Ryan Shaw <ryan@instructure.com>