Rename variables to match the keys in the JSON from REST Server
With the introduction of classType in the REST API, sectionCode can now be properly used to distinguish which sections are tutorials/labs for a given enrolment Section.

However, I was re-using the same variable 'section_code' to refer to 2 different things which was causing unexpected behaviour for sections that had no tutorials/labs. This results in Steve's script wanting to delete these sections.