DRY up react propTypes in newfiles
I did this to see if we could remove some of the 
duplicate logic and make it easier to enforce
that we are getting the right propTypes in
stuff in newFiles. I'm interested in your feedback
if you like this more or less.

Change-Id: I6e5093ba68889d15352eb59ed8016e25dbdc526b
Reviewed-on: https://gerrit.instructure.com/41631
Reviewed-by: Dan Minkevitch <dminkevitch@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>
QA-Review: Ryan Shaw <ryan@instructure.com>