a11y: Makes the To autocomplete accessible.
Adds a status and selection field that have an aria-live attribute.  Allowing
the fields to announce changes to it.  Each time a result is
calculated for the autocomplete, the fields are updated and
the screen reader announces the changes.

fixes CNVS-12803

Test Plan:
  - Activate a screenreader
  - Go to /conversations
  - Create a message.
  - When you navigate to the To field and start typing the
    autocomplete should popup.  The screenreader should also
    announce the number of items showing up in the list.

Change-Id: I80c38c412e0ce8eb3022ac26ce1bb733ef4b4b65
Reviewed-on: https://gerrit.instructure.com/41758
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Brad Horrocks <bhorrocks@instructure.com>
Reviewed-by: Mark Severson <markse@instructure.com>
Product-Review: Aaron Cannon <acannon@instructure.com>
QA-Review: Steven Shepherd <sshepherd@instructure.com>