show publish cloud to students in files
fixes CNVS-15768

If you are a student, you need a way to view that a file is only
accessible between certain dates. This commit shows the restricted
access publish cloud to students on files that have restricted dates
and allows them to hover over the icon, view the restricted dates but
they are not allowed to click on the button to change those dates.

Test Plan
As a student that can access new files
Given I have files or folders that have restricted dates
When I go to the filees section
Then I should see the restricted access icon next to those files/folders
And I should be able to hover over them to see the time restrictions
And clicking on them should not pull up the restricted access dialog

This goes without saying, everything should be accessible and
translated.

Change-Id: I0784b8f39ad1d0fa7df439ab07e0d9dedbe220ee
Reviewed-on: https://gerrit.instructure.com/42164
Reviewed-by: Clay Diffrient <cdiffrient@instructure.com>
Product-Review: Cosme Salazar <cosme@instructure.com>
Reviewed-by: Ryan Shaw <ryan@instructure.com>
QA-Review: Anna Koalenz <akoalenz@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>