Feature: Add support for Gmail Actions
This allows Gmail users to be able to click on a message's action
button directly, without having to open the email and press the link
in the footer.

https://developers.google.com/gmail/actions/

This reduces the amount of time and cognitive workload involved in common
tasks, such as checking grades or viewing new assignments.

This also encourages students to enter the Canvas website and view their
courses more frequently.