Verify we are getting the correct error message from the SubmissionsController
This test is to verify that the SubmissionsController properly
rejects assignment submissions that have files attached with illegal
file extensions, based on what is whitelisted by the instructor.

Previously the test was not setting the context of the attachment,
which caused the attachment to not be properly associated with the
user.  Because of that, the submission was properly being rejected
but it was being rejected because there was no file attached,
not because of the file extension.  We also weren't checking the
returned error message, so we thought everything was good when it
wasn't.  This patch set fixes those issues.

Fixes CNVS-16135

Test Plan:

None, this patch set only changes a unit test and doesn't actually
touch application code at all.

Change-Id: Ic263a5fd990db1b600e128fbe78134ed21f039a1
Reviewed-on: https://gerrit.instructure.com/42386
Reviewed-by: Simon Williams <simon@instructure.com>
QA-Review: Simon Williams <simon@instructure.com>
QA-Review: Sean Lewis <slewis@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>
Product-Review: Benjamin Porter <bporter@instructure.com>