handle errors when uploading files
fixes: CNVS-16435

test plan:
in new files, try dragging and dropping a folder
to upload, it. it should give you an error message
that it did not work, *but it should continue working
on the other uploads in the queue*

when QAing this, take the time to be thorough in testing
uploads. try killing your internet connection in the
middle of an upload (it should say the current upload
failed but offer a 'retry' link).

by QAing this thoroughly, it effectively also
closes: CNVS-15495

also fixes: CNVS-16710
Close link is not functional in the failed upload dialog

Change-Id: I28eb54fe70964223cc990a118033a17a70faac70
Reviewed-on: https://gerrit.instructure.com/43536
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Dan Minkevitch <dan@instructure.com>
QA-Review: Jahnavi Yetukuri <jyetukuri@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>