single shard transaction on stream item generation
fixes CNVS-15415

we suspect the accidental use of user_ids when it meant user_ids_subset
(the subset of user_ids local to Shard.current) could contribute to long
run times of this method when user_ids is moderately large and contains
ids from multiple shards. users is a busy table and trying to grab locks
on it unnecessarily is both slow and adverse to other users.

additionally, the only reason for the updates on the users is to
invalidate cache entries; give a little personal space on users that may
have already been updated very recently; it's highly unlikely that the
relevant caches have already been repopulated in those few seconds, and
if they were, a little cache staleness on stream items isn't the end of
the world.

Change-Id: Ibbda9af75aa0e727bd2e7d0a128ee2e3470433eb
Reviewed-on: https://gerrit.instructure.com/44045
Reviewed-by: Cody Cutrer <cody@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>
Product-Review: Jacob Fugal <jacob@instructure.com>
QA-Review: Jacob Fugal <jacob@instructure.com>