Create Outcomes:  Make "New Outcome" go away when clicked
When creating a new outcome, the default title is "New Outcome".  In
almost every case, the user wants to change the title.  When they click
the text box, we will now erase the default title for them.  We only do
this the first time it is created, so if the outcome is saved with the
default title of "New Outcome" we won't wipe out the title the next time
it is edited

Fixes CNVS-16763

Test Plan:

    - Create an outcome
    - Click in the title text box
    - Observe that the title is a placeholder that disappears when you
      start typing

Change-Id: If39c3aaef375254f57d6759fbde801711ddab142
Reviewed-on: https://gerrit.instructure.com/44066
Tested-by: Jenkins <jenkins@instructure.com>
Product-Review: Hilary Scharton <hilary@instructure.com>
Reviewed-by: Simon Williams <simon@instructure.com>
QA-Review: Adam Stone <astone@instructure.com>