allow file previews to have dynamic sizes
fixes CNVS-16871

Previously, if you went to new files and a file to preview, it was
stuck as always being 150px by 100px. Also there were some other css
issues associated with this. This commit fixes that by putting the
arrows on the side of the preview panel, streaching the preview frame
to expand and take up as much space as possible and fixing a few other
minor css things.

Test Plan
Given you have new files enabled
And you are on the new files page
And you have a few image and doc files uploaded
When you click on a file to view its preview
Then the preview should appear
And the next and back arrows should be on the far left/right side
And the center preview should expand to fill up the whole space between
the arrows

Change-Id: I0cbbea63b96f566daf5e2ee8d78d7fe5b36c4824
Reviewed-on: https://gerrit.instructure.com/44845
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Dan Minkevitch <dan@instructure.com>
QA-Review: Anna Koalenz <akoalenz@instructure.com>
Product-Review: Cosme Salazar <cosme@instructure.com>