validations on Version and SubmissionVersion
fixes CNVS-16961

don't create a SubmissionVersion when the Version fails to save. and if
we try to, don't let it be valid. add a data fixup to remove those that
had already happened (all of the instances of this found in production
came from cases where there was no corresponding Version the
SubmissionVersion should have been tied to).

Change-Id: I4e08acf41779ab88de85387b88db155955d3c50e
Reviewed-on: https://gerrit.instructure.com/44916
Reviewed-by: Cody Cutrer <cody@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>
Product-Review: Jacob Fugal <jacob@instructure.com>
QA-Review: Jacob Fugal <jacob@instructure.com>