Improves a11y for new files items cogs
Focus was being lost on modals opened from the
settings menu cog on each item.  This was because
it was trying to focus on the event.target (in this case
a link that was hidden on the screen).  This makes it
so that focus is returned to the settings cog which is
on the screen at all times.

fixes CNVS-17216

Test Plan:
	- Go to New Files
	- Use the cogs at the item level to open modals
	  (Move/Manage Usage Rights/etc)
	- Focus should go back to the settings cog when
	  the modal is closed.

Change-Id: I9bd9ea2a54306747949aa1336f88c5da91ca158b
Reviewed-on: https://gerrit.instructure.com/45232
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Dan Minkevitch <dan@instructure.com>
QA-Review: Nathan Rogowski <nathan@instructure.com>
Product-Review: Clay Diffrient <cdiffrient@instructure.com>