Fixes usage rights dialog styles
This commit also adds a way to specify a class to add to a tooltip.
All you need to do is "data-tooltip-class" and that class will be
applied to the tooltip when it opens.

fixes CNVS-17270

Test Plan:
	- Enable Usage Rights
	- Go to New Files
	- Select a mix of files with different usage rights and folders
          (You need at least 4 or 5 folders with different length names)
	- The warning icon should be spaced away from the warning text.
	- The tooltip text should be left aligned.

Change-Id: Ia5b9686dd870c50eca6c3ed778da8b2ba4305042
Reviewed-on: https://gerrit.instructure.com/45438
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Dan Minkevitch <dan@instructure.com>
QA-Review: Jahnavi Yetukuri <jyetukuri@instructure.com>
Product-Review: Clay Diffrient <cdiffrient@instructure.com>