Stop forcing existing assignments to post_to_sis
The original way this was implemented caused existing assignments show
that post_to_sis was true when the value hadn't ever actually been set.
This caused the feature to appear as on despite the fact that it was
not, this discrepancy caused some confusion as to what the actual state
of the feature was on existing courses.

Test Plan:
- With the Post Grades to SIS feature turned on and fully configured
- Edit an assignment created before the feature was enabled OR set an
  existing assignment's post_to_sis field to nil via the console
- The Post to SIS check box on the edit page should not be set
- Create a new assignment, the Post to SIS check box should default to
  being checked
- Uncheck the Post to SIS box and save the assignment
- Edit the same assignment again
- The Post to SIS check box should remain unchecked

Change-Id: I96783da06eab98b25e8c06ba4e09c517d2ca39ec
Fixes: SIS-744
Reviewed-on: https://gerrit.instructure.com/46374
Reviewed-by: Simon Williams <simon@instructure.com>
Product-Review: Tyler Pickett <tpickett+gerrit@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>
QA-Review: Jeremy Putnam <jeremyp@instructure.com>