Changes wording on usage rights
This commit modifies the wording shown on the restricted access dialog.
It also makes it so that the restricted access dialog is able to be submitted
without specifying a copyright holder.  It makes a small styling
change to fix spacing between the warning icon and the message text as
well.

fixes CNVS-18021
refs CNVS-18015

Test Plan:
   - Enable Usage Rights
   - Upload a file
   - Click to publish it.
   - The warning should say "Before publishing, you must set usage rights on your files."
   - The warning icon should be spaced away from the "B" in Before.
   - You should be able to publish the file without filling in the Copyright Holder field.

Change-Id: I0b0c7e97d13c05459516a34de7a4ad86818178d6
Reviewed-on: https://gerrit.instructure.com/47323
Reviewed-by: Dan Minkevitch <dan@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>
QA-Review: Jahnavi Yetukuri <jyetukuri@instructure.com>
Product-Review: Cosme Salazar <cosme@instructure.com>