add better error message to profile page
fixes CNVS-12169

If you've exceeded you quota for data usage and try to upload a profile
photo, the upload fails and you don't know why. This commit adds an
error message to the page that describes the error.

Test Plan
Given I'm a student
And profiles are enabled for my school
And I have an assignment that requires file uploads
When I upload a file to an assignment that exceeds my upload quota
And I go to my profile
And I attept to upload a user avatar photo
Then it should fail
And I should see an error messsage that explains why it failed
And the "submit" button should revert back to being clickable

Change-Id: I6ae4caf91e14d86ac645040e4e3d86b59123f218
Reviewed-on: https://gerrit.instructure.com/47724
Reviewed-by: Clay Diffrient <cdiffrient@instructure.com>
Tested-by: Jenkins
QA-Review: Jahnavi Yetukuri <jyetukuri@instructure.com>
Product-Review: Sterling Cobb <sterling@instructure.com>