fix teacher interaction report to work with submission comments
The submission_comments table has a `recipient_id` column which the code
was assuming would be the user receiving the comment.  This is actual
a denomralized field, since the comment is also attached to a submission
object and the recipient of the comment would always be the same as the
owner of the submission.  Unfortunately, this field never got populated,
so code that assumed it exist failed.  Rather than fully populate it,
this commit changes the code to query against the submission owner
directly.

fixes CNVS-15315

test plan:
- as a teacher in a course with an assignment
- go to speedgrader and leave a comment for a student
- go to the course roster page and click on the teacher
- click to see "student interactions" in the right sidebar
- there should be a "last interaction time" that matches the time you
  left the comment in speedgrader

Change-Id: Ie786efc7ae073aafd96662f52d2dbd27c27332e2
Reviewed-on: https://gerrit.instructure.com/48207
Tested-by: Jenkins
Reviewed-by: Steven Burnett <sburnett@instructure.com>
QA-Review: Adam Stone <astone@instructure.com>
Product-Review: Simon Williams <simon@instructure.com>