New Canvas UI
fixes:
CNVS-17905
CNVS-19188
CNVS-19185
CNVS-19191
CNVS-19197
CNVS-19187
CNVS-19189

This commit adds HTML/CSS that will serve as the
foundation for the new Canvas navigation and the foundation for
a responsive Canvas UI. It is this base structure we need to
continue to add the rest of the new UI features which include:

- Subnavigation panel
- Avatar in navigation (user sub-navigation)
- Sub-page revisions to bring in new ui
to them (there's a list of these)

ALL these changes are hidden behind the use_new_styles
flag, which no customers have turned on. In order to see them you
will need to turn on the feature flag.

THIS commit should NOT impact either the legacy or
k-12 UIs.

QA Test Plan:

Because this is a large feature to build out,
we are planning to merge it in manageable chunks.

As the feature is not finished yet, this
commit needs only to be tested to make sure it
hasn't broken the current Canvas navigation.

- Make sure the Use New Styles feature flag is off
  in the admin settings.

- Click every link in the current Canvas primary
  and secondary navigation and make sure everything
  still works and looks the same as it does on master.

Change-Id: I082aa0f4db6de143b4d9d4f193ca925de9716589
Reviewed-on: https://gerrit.instructure.com/52131
Reviewed-by: Ethan Vizitei <evizitei@instructure.com>
Tested-by: Jenkins
Product-Review: Colleen Palmer <colleen@instructure.com>
Reviewed-by: Chris Hart <chart@instructure.com>
QA-Review: Steven Shepherd <sshepherd@instructure.com>
Reviewed-by: Derek DeVries <ddevries@instructure.com>