capture replies before save; re-renderTree after
fixes CNVS-5128

This bug was caused by the removal of the replies array on the entry
object before saving the entry. (They are being removed because the
reply holds an instance of its parent in its attributes, as opposed to
just a reference to the parent. This causes a circular json error when
trying to prepare the model to send to the server. I'm not sure I agree
with the way we keep an instance of the parent, and it clearly runs
counter to what Backbone expects, but that is an issue for another day.)

After saving and re-rendering the EntryView, the replies were not present,
because they had been removed. I experimented with only sending up the
attributes that have changed, which would in theory prevent any issue
with leaving the replies on the entry, but this is not practical until
the API supports PATCH requests (or we want to significantly hack
Backbone.sync or Model#save).

Instead, I've udpated Entry#sync to track the replies array, if it is
present, and re-set the array after the API request is complete. Also,
EntryView listens on 'change:replies', and responds depending on what
the value provided is (if empty, remove the reference to the treeView;
if a colleciton, call `renderTree`.

Finally, I made a change to a selenium helper method used for replying
to discussion topics or entries. The method had been setup such that it
would always only reply to the topic, which seems counter to the intent
of the method (since it accepts an argument to target a specific topic
OR entry to respond to).

test plan:
- Create a discusion.
- Reply to the discussion topic.
- Make a comment on the reply in Step 2.
- Edit the discussion reply in Step 2.
- Observe that the form is removed, the updated entry is present, as are
  the entry's replies.

Change-Id: Iae98b06415e1617720b0a345ccd685bfe5cc5ae6
Reviewed-on: https://gerrit.instructure.com/49903
Tested-by: Jenkins
Reviewed-by: Simon Williams <simon@instructure.com>
QA-Review: Adam Stone <astone@instructure.com>
Product-Review: Simon Williams <simon@instructure.com>