Don't strip out lang attributes when sanitizing html
fixes CNVS-18648

This was suspected to be a tinymce problem,
but tiny wasn't stripping lang at all (which
is why the reporter was confused that it was
in the white list).

This adds rspec (and therefore specs)
and the sanitize dependency
to the canvas_sanitize gem, adds a spec
that proves the sanitization problem was in here,
and then changes the default config to allow the
lang element to show up on html elements.

TEST PLAN:
 - Open an RCE in Canvas
 - Use the HTML editor and paste in <p lang="es"> Hola. Como estas?</p>p>
 - Save
 - Edit and verify that the lang attribute hasn't been stripped

Change-Id: Iff9129a8cbff4b4efd004e619ef92c7395dda568
Reviewed-on: https://gerrit.instructure.com/50255
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
Product-Review: Jeremy Stanley <jeremy@instructure.com>
Reviewed-by: Dan Minkevitch <dan@instructure.com>
Tested-by: Jenkins
QA-Review: Jahnavi Yetukuri <jyetukuri@instructure.com>