TimeBlock accessibility
closes CNVS-19314

refs CNVS-15543

This applies a screenreader-only label to the date picker and
timepickers used in the scheduler, and some KB-only user tooltips
(though they are visible for all users).  The goal is to make
it clear what format to input dates and times in to users
for whom the rich selector isn't so great.

TEST PLAN:
- go to the scheduler (make sure scheduler is enabled, go to calendar,
   use scheduler menu)
- schedule an appointment group
- Using a screenreader with your eyes closed, ensure you can figure out
  what date format to submit for the appointment group with no
  other context.  Same for time fields next to it.
- tab through to the same fields and prove that you see
  a little tooltip telling you what format to enter the date and time
in.

Change-Id: Idce5cdcfc4169ee538083d745ddd44de67bf6846
Reviewed-on: https://gerrit.instructure.com/50859
Tested-by: Jenkins
Reviewed-by: Ryan Shaw <ryan@instructure.com>
QA-Review: Steven Shepherd <sshepherd@instructure.com>
Product-Review: Ethan Vizitei <evizitei@instructure.com>