make focus accessible for srgb set default grade modal
fixes CNVS-19218

When a user clicks "set default grade" in the individual gradebook
view the focus is set on the close modal button, and when they
click on "set default grade" in that modal, the focus returns to
the "set default grade" button on the page. When they click the
close button in that modal, focus returns to the "set default grade"
button on the page.

Test Plan:

1. In individual view gradebook, select an assignment.
2. Click "set default grade" and confirm focus is on the close
    button.
3. Click "set default grade" in that modal and confirm that focus
    returns to the "set default grade" on the page.
4. Click "set default grade" and then the close button and confirm
    that focus returns to the "set default grade" on the page.

Change-Id: I6cd16c581c7c6a1d428be5a4ef91d85b4d777a83
Reviewed-on: https://gerrit.instructure.com/51051
Tested-by: Jenkins
Reviewed-by: Jonathan Featherstone <jfeatherstone@instructure.com>
QA-Review: Nathan Rogowski <nathan@instructure.com>
Product-Review: Strand McCutchen <smccutchen@instructure.com>