Fixes possibly undefined variable use in SRGB
From time to time, visibilites ends up being undefined.  This
uses the ? operator to make sure that it works even when it is
undefined.  It was causing other JS in the stack at the time
to be skipped, thus leading to other problems.

This commit also fixes the focus handling for when that dialog
opens.  It was attempting to be set to an element that was not
able to be focused.  This makes it focus the close "X" to make
it consistent with current Canvas focus guidelines.

fixes CNVS-19215

Test Plan:
  - Go to Individual View Gradebook
  - Select a student and an assignment
  - Click the submissions details button
  - When opening the focus should be on the close button in the
    corner.
  - After updating the grade or adding a comment and clicking
    the appropriate submit button, the dialog should close.

Change-Id: I46d356be8737cbde5810f7452d0daf428f9a860f
Reviewed-on: https://gerrit.instructure.com/51065
Reviewed-by: Cameron Sutter <csutter@instructure.com>
Tested-by: Jenkins
QA-Review: Nathan Rogowski <nathan@instructure.com>
Product-Review: Clay Diffrient <cdiffrient@instructure.com>