Stop error report creation from failing on encoding mismatch
fixes CNVS-19610

On several development machines (but not production apparently) the
request.env hash has strings in it that are ASCII_8BIT
which is usually not a big deal.  However,
in ErrorReport we serialize part of that hash, and we have
some code in "gems/utf8_cleaner" that makes sure we don't bomb on writes
because of bad UTF8 characters by force-encoding things we're going to
serialize.  The problem is that the strings in the request env hash are
(rightfully) frozen.  To make this not bomb, we need to deep-copy
everything we take directly from that hash, which this commit does
(after first writing a spec to expose the problem)

Change-Id: I53ef27ba2ea1fca9a3e5bd91c2840479bdb45fe3
Reviewed-on: https://gerrit.instructure.com/51247
Tested-by: Jenkins
Reviewed-by: Cody Cutrer <cody@instructure.com>
QA-Review: Jeremy Putnam <jeremyp@instructure.com>
Product-Review: Ethan Vizitei <evizitei@instructure.com>