add a parameter to user creation for forcing self-registraion flow
closes CNVS-19046

Marketing is moving to take over the FFT form, which means
the user creation will occur via API, but they'll still need
to get the self-registration style emails to get passwords right
and such.

This provides a parameter to do so.  It also pulls out a handful
of lines from the users_controller all related to how and which
notifications get sent, and moves them into a policy object, which
has some predicate methods for the controller to make decisions off
of, and also does the notification dispatching.

Change-Id: I0513062aedf4258b7d9f51a46b1d8e42b2a2582a
Reviewed-on: https://gerrit.instructure.com/51608
Tested-by: Jenkins
Reviewed-by: Jacob Fugal <jacob@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Product-Review: Ethan Vizitei <evizitei@instructure.com>