Say "web browser" instead of "internet browser"
The term "Internet browser" is not used professionally because is it not
technically correct. Many programs use the Internet: E-mail clients such
as Thunderbird and Outlook, instant messangers, smartphone apps, game
consoles, and much more. Web browsers are just another kind of program
that uses the Internet; they use it specifically to access the
collection of URI-accessible documents known as the World Wide Web.