fix anonymous peer review checkbox on graded discussion form.
fixes CNVS-19782

The checkbox used to determine whether or not an assignment's peer
reviews should be considered anonymous, which is part of a backbone
view, did not function properly if it was 'nested' (i.e., if the name on
the field should be `assiginment[anonymous_peer_reviews]`, instead of just
`anonymous_peer_reviews`. The `checkbox` helper is already setup to
handle this; we just needed to pass the `nested` value to it, and to
update how we determine whether or not the checkbox should appear
checked, so as to avoid manually setting the checkbox's `name`
attribute.

Additionally, this patch fixes a bug in the
`RubricAssessment#considered_anonymous?` method, which would blow up if
the model did not have a related rubric_association. I've added a
protection to two other methods that would have thrown exceptions in the
absence of a rubric_assocation as well.

Finally, I updated the logic in the SubmissionComment policy
declaration, so that when trying to `:read_author`, a student can always
see the name of a teacher, even if the comment is marked as anonymous.

test plan:
- as a teacher, create a graded discussion topic. When doing say, enable
  peer reviews, and enable anonymous peer reviews.
- as a student, submit an entry for the graded discussion topic.
- as a teacher, assign a peer review to a student for the submission
  above.
- as the reviewer student, submit a review.
- as the reviewed student, navigate to the submission details page
  (/courses/:course_id/assignments/:assignment_id/submissions/:user_id)
  and observe the comment list on the right hand side of the page.
- observe that the name of the reviewer student is not present, and
  instead it says 'Anonymous User'.
- as the teacher, navigate to the same submission page, and observe the
  comment list on the right hand side of the page.
- observe that the name of the reviewer student IS present.
- as the teacher, navigate to the same submission page, and observe the
  comment list on the right hand side of the page
- as the teacher, leave a comment on the submitted assignment.
- as the reviewed student, navigate to the submission details page, and
  observe the comment list on the right hand side of the page.
- observe that the name of the reviewer teacher IS present.

Change-Id: I572dd7fa319cd784e59f00057898fcea5349c899
Reviewed-on: https://gerrit.instructure.com/52099
Tested-by: Jenkins
Reviewed-by: Mike Nomitch <mnomitch@instructure.com>
QA-Review: Adam Stone <astone@instructure.com>
Product-Review: Hilary Scharton <hilary@instructure.com>