Make term editing dates accessible
closes CNVS-19316

refs CNVS-15543

see the description and comments of CNVS-15543 for full narrative

This applies a screenreader-only label to the date pickers in
the "edit term" section that gives formatting information,
and also adds an on-focus
tooltip for KB-only users that gives format
information.

TEST PLAN:
- go edit any term (/accounts/x/terms as an admin)
- make sure as a screenreader user you can understand what format to
  submit your dates in, there should be 10 fields
- tab through to the date/time fields and prove that you see
  a little tooltip telling you what format to enter the date and time
in.

Change-Id: I96e69dea9e071f488b03e3cc7da7f55944760441
Reviewed-on: https://gerrit.instructure.com/52729
Tested-by: Jenkins
Reviewed-by: Jason Madsen <jmadsen@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Product-Review: Ethan Vizitei <evizitei@instructure.com>