Deleted upload submissions no longer download
When zipping assignment submissions for download, deleted files should not be
added to the zip.

fixes CNVS-12167

There is an edge case that is covered - that of a small course where maybe 1 or
2 submissions may exist, and the file for each has been deleted. By marking
success if the file is deleted, we keep it from erroring out, and will download
an empty zip. Jenkins shows no regressions for this, so we should be good.

Test Plan:
  A: As a teacher:
    1. Create a file-upload assignment as a teacher
  B: As a student:
    1. Submit a file upload for the assignment
    2. Delete the upload (/users/:user_id/files)
  C: As a teacher:
    1. View the students submission in speed-grader
    2. Click download icon
    3. Verify that a page not found error is displayed
    4. Return to assignment page
    5. Click "Download Submissions"
    6. Verify that the zip file downloaded
    7. Double-click on zip file
    8. Verify that the resulting file does not contain the deleted file
       The resulting file may simply be a .cfgz or similar file. This signifies
       an empty zip file in this case.

Change-Id: I6344cadb92d4e13bd23ade4c4e34111751bcecd1
Reviewed-on: https://gerrit.instructure.com/53574
Tested-by: Jenkins
Reviewed-by: Strand McCutchen <smccutchen@instructure.com>
QA-Review: Adrian Foong <afoong@instructure.com>
Product-Review: Cody Poll <cpoll@instructure.com>