give feedback on bad email formats in password reset
closes CNVS-19932

If the user submits something obviously not-email-ish, we've
just said "sure we're sending you an email" before, which is good for
security but bad for user experience.  This change gives some feedback
to the user if the submission is not an email, but makes no remarks
about
whether it's an email we know about on any account anywhere.

also changes some of the documentation on our "formSubmit" jquery
plugin which was shown to be inaccurate

TEST PLAN:
 1) Access any canvas auth login page and click "Don't know your
password?"
 2) hit the spacebar, then enter
 3) notice message telling you that's not an email
 4) make sure you can still reset your password with your real email

Change-Id: I87e57e9c71298b11d474d58a2615a95d0ce70021
Reviewed-on: https://gerrit.instructure.com/54400
Tested-by: Jenkins
Reviewed-by: Rob Orton <rob@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Product-Review: Ethan Vizitei <evizitei@instructure.com>