don't use symlinks to @include bower/vendor css
the new version of node-sass that we are using
understands how to @impor files that have .css
extensions directly.  we used to have to do a hack
where we create a symlink to the .css file from
a file with a .scss file extension. but we
don't have to do that now

Change-Id: I81f2e57a33aaa1e19e03f2d6c6d9cc0679d4bce1
Reviewed-on: https://gerrit.instructure.com/54807
Reviewed-by: Ethan Vizitei <evizitei@instructure.com>
Tested-by: Jenkins
Reviewed-by: Colleen Palmer <colleen@instructure.com>
Product-Review: Colleen Palmer <colleen@instructure.com>
Reviewed-by: Jacob Fugal <jacob@instructure.com>
QA-Review: Ryan Shaw <ryan@instructure.com>