spec: fix fragile timezone TimeBlockRow spec
when the browser timezone is later than Detroit (e.g. UTC on a
continuous integration box), and the time is pre-midnight Detroit but
post-midnight in the browser timezone, then the former construction of
a "fudged midnight" would choose the wrong day (the UTC date instead of
the Detroit date), and as such would come later than the current time
instead of the intended earlier.

while fixing, also clarify the behavior with appropriate variable names,
and fix the spec (and other similar specs) to produce failed tests
instead of uncaught errors. (when validation is expected to fail, but
doesn't, the associated_error_box won't be there.)

test-plan:
 - wait until after 8am Detroit time (or set system clock)
 - set system timezone to Auckland (New Zealand)
 - specs still pass

Change-Id: Ifb587455fbcdb76dbefaac8b17ff2b3ad76e4da9
Reviewed-on: https://gerrit.instructure.com/56019
Reviewed-by: Simon Williams <simon@instructure.com>
Tested-by: Jenkins
Product-Review: Jacob Fugal <jacob@instructure.com>
QA-Review: Jacob Fugal <jacob@instructure.com>