fixes gradebook cell mis-alignment
the total_grade column's width was not being
parsed as an integer, therefore when adding all
the column's widths together to get the
width of the gradebook, the width was
being returned as a string. this caused
the gradebook to have mis-aligned columns
and prevented editing of grades until the
total_grade column was resized.

closes CNVS-21117

test plan
  - resize the total_grade column in gradebook
  - reload the page
  - the gradebook should load properly:
    - the cells should be editable
    - the cells should have proper styling

Change-Id: Id202817ab649f88255218e4d63b9ded9ce79556e
Reviewed-on: https://gerrit.instructure.com/56157
Tested-by: Jenkins
Reviewed-by: Josh Simpson <jsimpson@instructure.com>
QA-Review: Adrian Foong <afoong@instructure.com>
Product-Review: Josh Simpson <jsimpson@instructure.com>