don't let course notifications bomb with no prior_version
closes CNVS-21124

Course notifications for catalog were not getting sent out.
Upon sending them manually to debug, we found that courses with
no "prior_version" were throwing an exception from the predicate
block that evaluates whether to send the new_course notification.

This bombs the whole job (catalog sends notifications once an hour,
and a single exception stops the job from running).

This is not usually a problem when sending notifications in a save
cycle, but these records are being loaded and broacasted with no
save, which is why they break in this case.

TEST PLAN:
 1) load a course from the database but don't save it
 2) call "broadcast_notifications"
 3) it should not throw an exception

Change-Id: Ia9a07080b5b54e4f732fb51722a678650cce3b76
Reviewed-on: https://gerrit.instructure.com/56186
Reviewed-by: Jacob Fugal <jacob@instructure.com>
Tested-by: Jenkins
Product-Review: Bryan Petty <bpetty@instructure.com>
QA-Review: August Thornton <august@instructure.com>