replace slashes with underscores in gradebook export filename
When using local storage instead of S3, having a course short name with a slash in it will cause an error (No such file or directory @ rb_sysopen) when trying to export the Gradebook as a CSV. attachment_fu is receiving the slash as part of the file path, and erroring out because it can't find the file.

This is a band-aid fix until I can fix it properly in AttachmentFu and submit it back to Instructure.

Test plan:
  create a course with a slash in the short name
    (e.g. CHEM 121 / CHEM 122)

  attempt to export the gradebook
  ensure the gradebook file was exported and
    downlaoded correctly

Signed-off-by: Graham Ballantyne <grahamb@sfu.ca>

Conflicts:
	app/controllers/gradebook_csvs_controller.rb