make content migration date shifting
closes CNVS-19335
refs CNVS-15543

This applies a screenreader-only label to the datetime pickers
used in the content migrations UI, and some KB-only user tooltips
(though they are visible for all users).  The goal is to make
it clear what format to input dates and times in to users
for whom the rich selector isn't so great.

TEST PLAN:
- go to make a content migration (/courses/x/content_migrations)
- setup a course copy and indicate you want to shift dates
- Using a screenreader with your eyes closed, ensure you can figure out
  what date format to submit for the date shifting with no
  other context.
- tab through to the same fields and prove that you see
  a little tooltip telling you what format to enter the date and time
in.

Change-Id: I91cb4f471b735e6954e1a55923167c2b151b8015
Reviewed-on: https://gerrit.instructure.com/56709
Reviewed-by: Jacob Fugal <jacob@instructure.com>
QA-Review: Jeremy Putnam <jeremyp@instructure.com>
Tested-by: Jenkins
Product-Review: Ethan Vizitei <evizitei@instructure.com>