show green banner on successful gradebook upload
when a gradebook is successfully uploaded, a
green success banner will show.

closes CNVS-20359

test plan:
import a Gradebook CSV (with changes made to scores).
verify that after the gradebook is done uploading,
you are redirected to the Gradebook and there is a
green banner at the top of the page that says
'Upload successful'. refresh the page and verify that
a green banner does _not_ appear with the success
message.

Change-Id: I29346498a3c1ae7158e7ac2f11de3f7ca5b8d806
Reviewed-on: https://gerrit.instructure.com/57062
Tested-by: Jenkins
Reviewed-by: Cody Poll <cpoll@instructure.com>
QA-Review: Derek Hansen <dhansen@instructure.com>
Product-Review: Spencer Olson <solson@instructure.com>