fix rce image linking in Firefox
closes CNVS-21204

make firefox treat rce image linking appropriately.  Prior to this
patchset it would make the alt text a link instead of wrapping the image
tag in a link tag.  The reason is the way we decide whether to wrap
DOM content, or make a new text link and replace existing DOM.  In the
case of some DOM element which has HTML, but for which
"editor.selectContent({format: 'text'})" returns a blank string, we've
usually interpreted that as something image-like that should probably
just be wrapped.  The problem is that Firefox gives us the alt
text when we ask for the "text" of that node, throwing off our
check.  This patchset makes the text extraction explicitly check
for the case of an image node with alt text and treat that the same
for the purpose of content sniffing, and seems to fix the problem.

TEST PLAN:
 1) use firefox
 2) go to any RCE
 3) embed an image
 4) now select the image and try to make it a link
 5) it should work

Change-Id: I65695b1d73d3c4855c8d63848578ba94a4c5c672
Reviewed-on: https://gerrit.instructure.com/57181
Tested-by: Jenkins
Reviewed-by: Jacob Fugal <jacob@instructure.com>
Product-Review: Ethan Vizitei <evizitei@instructure.com>
Reviewed-by: Ryan Shaw <ryan@instructure.com>
QA-Review: August Thornton <august@instructure.com>