(NUI) Fix broken permissions and files layouts
Fixes: CNVS-23435, CNVS-23601

Scoping the new UI flexbox layout to the absence of
the .full-width body class broke a couple of layouts,
so we are adding a new class to full-width pages
in Canvas (speedgrader, conversations) that break
when the flexbox layout is added.

This will fix the broken Files and Permissions layouts.

Test plan:
- Turn on the New UI
- Go to [your course] > Files. Make sure the layout
  is back to normal (two columns)
- Go to Admin > Permissions. Make sure the layout
  is back to normal (two columns)
- Also check Speedgrader and Conversations to make
  sure they're working as usual.

Change-Id: I2119d6a49d95d7a7d7ebc792bb2fe014f3ef8da2
Reviewed-on: https://gerrit.instructure.com/63929
Reviewed-by: Jennifer Stern <jstern@instructure.com>
Product-Review: Jennifer Stern <jstern@instructure.com>
Tested-by: Jenkins
QA-Review: Nathan Rogowski <nathan@instructure.com>