Fix Canvadoc.rb previewing of peer reviews
Fixes CNVS-23648

This fixes a bug where Canvadoc.rb would fail to generate previews
for files with any type of peer-review set. This is most easily
seen with RTF files which is one file type Canvadoc uses to
render. The cause for this error was due to a misspelling of
"anonymous" which was instead spelled: "anonymouis". As such
the spelling is now corrected, and has been checked to make sure
it lines up with the assignment model.

Change-Id: If54858646b9c2e0882ff8b444d17cba01344ecf0
Reviewed-on: https://gerrit.instructure.com/63997
Reviewed-by: Derek Bender <djbender@instructure.com>
Tested-by: Jenkins
QA-Review: Caleb Guanzon <cguanzon@instructure.com>
Product-Review: Keith T. Garner <kgarner@instructure.com>