fix visibility of total column tooltip in gb2
The tooltips for the total column in gb2 were
being cut off on the left when the total column
were moved to the front. Added some css to
re-size the tooltip box so that it is always
visible.

fixes CNVS-23787

Test Plan:
1. Create a course with at least 1 student
2. Create an assignment and mute it
3. Move Total column to the front of the Gradebook
4. Hover over warning icon in Total column and
notice the position of gradebook-tooltip-last
5. Verify tool tip is not cut off on the left and is
visible.

Change-Id: I1a2387b31a0945a82922d494246625a7e3a50a07
Reviewed-on: https://gerrit.instructure.com/66151
Tested-by: Jenkins
Reviewed-by: Strand McCutchen <smccutchen@instructure.com>
QA-Review: Jason Carter <jcarter@instructure.com>
Product-Review: Christi Wruck