spec: fix coffeescript module namespaces
in qunit, tests run in the namespace of the last module call that
happened. so if you say:

  module('module1');
  test('test1', function() { ... });

then test1 runs in the namespace of module1.  you can pass module1 an
object as a second argument which are functions that will be defined in
the module namespace, and some modules (like setup and teardown) have
special functionality.  unfortunately, you can also pass a test in the
module object and it will happily run, but in the namespace of the
previous module definition.

this is all exacerbated by coffeescript! because in coffeescript:

  module 'module1'
  test 'test1' -> ...

is correct, but

  module 'module1',
  test 'test1' -> ...

runs the test in the previous namespace.  notice the difference? ONE
COMMA! I did an audit of these instances and fixed as many as I could
fine, as well as one legit test issue that came out of fixing them.

Change-Id: I79ea0c673df49e6c09151b50cf765ed9ec760a0d
Reviewed-on: https://gerrit.instructure.com/70950
Reviewed-by: Ethan Vizitei <evizitei@instructure.com>
Reviewed-by: Jon Jensen <jon@instructure.com>
Product-Review: Jon Jensen <jon@instructure.com>
QA-Review: Jon Jensen <jon@instructure.com>
Tested-by: Jenkins