escape student names that begin with equals sign in gradebook exporter
When exporting to a CSV when a student name beings with an equals, the
gradebook exporter will set the name to be a forumula that explicity
sets the cell to name. This will allow a teacher using Excel to be
safe from potential security issues caused by a student changing their
student name name to a function or macro.

This commit also contains a fix to an existing spec that didn't test
what it was intending to.

fixes CNVS-25317

test plan:
 - Have a course and gradebook that has a student with the name of
   =sum(A)
 - Export the gradebook to a CSV.
 - Open the CSV in a text file and note that the name looks like
   ="=sum(A)"
 - Open the CSV in excel and note the student name is properly
   displayed in the grid as =sum(A)

Change-Id: I1ff77f1c49f93d997cb5dcde8ad0d2239edbc4c7
Reviewed-on: https://gerrit.instructure.com/72914
Reviewed-by: Spencer Olson <solson@instructure.com>
Reviewed-by: Derek Bender <djbender@instructure.com>
Tested-by: Jenkins
QA-Review: KC Naegle <knaegle@instructure.com>
Product-Review: Keith T. Garner <kgarner@instructure.com>