Take out wide breakpoint in Inbox
refs: CNVS-27362

We need to get rid of the "wide" breakpoint variable here, but considering
this page has a layout that would break with our current desktop breakpoint
(which is going to be re-visited soon), we just hardcoded 1200px in for
the time being. So there should be no changes visually in this commit.

Test plan:
- Go to inbox
- Nothing should change in the UI

Change-Id: Ic40afcc672c55701c28c733ae0bae917cd7287b2
Reviewed-on: https://gerrit.instructure.com/74032
Tested-by: Jenkins
Reviewed-by: Pam Hiett <phiett@instructure.com>
Product-Review: Pam Hiett <phiett@instructure.com>
QA-Review: Nathan Rogowski <nathan@instructure.com>