use a button for "Show All Details" on Student grades
When a screen reader user logged in as a student navigates to the Grades
page of a course and focuses on the "Show/Hide All Details" button, it was
previously read as a link. It should be read as a button. This change
implements that.

closes CNVS-28045

test plan:
1. In Safari with VoiceOver enabled, log in as a Student
2. Navigate to the "Grades" page of a Course
3. Move VO focus to the "Show All Details" button
4. Verify that the button is read as a button

Change-Id: I7b7a86194c11eb97a237879c3d607c1170b35944
Reviewed-on: https://gerrit.instructure.com/75890
Reviewed-by: Spencer Olson <solson@instructure.com>
Reviewed-by: Derek Bender <djbender@instructure.com>
Tested-by: Jenkins
QA-Review: KC Naegle <knaegle@instructure.com>
Product-Review: Christi Wruck