update nil checking pattern for grading standards
for the grading standard partial view, three places had
different patterns for checking whether the object passed to
the partial was nil despite the partial checking for nil itself.
this commit removes the redundant checking and relies on the
checking in the partial. also, the nil checking and use
of variables in the partial were cleaned up.  this cleanup also
fixed an issue where the name of the default grading scheme was
not being shown despite the data being shown.  this was seen
in the Edit Discussion Topic and Course Settings pages.

closes CNVS-29329

test plan:
- create a course if necessary
- navigate to the "Course Details" tab of the "Settings" page
- set the "Enable course grading scheme" checkbox
- click the "view grading scheme" link that appears
- ensure that "Default Grading Scheme" is the title of the table
- create a discussion if necessary
- edit the discussion
- set the "Graded" checkbox in the "Options" section
- select "GPA Scale" from the "Display Grade as" dropdown
- click the "view grading scheme" link that appears
- ensure that "Default Grading Scheme" is the title of the table
- create an assignment if necessary
- edit the assignment
- click the "More Options" button
- select "GPA Scale" from the "Display Grade as" dropdown
- click the "view grading scheme" link that appears
- ensure that "Default Grading Scheme" is the title of the table

Change-Id: Iae0b317ec2bc95e0a94eba1c2182022732ea898b
Reviewed-on: https://gerrit.instructure.com/79422
Tested-by: Jenkins
Reviewed-by: Spencer Olson <solson@instructure.com>
QA-Review: KC Naegle <knaegle@instructure.com>
Product-Review: Keith T. Garner <kgarner@instructure.com>