spec: (hopefully) fix flaky spec
sometimes selenium doesn't click on the submit button and instead just
focuses. click_option doesn't blur the dropdown, and it's still open
until the next interaction. although it doesn't overlap the button,
sometimes it causes the click to misfire (perhaps a subtle dom-shift
when the change event fires?). if we blur out of the select before we
submit, in theory it should be able to click it every time. in theory.

Change-Id: Ic1160c8c1d6d3f44d2c77129acdb52058a10f541
Reviewed-on: https://gerrit.instructure.com/79632
Tested-by: Jenkins
Reviewed-by: Landon Wilkins <lwilkins@instructure.com>
Product-Review: Jon Jensen <jon@instructure.com>
QA-Review: Jon Jensen <jon@instructure.com>