spec: fix more specs that fail when newUI is on
refs: CNVS-29532

ENV['CANVAS_FORCE_USE_NEW_STYLES'] is true in the
jenkins hourly NewUI only build. so each of these
were failing there.

as far as the formatting of these changes, I tried
to do it in a way to make it so the process of
removing the oldUI code path this summer is explicit,
easy and clear.

Change-Id: I89dc6ad6add44234c29b52ea786888e9b74481e4
Reviewed-on: https://gerrit.instructure.com/80706
Tested-by: Jenkins
Reviewed-by: Matt Fairbourn <mfairbourn@instructure.com>
QA-Review: Matt Fairbourn <mfairbourn@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>