ensure all bundles are picked up by rjs, refs SD-1235
anything in a subdir (e.g. /legacy/) would not get bundled up. webpack is
imminent (in theory) and already does this correctly, but this is still a
good thing to do in the meantime so that:

* we speed up page loads (no more async, dependencies now get uglified)
* users don't load stale dependencies post-deploy (they aren't
  fingerprinted)

test plan:
1. canvas should work
2. when running optimized, any pages that use legacy bundles should just
   load a single js files, instead of a whole bunch through requirejs

Change-Id: I2004310cc0675a0ebeb49cdc70d8909e28cfb3b1
Reviewed-on: https://gerrit.instructure.com/82322
Tested-by: Jenkins
Reviewed-by: Ryan Shaw <ryan@instructure.com>
Product-Review: Jon Jensen <jon@instructure.com>
QA-Review: Jon Jensen <jon@instructure.com>