alert screenreader on close discussion reply
fixes CNVS-29499

Test Plan
in Firefox with NVDA on, navigate to a discussion with a reply
Move focus to the Collapse bar at the top of a reply.
Press Enter and note that, while the reply does collapse, the SR
is notified.

Change-Id: I65efaf7c4259b46d7978098ec068cc59ade77913
Reviewed-on: https://gerrit.instructure.com/83246
Reviewed-by: Felix Milea-Ciobanu <fmileaciobanu@instructure.com>
Tested-by: Jenkins
Product-Review: Steven Burnett <sburnett@instructure.com>
QA-Review: Steven Burnett <sburnett@instructure.com>