bump inst-jobs to 0.11.3
fixes CNVS-30389

test plan: this may be difficult to test, but the rough idea is that if
you have a job that is queued for shard X in the future, then shard X is
deleted, and then the job tries to run.  it should fail, but not kill
the job worker.

Change-Id: Ib2d7858e3da61366eac4ea89482a9696773eaf33
Reviewed-on: https://gerrit.instructure.com/85151
Reviewed-by: Cody Cutrer <cody@instructure.com>
Tested-by: Jenkins
QA-Review: Benjamin Christian Nelson <bcnelson@instructure.com>
Product-Review: Simon Williams <simon@instructure.com>