Clarify suggestions for using a supported bundler version
If the wrong bundler version is in use it will always be helpful
to suggest explicitly installing a version that will work.

If a newer bundler version is in use then we should also explain
how to use the appropriate version (and avoid the newer version).

Additionally bundler outputs a snippet of the code when it exits early
and having the message in a variable makes that snippet less confusing
to see.

This should help avoid issues like gh-855.

Test Plan:
  - Start with new ruby installation or uninstall all bundler versions
  - Install too-old version: `gem install bundler -v 1.7.0`
  - `bundle install`
    - Should see a message explaining version in use vs desired
      with suggestion to install a specific version.
  - Install desired version: `gem install bundler -v 1.11.2`
  - `bundle install`
    - Should work
  - Install too-new version: `gem install bundler`
    - Should see message explaining version in use vs desired
      with a suggestion to install a specific version
      then instructions on how to use the older version:
      either specify the version desired or uninstall the too-new
      version.
  - Run the command suggested: `bundle _1.11.2_ install`
    - Should work
  - Run the suggested "gem uninstall" command
  - `bundle install`
    - Should work