display correct text on button after comment submission in speedgrader
In the recent Speedgrader redesign the wording on the submit comment
button was changed to submit. It was changed in the erb, but missed in
the javascript. This updates the javascript to match.

fixes CNVS-30526

test plan:
 - Load speedgrader for an assignment for a student
 - Submit a comment for that student
 - Notice the button text change from Submit to "Submitting..." and
   back to Submit.

Change-Id: I5634c180f54f8cd55ef9e9741634d2fbc256c2e5
Reviewed-on: https://gerrit.instructure.com/85804
Reviewed-by: Spencer Olson <solson@instructure.com>
Tested-by: Jenkins
QA-Review: Gentry Beckmann <gbeckmann@instructure.com>
Product-Review: Keith T. Garner <kgarner@instructure.com>