Allow items to be moved into new modules without refresh
This makes it so that the content portions of a newly
created module receive the proper module id.  The move to
dialog feature attempts to find by this id in order to
perform the move on the client, when not found it was
silently failing.  This should prevent that from ever
happening as this was the only case where the ids on the
content were not set.

closes CNVS-31883

Test Plan:
  - Go to the modules page
  - Create a module with some items in it
  - Create another module
  - Use the cog menu on a module item to move it to
    the newly created module.
  - It should move without problem.

Change-Id: Ibab0c92470d19491a93140bbd9fba44221951e82
Reviewed-on: https://gerrit.instructure.com/91028
Tested-by: Jenkins
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
QA-Review: Heath Hales <hhales@instructure.com>
Product-Review: Chris Ward <cward@instructure.com>