update react-modal dep in package.json
closes: CNVS-33970

the old version of react-modal was using an older version of lodash.
it only needed the _.assign method but with the old version of lodash
it brought in a ton of extra JS. by using this newer version our bundle
size should be noticably smaller and the webpack build time should be
shorter.

test plan:
* go to a page that has a react-modal & make sure it works
e.g.: go to theme editor and open the modal to name the theme

Change-Id: I9b82c1abe2b10a1977bb171c43878ada470400d3
Reviewed-on: https://gerrit.instructure.com/98149
Reviewed-by: Clay Diffrient <cdiffrient@instructure.com>
QA-Review: Jeremy Putnam <jeremyp@instructure.com>
Product-Review: Clay Diffrient <cdiffrient@instructure.com>
Tested-by: Jenkins