simplify and speed up files index json response
as far as I can tell this response is now only used for rendering the
"upload folder" select box in the wiki sidebar, so it can be drastically
simplified for this single use case.  note that in a course with tons of
folders we are still sending them all down.  we may need to further
limit this down to a certain "max depth" (which we already do on the
modules insert item dialog)

fixes CNVS-34482

test plan:
- make sure uploading files to folders in the wiki sidebar still works
and the folder structure still looks right in the select box
- double check other areas that show folder hierarchy (I couldn't find
any others that use this, and tests didn't find any, but I still worry
we might have missed one)

Change-Id: I135d1c71cd3ee77ede3dcb19249275b12a383995
Reviewed-on: https://gerrit.instructure.com/100890
Tested-by: Jenkins
Reviewed-by: Cody Cutrer <cody@instructure.com>
QA-Review: Jeremy Putnam <jeremyp@instructure.com>
Reviewed-by: Ryan Shaw <ryan@instructure.com>
Product-Review: Simon Williams <simon@instructure.com>