spec: fix byebug+timeouts, consolidate logic
i had thought it was working w/ byebug, but apparently not. presumably
sometime after the initializer was written, byebug introduced its
freeze-all-other-threads behavior. so make it work again, but make it
like the Pry approach in that if we have ever debugger'd, timeouts
become disabled for the rest of the process' lifetime (and not just if
we are currently debugger'ing, since that's impossible w/ byebug since
the timeout thread is frozen)

test plan
1. put a debugger in a spec and continue after a minute
2. confirm you don't get hit by the spec timeout
3. put a binding.pry in a spec and continue after a minute
4. confirm you don't get hit by the spec timeout

Change-Id: I78f2ed4c2501273564cfed77504ec14d4b442d66
Reviewed-on: https://gerrit.instructure.com/100936
Tested-by: Jenkins
Reviewed-by: Shawn Meredith <shawn@instructure.com>
Product-Review: Shawn Meredith <shawn@instructure.com>
QA-Review: Shawn Meredith <shawn@instructure.com>