Add DUE_DATE_REQUIRED_FOR_ACCOUNT to the js_env
Added DUE_DATE_REQUIRED_FOR_ACCOUNT to the js_env in the quizzes
controller.  This will be used in to verify if a date is required for
the quiz, and this functionality will appear in the next
commit.

Refs: SIS-2636

Test Plan: QA will review the functionality in the final commit.

Change-Id: If6bf2213b7484b06f05b1b6a062153a95c740c72
Reviewed-on: https://gerrit.instructure.com/101990
Tested-by: Jenkins
Reviewed-by: Brad Humphrey <brad@instructure.com>
Product-Review: Stewie aka Nicholas Stewart <nstewart@instructure.com>
QA-Review: Stewie aka Nicholas Stewart <nstewart@instructure.com>