Ignore query parameters for developer key redirect uri
When using the new developer key redirect_uris, the final redirect_url should
exactly match one of the provided urls. However, some LTI apps require request
specific parameters to allow cookieless sessions. This simple fix just removes
the query parameters before checking if it is a valid uri. The input is
converted to a string before splitting to prevent an error on nil as an input.

Test Plan:
  - Find an LTI app that uses request specific redirect_uris and legacy
    redirect_uri for the developer key.
  - Change the developer key to use the new redirect_uris instead of
    redirect_uri.
  - Verify that the oauth authorization page still functions and not raises an
    error for an incorrect redirect_uri