clean up old webpack artifacts after each build
closes: CNVS-36233

webpack adds a fingerprint to the filenames that it outputs, which 
means that if the content changes, the filename changes. so over 
time, unless you manually know to rm -rf public/dist/webpack* 
periodically, you'll start accumulating a ton of javascript files 
in that dir that you don't need which can slow down your editor's 
grepping, take up a ton of space, and confuse you. this change makes 
it so right before webpack emits all of the new bundles, it cleans up 
the ones that were there. it only cleans up the ones in the dir it is 
outputting to, so you can still have prod bundles in 
public/dist/webpack-production/ and dev bundles in 
public/dist/webpack-development

test plan:
* run webpack before you check this out
* check this out
* run yarn then run webpack again
* you should see something in the webpack output like:
  WebpackCleanupPlugin: 220 file(s) deleted.
* and there should be only one copy of each bundle in the 
  public/dist/webpack-development dir.

Change-Id: Ide7a7f53c255c7aa5ad09fb4ff0bcfd7436b2852
Reviewed-on: https://gerrit.instructure.com/108147
Tested-by: Jenkins
Reviewed-by: Clay Diffrient <cdiffrient@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>
QA-Review: Ryan Shaw <ryan@instructure.com>