recreate the pg_collkey based indexes to use collation level 3
closes CNVS-38143

this commit aligns the collation schemes for postgres, ruby and
javascript to the same level.  this means that when sorting an
array of strings in any of the three environments, the result
will be identical.

this also stops relying on default collation strengths in ruby
and postgres and explicitly specifies the default collation
strengths to avoid surprises later.

test plan:
* Verify migrations pass with multiple shards
* Verify migrations rollback properly individually

Change-Id: I6023b364c260c35b85a8d6364ccc346e2ef7d159
Reviewed-on: https://gerrit.instructure.com/118675
Tested-by: Jenkins
Reviewed-by: Derek Bender <djbender@instructure.com>
Reviewed-by: Cody Cutrer <cody@instructure.com>
Product-Review: Keith T. Garner <kgarner@instructure.com>
QA-Review: Keith T. Garner <kgarner@instructure.com>