wait for grid before handling loaded submissions
fixes CNVS-39366

QA Notes:

There is only one non-deterministic ways to likely force the
scenario which revealed this bug: Create a bunch of assignments
and custom columns, while using only one student.

This is NOT guaranteed to produce the bad behavior prior to this
commit. However, if the issue is consistent with this setup
before testing this commit and is not present afterward, this
issue should be considered resolved.

test plan:
 1. Create a course with:
    a. 4 grading periods
    b. 10 assignment groups
    c. 40 assignments
    d. 10 custom columns
    e. one student
 2. Check out master
 3. Open Canvas in Chrome with dev tools open to the web console
 4. Visit Old Gradebook
 5. Verify that an error is logged to the console:
    * `something...something removeCellCssStyles`
 6. Apply this commit
 7. Open Canvas in Chrome with dev tools open to the web console
 8. Visit Old Gradebook
 9. Verify that the error above does not appear
10. Verify that Old Gradebook loads successfully

Extra Note regarding Step 5:

Another indicator that the conditions for this issue are present
is to open the network tab in Chrome dev tools and verify that
the request for submissions completes prior to at least one of
the following other requests:
 * assignment_groups
 * custom_gradebook_columns
 * grading_period_assignments

Observing that this is occurring after the commit is applied will
be verification that the conditions for this issue are still
present without being a bug anymore.

Meow!

Change-Id: I506dc8f1c2ed04f560f9c1fc0e019c32d1ef1e05
Reviewed-on: https://gerrit.instructure.com/126425
Tested-by: Jenkins
Reviewed-by: Spencer Olson <solson@instructure.com>
Reviewed-by: Matt Taylor <mtaylor@instructure.com>
QA-Review: Anju Reddy <areddy@instructure.com>
Product-Review: Keith T. Garner <kgarner@instructure.com>