move things from “devDeps” to “dependencies”
Closes CNVS-39744

we used to only have the things needed to compile brandable_css in
“Dependencies”. This was because, when we used to have to run 
brandable_css on a job server when someone pressed “preview” in the 
theme editor, we only wanted to have the minimum amount of node_modules
installed there to do that. But now that we no longer have to do
any brand-specific css compiling we can just do what is considered
normal and put anything that actually goes on the page in “dependencies”
and leave all the build tooling, linting, etc stuff in devDependencies.

This also makes it so it is really easy to see the cases that we
might be loading 2 versions of something in our webpack bundles
by running `yarn install --prod --flat`. Whatever that prompts you
about is something that will be included more than once in our webpack
bundles.

Test plan:
* run `yarn install`
* it should not generate a new yarn.lock because everything is the
  same versions as before.

Change-Id: I87b88f50d2cdbe494f928e5418e8540da9cc084c
Reviewed-on: https://gerrit.instructure.com/128244
Tested-by: Jenkins
Reviewed-by: Felix Milea-Ciobanu <fmileaciobanu@instructure.com>
Reviewed-by: Clay Diffrient <cdiffrient@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>
QA-Review: Ryan Shaw <ryan@instructure.com>