avoid mastery scaling if mastery pts or pts possible is 0
fixes OUT-1591

test plan:
- create an outcome that uses either decaying avg or
  n mastery
- set its points possible to 0 (set all criterion to 0)
- create another outcome using one of the above methods
  and set its mastery score to 0
- create a question bank, add questions, and align it to
  the two newly created outcomes. Due to the setup of
  the outcomes, any mastery percent set on the alighment
  should suffice
- align the bank questions to a quiz
- as a student, take the quiz
- as the teacher, the LMGB should properly load

NOTE: The scores shown for this specific bank will not
look correct in the LMGB, since the percentage on the question
will be trying to multiply by 0, thus coming back as 0 even if the
student did not get a 0 on the bank. This is due to the percentage
scaling system not anticipating someone having 0 points possible
on an Outcome rubric. This issue will be addressed in a separate
ticket.

Change-Id: I61dd49e8fba62bb960cadff6af790c402bfced3b
Reviewed-on: https://gerrit.instructure.com/129455
Reviewed-by: Michael Brewer-Davis <mbd@instructure.com>
QA-Review: Andrew Porter <hporter-c@instructure.com>
Tested-by: Jenkins
Product-Review: Sidharth Oberoi <soberoi@instructure.com>