fix flaky javascript spec by using callback prop
fixes GRADE-642

There is a test that depends on a race condition to pass. The
only deterministic way to avoid this is to pass along a callback
option that is called when the appropriate state is achieved,
prior to continuing with the assertions.

test plan:
 1. Ensure the account has avatars enabled
 2. Visit your user's profile
 3. Upload an avatar from an image
 4. Refresh the page
 5. Verify the avatar was created successfully

Change-Id: Ifc6230e9819b4383cfb691daef744dd75d996c88
Reviewed-on: https://gerrit.instructure.com/131871
Reviewed-by: Aaron Shafovaloff <ashafovaloff@instructure.com>
Tested-by: Jenkins
Reviewed-by: Keith T. Garner <kgarner@instructure.com>
QA-Review: Anju Reddy <areddy@instructure.com>
Product-Review: Keith T. Garner <kgarner@instructure.com>