Fix New Activity button gap in IE11
Fixes: ADMIN-110, UIDEV-111

This commit fixes the gap between the Canvas dashboard
header and the New Activity button in IE11:
https://cl.ly/2e2C1w3c0Z3W

IE11 does not support `position: sticky`, so the
button will not stick at the top-left of the Planner
on scroll, like it does for Edge and other supported
browsers.

Test plan:
- Open your planner dashboard as a student in IE11

- Verify that there is no longer a gap between the
  button and the dashboard header:
  https://www.screencast.com/t/lhkL5M3t5

- Verify that Edge and other supported browsers
  have not been affected by this change. NOTE:
  There is a separate problem with Firefox and
  the New Activity button that will be addressed
  in ADMIN-520 -- so expect to see the offset
  wrong in Firefox. That is a problem with Planner
  code, not Canvas.

Note that IE11 sometimes has problems with text
alignment (see the screenshot above). This is not
related to this commit.

Change-Id: Ic0786ea03c4e4e076f1a482648c501beeb7ded51
Reviewed-on: https://gerrit.instructure.com/132028
Reviewed-by: Ed Schiebel <eschiebel@instructure.com>
Tested-by: Jenkins
QA-Review: Dan Sasaki <dsasaki@instructure.com>
Product-Review: Chris Hart <chart@instructure.com>