change procedure to add not null constraint to large scores table
There is concern that using change_column to add a NOT NULL constraint
on a large table will lock the table for too long. This uses a
procedure that locks the table for a short period of time to add the
constraint, and then kicks off a validation that will run outside of
the lock.

closes GRADE-668

test plan:
 - The migration runs
 - Check that the NOT NULL constraint is on the table.

Change-Id: Idc11f170768cd4af9820e73950c5cfcc6a3354de
Reviewed-on: https://gerrit.instructure.com/133002
Reviewed-by: Rob Orton <rob@instructure.com>
Reviewed-by: Shahbaz Javeed <sjaveed@instructure.com>
Tested-by: Jenkins
QA-Review: Jeremy Neander <jneander@instructure.com>
Product-Review: Keith T. Garner <kgarner@instructure.com>