Fetch more users on scroll in collaboration picker.
For some reason, when a UserCollection is passed in a comparator
as an argument, it immediately gets an element.  However, if I
instead set that comparator on the next line then it doesn't get
an element right away.

This is breaking scrolling because PaginatedView checks for
whether the collection is "empty" in deciding whether to stop
listening for scrolling.

Yeah, this might be the stupidest bug fix I have ever made.

Fixes COMMS-703

Test Plan:
* Have a course with more than 50 students (make it a lot more).
  I suggest using rails console.  I did something like:
  `c = [some course]
   for i in 1..200 do
     u = User.create!(:name => "User #{i}")
     u.register!
     e = c.enroll_student(u)
     e.accept!
   end`
* Go get some snacks while these things are being made.
* Go to `courses/:course_id/collaborations`
* Scroll down.  Pagination should now work on
  hitting the bottom.

Change-Id: Ic547ee0c17cb3590ee77b1d0673d7953450b307e
Reviewed-on: https://gerrit.instructure.com/138909
Reviewed-by: Steven Burnett <sburnett@instructure.com>
Reviewed-by: Aaron Kc Hsu <ahsu@instructure.com>
Tested-by: Jenkins
QA-Review: Landon Gilbert-Bland <lbland@instructure.com>
Product-Review: Venk Natarajan <vnatarajan@instructure.com>