harden PaginatedCollection loading
The first issue addressed here was that it was possible to incorrectly
kill api param arguments if you asked for a "page" when a pagination url
for that page didn't already exist.  We sometimes use this pattern to
always ask for the "next" page, even on initial load.

The second issue was that we fire backbone events from the dataFilter
that is called post xhr, and we mark the collection as not in
a "loading" state at that point, but we don't yet check if all pages
have loaded.  This could cause a subsequent, erroneous request to be
generated from a callback to the dataFilter events.

Also, sometimes specs are flaky because of a real problem! \o/

fixes ADMIN-821

test plan:
- have a course with student and teachers
- repeatedly load the new group dialog as a student
- teachers should never show up
- it should only make one xhr request to load users

Change-Id: I8ddcc77764e8a01994a495992d0fcb1786359d17
Reviewed-on: https://gerrit.instructure.com/142340
Reviewed-by: Ryan Shaw <ryan@instructure.com>
Tested-by: Jenkins
QA-Review: Deepeeca Soundarrajan <dsoundarrajan@instructure.com>
Product-Review: Simon Williams <simon@instructure.com>