Make the wiki page input box get focus properly
This adds a wait so that canvas-rce doesn't take
focus away from the input. There was already logic
to do this, but some browsers (Firefox) regularly
handled it out of order making focus go to the RCE
instead.  Unfortunately, we need the `focus` option
for the RCE so that the sidebar registers the RCE
instance as the active RCE

closes CORE-1016

Test Plan:
  - Make a new wiki page
  - Notice focus is on the title input

Change-Id: I149a649a412302221a328c1bcf08eb11bf9e9e17
Reviewed-on: https://gerrit.instructure.com/142801
Reviewed-by: Ryan Shaw <ryan@instructure.com>
Tested-by: Jenkins
QA-Review: Jeremy Putnam <jeremyp@instructure.com>
Product-Review: Clay Diffrient <cdiffrient@instructure.com>