Make styling of blueprint button on course item row better
Note this is still not perfect, the keyboard only focus border goes a
bit lower then the other elements, which is way harder to fix then it
has any right to be. But this gets everything alligned correctly and
still passes a11y, so we don't want to spend any more time working on
it. There is a tech debt issue to switch this out to a ToggleIcon at
some point in the future, at which point none of this matters anyway as
there will be consistancy for all those buttons

Fixes COMMS-961

Test Plan:
  * Create a master blueprint course
  * Create some discussion in this course
  * Notice that the blueprint lock/unlock button has some proper spacing
    around it and the other buttons / menu icon
  * Add some announcements to this course
  * Notice that the blueprint lock/unlock button looks correct on the
    announcements page as well
  * Using keyboard navigation, make sure that the buttons aren't using
    the built in browser focus highlighting (it should be blue)

Change-Id: Ie5b4e29b5a6eeb69a030d179debf0264bcc54b61
Reviewed-on: https://gerrit.instructure.com/143623
Tested-by: Jenkins
Reviewed-by: Aaron Kc Hsu <ahsu@instructure.com>
Reviewed-by: Venk Natarajan <vnatarajan@instructure.com>
Product-Review: Steven Burnett <sburnett@instructure.com>
QA-Review: Steven Burnett <sburnett@instructure.com>