The Bulk of the RTL CSS changes
fixes: CORE-1172

As stated in the last commit message, the RTL changes were split into 2
commits. this is the much bigger of the two but the test plan can be
much simpler. the net result of this change is that absolutely nothing
should change in any of our existing css output. this just makes it so
all of these places where we said:
margin/padding/float/border/border-radus: (left or right), it does the
opposite in RTL so it flows right.

test plan:
* before you check this out, compile all the css in app.
* copy the contens of public/dist/brandable_css to a folder
* check this commit out
* run brandable_css again and diff the output of that folder with
  the old folder.
* there should be absolutely no changes

* verify that for the "_rtl" variants, the right and left values are
  swapped whenever they need to be.

* turn on the RTL feature flag and browse around, everything should
  be RTL and all of the margins, padding, spacing, arrow icons, etc
  should be flipped and make sense for an RTL layout.

Change-Id: I9b70896ce6b262cc45623993de2db61810a75b27
Reviewed-on: https://gerrit.instructure.com/143972
Reviewed-by: Brent Burgoyne <bburgoyne@instructure.com>
Tested-by: Jenkins
QA-Review: Jeremy Putnam <jeremyp@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>