fix editing a planner item outside loaded range
fixes ADMIN-940

If you created or edited a planner item so it was outside the currently
loaded date range, it would appear to be the only item on that date,
even though if you loaded more the rest of the items on that date would
appear. This is confusing, so instead of displaying that item, we put
a success alert indicating the item was successfully saved and we don't
show that item on a date that hasn't been loaded.

test plan:
- the number of test cases is large and complex and depends somewhat
on data that you can't see in the interface. See the
save-item-reducer.spec.js file to get an idea of what to test.

- Generally you'll want to have 9 or 10 items on today and shorten your
screen such that only today loads. Then add items for tomorrow and
yesterday, and then make sure they load along with the exisiting items
when you load the future or past. Also, with lots of items near the
present, add items in the far future and past, and make sure they load
as well when you scroll to their date.

Change-Id: I8183ec3cbcd90ab1fcf5c13bb17ba671a0f18d90
Reviewed-on: https://gerrit.instructure.com/150238
Tested-by: Jenkins
Reviewed-by: Ed Schiebel <eschiebel@instructure.com>
Product-Review: Jon Willesen <jonw+gerrit@instructure.com>
QA-Review: Jon Willesen <jonw+gerrit@instructure.com>