Make discussions index faster by pre`fetch`ing
closes: CORE-2750

This change makes it so we fire off window.fetch requests for each of
the api requests we’re going to need to show all the discussions on
the page up in the <head>, instead of waiting until after all the
javascript has loaded and ran. By starting those requests earlier
the servers will start working on them while we are still loading all
our JS and the page will complete loading sooner. This should help
our end-user-page-speed metric and may be a pattern we can use in other
places too.

Test plan:
In an environment that has multiple rails processes handling requests,
(like if you are running Pow locally, or on beta, or portals, but not if
you are just running a single-threaded `rails server` from the cli)
* go to courses/x/discussion_topics
* the page should load a lot faster. As in, you should see your actual
  discussions quicker.
* to especially see the effect, open chrome devtools and go to the
  Network panel and turn on the simulate “Fast 3g” connection option.

Change-Id: Id381c3581d9de91c7c50dba2ee380b36d5587cf1
Reviewed-on: https://gerrit.instructure.com/188651
Tested-by: Jenkins
Reviewed-by: Steven Burnett <sburnett@instructure.com>
QA-Review: Steven Burnett <sburnett@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>