instUI5 upgrade: meaningful gradezilla/grading stuff
refs: CORE-1437
closes: CORE-1489

the commit to upgrade gradezilla/grading stuff to instUI 5 has been
split in 2. The other, that this is based on takes care of the trivial
stuff: components whose API has not changed, who just need to have their
import paths  renamed to the new location in the split-up instUI
packages.

This one takes care of the stuff that actually has meaningful changes
like <NumberInput> and <Select>. it will need more scrutiny, but
since the changes are isolated to just a couple components it won’t
be as scary.

Test plan:
* verify these components still work as they used to and tests pass
* look at the things that @solsen found in his Code Review and make sure
  they work

Change-Id: I72996ead775793237fc9a527f3da366f8ad6a6e3
Reviewed-on: https://gerrit.instructure.com/151772
Tested-by: Jenkins
Reviewed-by: Spencer Olson <solson@instructure.com>
Reviewed-by: Keith T. Garner <kgarner@instructure.com>
QA-Review: Rohan Cheeniyil <rcheeniyil@instructure.com>
Product-Review: Pert Eilers <peilers@instructure.com>