Speed up rendering of assignment groups
Note that this is not a proper fix. What is happening here is the inital
paginated API requests are being rendered in one go, then each
individual assignment group from subsequent pagination requests are
causing *each* assignment group to be fully re-rendered (including all
the assignments in the assignment group). This if we had 100 assignments
in total, a per_page size of 10, and 15 assignment groups, the render
function for the AssignmentGroupListItemView would be initially called
600 times (100 initially, and 100 more for each of the 5 assignment
grouns after the initial pagination request).

Right off hand, a proper fix could be done in a couple different ways:

  * Have some sort of componentShouldUpdate() functionality in backbone
    to avoid useless re-renders
  * Stuff all the AssignmentGroups into the ENV and override `fetch()`
    to grab them from the ENV instead of hitting the API.

However, as i understand it, this page is going to be completely redone
soon, and we don't want to spend a bunch of time hacking away on this
just to throw it all away for the re-work. Thus we are hoping that in
increase of the initial per_page load count here will offset the
common cases where this issue is being a problem.

Refs COMMS-1104

Test Plan:
  * Create an bunch of assignment groups and assignments in one of the
    groups.
  * Load the page as a teacher and a student, and make sure in both
    cases the page loads and is interactable with reasonably fast

Change-Id: I33560d9a9443da2943cae81e413a3bf1d39d7ab7
Reviewed-on: https://gerrit.instructure.com/155152
Reviewed-by: Steven Burnett <sburnett@instructure.com>
QA-Review: Steven Burnett <sburnett@instructure.com>
Tested-by: Jenkins
Product-Review: Steven Burnett <sburnett@instructure.com>