upgrade react-apollo so it doesn’t use own lodash
The version of react-apollo that we were using had a hard dependency
On lodash 4.17.5 but the rest of all of our stuff was using 4.17.10.
That meant that react-apollo would use it’s own copy of lodash leading
To unnesisary JS bundle size in the browser.

But the latest version changes it’s hard dependency to 4.17.10 so it can
Share with everything else

Test plan:
* pages that use react-apollo should not download their own copy of
  lodash

Change-Id: I63068ce83d9ad834cbcd5fc3174c7cdf411e4409
Reviewed-on: https://gerrit.instructure.com/155245
Reviewed-by: Aaron Kc Hsu <ahsu@instructure.com>
Tested-by: Jenkins
Product-Review: Ryan Shaw <ryan@instructure.com>
QA-Review: Ryan Shaw <ryan@instructure.com>