Find correct item when clicking Today button
fixes ADMIN-1249

the code prevously looked for the item whose time was closest to today,
but that ignored the fact that in the UI, items are first grouped so
that the next item after today visually, might not actually be the
absolutely next item due. This was confusing.

test plan:
  - have a student with items in the past and future, scroll into the
    past, then click today
  > should focus on the first visible item in the future
  - have one item in one class AAA due at 2pm today
  - have another item in class BBB due at 1pm today
  - click today
  > should focus the item in AAA, which is first in the UI, even
    though it's due after the item in BBB

Change-Id: I142ae814845b35df52c4a79218f959f9f5ce90db
Reviewed-on: https://gerrit.instructure.com/157696
Tested-by: Jenkins
Reviewed-by: Jon Willesen <jonw+gerrit@instructure.com>
QA-Review: Jon Willesen <jonw+gerrit@instructure.com>
Product-Review: Ed Schiebel <eschiebel@instructure.com>