Change duplicate tool installation check
Fixes PLAT-3688

Test Plan:
Verify a warning is shown when installing a tool with the same launch
url as another active tool in the context.

Verify a warning is shown when installing a tool with no launch url
(just domain) when the domain matches the domain of another active tool
in the context.

Change-Id: I8d0a184fff9070c5e38f01409d403c1617c4070a
Reviewed-on: https://gerrit.instructure.com/163348
Tested-by: Jenkins
QA-Review: August Thornton <august@instructure.com>
Reviewed-by: Marc Alan Phillips <mphillips@instructure.com>
Product-Review: Jesse Poulos <jpoulos@instructure.com>