Add a check for every assignment in the to_do_list_presenter
considerations:
* note that the check goes through every assignment and runs a filter
on them. this is because the contexts object handed into the presenter
can be nil and because the assignments can span multiple courses. i'm
open to suggestions on what would be better.
* i would suggest to check for a similar bug for assignments needing
moderation and submitting, because the to do list is built the same
way for assignments in those states and how it used to be for ones
needing grading.

fixes: ADMIN-1459

Test plan
run through the repo steps on the ticket and make sure the todo
list on the dashboard doesnt have stuff they cant grade.

Change-Id: I7b678bad0d60547a8b93cd6003b5bb9d22570d20
Reviewed-on: https://gerrit.instructure.com/165936
Reviewed-by: Carl Kibler <ckibler@instructure.com>
Reviewed-by: Mysti Sadler <mysti@instructure.com>
Tested-by: Jenkins
QA-Review: Anju Reddy <areddy@instructure.com>
Product-Review: Rex Fleischer <rfleischer@instructure.com>