allow all day on calendar model to be set back to false
when start_at and end_at on the calendar_model goes to the
beginning of the day, it automatically set all_day to true. but it
was unable to do so if start_at and end_at was then filled. now
the model can correctly set all_day when the time values change.

Test-Plan:
* make an event on the calendar with times during the day
* update the event to be all day
* update the event again to have times

fixes: ADMIN-1466
Change-Id: Ie16f76e88fe49b3ad68ae824280710ae0edcfefd
Reviewed-on: https://gerrit.instructure.com/166508
Tested-by: Jenkins
Reviewed-by: Jon Willesen <jonw+gerrit@instructure.com>
QA-Review: Jon Willesen <jonw+gerrit@instructure.com>
Product-Review: Rex Fleischer <rfleischer@instructure.com>