Fix index query for LTI 1.3 tools
Previously index queries were being run against all
the context tools for a devkey, which would mean that
the account would show as being installed with a tool
if any of the courses or subaccounts had installed it.
Now it searches from the context up to see if it is
installed, not the other way around.

Test Plan:
 - install a tool in an account (1.3)
 - note that it shows in a course and it can't be installed
 - uninstall the tool, note that it is removed in teh course
 - install that tool in the course
 - note that it is not installed at the account
 - install the tool at course then at account, note that there
   are two tools now
 - remove the course tool and note that there is now only one

Change-Id: I9985f80e68dbc99747f3467b1d69ef329713cf16
Reviewed-on: https://gerrit.instructure.com/167772
Tested-by: Jenkins
Reviewed-by: Weston Dransfield <wdransfield@instructure.com>
QA-Review: Weston Dransfield <wdransfield@instructure.com>
Product-Review: Marc Phillips <mphillips@instructure.com>