fix babel hack so it works with istambul coverage
when it runs into classes that have fat arrow methods, decaffeinate
Adds a hack at the top of the generated constructor function to allow
it to do `this.fatArrowMethod = this.fatArrowMethod.bind(this)` BEFORE
it calls `super` so it can recreate exactly the behavior of what 
coffeescript did.

That works fine and dandy but when you run with Istanbul it adds a bunch
Of generated code to track line execution and so when it was looking 
for thisFn.indexOf(';') in the hack,  it finds a semicolon inserted 
by istanbul

The actual app code was fine, it was just when you enable istambule that
It would break.

Long term, the better thing to do would be to modify the actual code so
It doesn’t need the hack but that requires careful examination of 
Execution order and human intervention and I didn’t want to do that as
Part of a “decaffeinate” commit.

Test plan:
Run
COVERAGE=true DISABLE_HAPPYPACK=1 yarn jspec-watch \
  spec/coffeescripts/views/accounts/admin_tools
* it should work
* the master build should pass after this is merged

Change-Id: Ib29f9d57134f992561f082ba75edca5c745bc91e
Reviewed-on: https://gerrit.instructure.com/167915
Tested-by: Jenkins
Reviewed-by: Clay Diffrient <cdiffrient@instructure.com>
QA-Review: Ryan Shaw <ryan@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>