Fix scrolling to new activity dot
changed the logic so dynamic ui manager is given the ids of the 2 dom
elements we're most interested in when scrolling: the planner header
and the new activity button. Now it can look for them and compute the
necessary offset from the top of the window as needed.

this will become necessary when the phone layout is finished,
since the header will wrap and it's height becomes dynamic.

fixes ADMIN-1486

test plan:
  - have a planner with some new activity in the past
  - click the new activity button
  > expect planner to load the past data and scroll the
    blue dot into view just below the new activity button
  - press the button again
  > expect the next dot to scroll into view.

Change-Id: I82ec54648466a7f9499296a8292c86721d2f847e
Reviewed-on: https://gerrit.instructure.com/168551
Tested-by: Jenkins
QA-Review: Jon Willesen <jonw+gerrit@instructure.com>
Product-Review: Ed Schiebel <eschiebel@instructure.com>
Reviewed-by: Jon Willesen <jonw+gerrit@instructure.com>