A2: add user search to submissions in graphql
This adds a bunch of search options to the assignment
submissionsConnection in graphql. Not all of the potentially needed
orders and filters are implemented, but this gets the basic concept and
the simple stuff implemented.

closes ADMIN-2324

test plan:
- go to the graphiql interface and query an assignment and its
  submissionsConnection. You should see a bunch of options in the filter
  input of the submissionsConnection.
- try out all the filters and orders and see that you get the results
  that you'd expect
car
Change-Id: I541e333dc544f2258c9c2359124bac0d24277abc
Reviewed-on: https://gerrit.instructure.com/184727
Tested-by: Jenkins
QA-Review: Carl Kibler <ckibler@instructure.com>
Product-Review: Jon Willesen <jonw+gerrit@instructure.com>
Reviewed-by: Carl Kibler <ckibler@instructure.com>