Fix unlimited scrolling for students managing a group
The first issue is that we have pagination setup through unlimited
scrolling, but were not loading enough users to fill up the modal and
actually have the browser disply the scroll bar. Once the scroll bar is
properly displayed, utilizing it for unlimited scrolling works as
expected.

The second issue discovered was we were doing the sorting of users on
the frontend instead of the backend. Coupled with unlimited scrolling,
you could reach the end of the scroll bar and end up loading new users
at the top of the list instead of at the bottom. To fix this, we added a
new `sort` option to the course user search api and stopped sorting
these on the frontend.

Fixes COMMS-1988

Test Plan:
  * Create and publish a course
  * Create a bunch of users in the rails console and enroll them in the
    course:

    ```
    a = Account.first
    c = Course.last
    (1..35).each do |i|
      u = User.create!(name: "Minion#{i}", short_name: "minion#{i}")
      u.accept_terms
      u.register!
      u.email = "#{u.short_name.downcase}@pixar-foo.com"
      u.save!
      pseudonym = a.pseudonyms.build(user: u, unique_id: u.email,
        password: 'asdfjkl;', password_confirmation: 'asdfjkl;')
      pseudonym.save_without_session_maintenance
      enrollment = c.enroll_user(u, 'StudentEnrollment')
      u.save!
      enrollment.course = c
      enrollment.workflow_state = 'active'
      enrollment.save!
    end
    ```
  * Create a new group set and group in the course.
  * Add any one student to the group and set them as the leader
  * Login as that student and navigate to the group *on the course*.
    Something like: http://pixar.canvas.test/courses/2/groups
  * Click on the `manage` button on the group, and notice that it loads
    the users in sorted order, and that you can scroll down to load more
    users which also show up in sorted order at the bottom.

Change-Id: I04e8a734833fb27e2cf73289c748894f26d9919c
Reviewed-on: https://gerrit.instructure.com/188886
Tested-by: Jenkins
Reviewed-by: Steven Burnett <sburnett@instructure.com>
QA-Review: Aaron Hsu <ahsu@instructure.com>
Product-Review: Landon Gilbert-Bland <lbland@instructure.com>