add submit button to content upload in A2
This only adds the functionality of rendering the submit button after a
file has been uploaded to an assignment. The button currently does
nothing and isn't wired up to anything in the backend. The submit button
should be in a sticky footer fixed to the bottom of the page that
appears when a file has been uploaded in the upload tab. Similarly if
there are no files or if the files have been removed the footer and
submit button should disappear.

fixes COMMS-2014

Test Plan
* Create a course
* Add a student to the course
* Create an assignment in the course
* Masquerade as the student and navigate to the assignment in A2
* Hide the masquerade bar at the bottom of the browser
* Notice that there is no submit button
* Upload a file on the upload tab
* Notice the footer with the submit button
* Remove the uploaded file
* Notice that the footer and submit button have been removed

Change-Id: Ia2c2ac1a0ed56c01961f2f25439aec3de419fc04
Reviewed-on: https://gerrit.instructure.com/191280
Tested-by: Jenkins
Reviewed-by: Steven Burnett <sburnett@instructure.com>
QA-Review: Steven Burnett <sburnett@instructure.com>
Product-Review: Matthew Lemon <mlemon@instructure.com>