make conversations dropdown focus/selected styling a11y friendly
The selection in the top right of conversations has styling
problems where the contrast of the hover state is not dark
enough to create a contrast. The old borders also prevent
the styling from having a contrast even if we darkened the
background on hover to match the rest of canvas. The
solution here as suggested by our a11y team is to remove the
border and make the hover background have a good contrast
color. This fix also required changing the hover color
state of the text inside the selection list view.

fixes COMMS-2032

Test Plan
* Navigate to conversations
* Do a a11y ally sweep of the top right search bar
* Make sure to navigate the component in contrast mode
 as well

Change-Id: Ie7f139321f26611b19c74df22d2ce1ac30a2f6ce
Reviewed-on: https://gerrit.instructure.com/191101
Tested-by: Jenkins
Reviewed-by: Steven Burnett <sburnett@instructure.com>
QA-Review: Steven Burnett <sburnett@instructure.com>
Product-Review: Nathan Nuclear <nathan@instructure.com>