make unread part of the title for discusions/announcements
fixes COMMS-2038

When we have an object that's unread, we should include that information
as a part of the title, rather than as an independent piece of text for
SR users. This ensures that it's an intentional part of the object,
rather than just being associated to the object by proximity.

Test Plan
- create a course and add a student
- create several announcements and discussions in the course
- masquerade as the student
- use a SR to navigate through the discussions and announcements
- notice that the unread indicator is now part of the announcment and
discussion title

Change-Id: Ia3e469171c3b066cb17ad09ffa1bf504cdfca4cc
Reviewed-on: https://gerrit.instructure.com/192098
Tested-by: Jenkins
Reviewed-by: Steven Burnett <sburnett@instructure.com>
QA-Review: Steven Burnett <sburnett@instructure.com>
Product-Review: Matthew Lemon <mlemon@instructure.com>