Do not update version if posted_at changed
Do not consider changes to a submission's posted_at attribute to merit a
change to the most recent submission version (as, for example, a
newly-issued grade would). When a submission *is* updated by a change to
a worthier attribute, the current posted_at time will still be copied to
the version object, as is standard. The intent is to fix a potential
race condition on the latest submission version that could occur when a
submission received a grade and a comment at almost exactly the same
time.

fixes GRADE-2186

Test plan:
- Create an assignment with text entry or online submissions
- As a student, submit something to it
- In the console, verify that a version object has been created:
  > submission.versions.count # should be 1
- As a teacher, leave a comment on the submission (do not grade)
- In a console, check that a new version was *not* created, and that
  the existing version was not updated to include posted_at even though
  the submission itself was
  > submission.reload.versions.count # should still be 1
  > submission.versions.first.yaml.posted_at # should be nil
- As a teacher, give the submission a grade
- In a console, check that the submission version has now been updated
  with the grade info and posted_at (since a change to an actually
  significant attribute prompted re-saving the version)
  > yaml = submission.versions.first.yaml
  > yaml.grade # should be the grade you issued
  > yaml.posted_at # should be the original posted_at time

Change-Id: I4cd8cf15e39aac788fe9e582cbf5ee0993dd7743
Reviewed-on: https://gerrit.instructure.com/192903
Reviewed-by: Landon Gilbert-Bland <lbland@instructure.com>
Reviewed-by: Gary Mei <gmei@instructure.com>
Tested-by: Jenkins
QA-Review: Adrian Packel <apackel@instructure.com>
Product-Review: Keith Garner <kgarner@instructure.com>