Include message info in Conversations' reply and menu icons
When a conversation is opened and a screen reader is being used
to navigate the conversation, the screen reader only reads "reply"
and "message actions" for the respective icons, giving the user no
information about which message the icon refers to. These icons
have been updated to include the body of the message in the screen
reader's voiced message to the user for the Reply and Message
Options icons.

Reproduction Steps:
1. In Safari with VoiceOver enabled, navigate to the Conversations
page
2. Open a message and move focus to the reply and menu icons shown
above the most recent message
3. It only says "Reply" and "message actions", not what it will
affect
4. Move to the reply and menu icons of the conversation (to the
right of the subject line)
5. It only says "Reply" and "message actions", not what it will
affect

Test Plan:
1. Repeat steps 1 and 2 from Reproduction Steps.
2. Notice the screen reader now says "Reply to <message>" and
"Message actions for <message>," giving the user a description of
the message that their actions will affect.
3. Move to the reply and menu icons of the conversation
(to the right of the subject line)
4. Notice the screen reader now says "Reply to conversation
<subject>" and "Message actions for conversation <subject>".
3. Create a conversation with no subject. Move to the hidden
reply and menu icons of the conversation (to the right of the
subject line) previous steps; notice the screen reader now says
"Reply to conversation" and "Message actions for conversation."

fixes COMMS-2037

Change-Id: I5358ae994c276a185d1068e73707463790eaf682
Reviewed-on: https://gerrit.instructure.com/193105
Tested-by: Jenkins
Reviewed-by: Matthew Lemon <mlemon@instructure.com>
Reviewed-by: Landon Gilbert-Bland <lbland@instructure.com>
QA-Review: Aaron Hsu <ahsu@instructure.com>
Product-Review: Ryan Norton <rnorton@instructure.com>