move bundles/common into main so webpack is not confused
Because common.js was in /bundles that technically means that it is a
“Bundle” that you could js_bundle onto a page. But it is just code
that should be ran as part of the “main” entry. Nothing ever
calls js_bundle(:common). So this helps webpack know that it doesn’t
have to separate it into its own bundle.

Reviewers: note that nothing from the code moved from common.js was 
changed, except I changed the “../“ in the import paths to “./“ because
Of location change.

Test plan:
* webpack build should be successful
* there should not be a dist/webpack-dev/common-c-<hash>.js chunk

Change-Id: If867432a3e638bf14c0edce46d50a8472ff68867
Reviewed-on: https://gerrit.instructure.com/193995
Tested-by: Jenkins
Reviewed-by: Steven Burnett <sburnett@instructure.com>
QA-Review: Ryan Shaw <ryan@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>