order account reports off of updated_at
the index that is in place is based off of
updated_at instead of created_at and the query to
get the account reports is really slow without on
created_at. We could make a new index and drop the
old index but would require more work, and they
*should* be the same because updated_at will only
change when it is deleted, but then will be
filtered by the .active on the same scope.

fixes CORE-3010

test plan
 - specs should pass

Change-Id: I6bda587dabda9e5907f185587aa7da22c3977468
Reviewed-on: https://gerrit.instructure.com/195557
Tested-by: Jenkins
Reviewed-by: Chris Mena <cmena@instructure.com>
QA-Review: Rob Orton <rob@instructure.com>
Product-Review: Rob Orton <rob@instructure.com>