fix rce course link drag image and firefox drag
setting position relative on the element fixed the dragging ghost
image that included previous elements in the tray.

drag and drop was completely broken in firefox. it turns out there is
a long standing bug when button elements cannot be dragable in
firefox. the solution was to make the element an anchor with a button
role.

fixes CORE-2864

test plan:
- make sure the new rce flag is enabled
- open the course links tray
- drag links of multiple types (Pages, Assignments, Quizzes, etc) in
  to the editor
- the drag image should only be the row for the link
- it should insert the corret link in the editor
- test with multiple browsers (including firefox)

Change-Id: I1feaa62d46833949d5c8b736d46a7adcb1a3fd9c
Reviewed-on: https://gerrit.instructure.com/195570
Tested-by: Jenkins
Reviewed-by: Ryan Shaw <ryan@instructure.com>
QA-Review: Ed Schiebel <eschiebel@instructure.com>
Product-Review: Brent Burgoyne <bburgoyne@instructure.com>