Don’t let page_views slow down page unload
See this? https://cl.ly/68e723db54c8 that’s a beforeuload in
page_views.js that is taking 335ms in prod to run _before_ the browser 
can navigate to any other pages.

what we are trying to do there in page_views is log the last page view
And we can’t use an xhr to do so because it would get aborted when we
go to a new page. Where there is a browser API to solve that exact 
problem: 

https://developer.mozilla.org/en-US/docs/Web/API/Navigator/sendBeacon

This is a screenshot of after: https://cl.ly/fed0076a15eb 
that sendBeacon call only takes 0.4ms

Oh, and we don’t need to do any of this unless ENV.page_view_update_url
Is set, so we can defer loading any of it unless that is present

Test plan:
* page views should work exactly as before
* turn on page views
* go to a page and sit there for more than 30 seconds
* clear your rails log and your browser log
* click to go to a new page
* you should see a request in your browser Network log to
  page_views/xxxx-xxxx-xxx-xx-xxx?page_view_token=xxxx
* look at your rails server logs, you should see something like:
  Processing by PageViewsController#update as */*
  Parameters: {"interaction_seconds"=>"1005”, …
  ”id"=>"36bcbd62-3a11-44bc-8dd2-69410050ff74"}
   

Change-Id: Ie67cc01dd1524c75916a26fbeffa67a01f490adc
Reviewed-on: https://gerrit.instructure.com/196416
Tested-by: Jenkins
Reviewed-by: Clay Diffrient <cdiffrient@instructure.com>
QA-Review: Clay Diffrient <cdiffrient@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>