Open RCE links in a new window
There is a bug in tinymce demonstrated in http://fiddle.tinymce.com/YJgaab/1
and reported as an issue at https://github.com/tinymce/tinymce/issues/5036
where the default_link_target is not set on an anchor if the target_list
is not shown in tinymce's link dialog. Because of this, external links
do not open in a new window.

CORE-3065 was created to eventually resolve the issue

closes CORE-2667
refs CORE-3065

test plan:
  - add a course link or course document link to your rce
  - save whatever you were editing
  - click the link
  > expect it to open in a new browser tab (or window,
    depending on your browser settings)
  - add a navigation course link, save, and click the link
  > expect it to load in the current window
  - add an external link, save, and click
  > while it should open in a new tab, it does not due
    to the tinymce bug described above.

Change-Id: I2befef3f41ae85bb33d11d561fd909862942bee3
Reviewed-on: https://gerrit.instructure.com/196958
Tested-by: Jenkins
Reviewed-by: Ryan Shaw <ryan@instructure.com>
Product-Review: Ed Schiebel <eschiebel@instructure.com>
QA-Review: Ryan Shaw <ryan@instructure.com>