add fake Calculated By string in a2 header
to visually demonstrate what the 'Calculated by' element
will look like for the InstructureCon demo we are adding a
non-functional string into the header.

this is not intended to be interactable, but per designs we want
it to look like a link and are using an empty anchor tag for this
purpose. empty anchor tags should never be used as they're not
keyboard accessible, but because this will be replaced shortly
after the demo and will never be interactable it is sufficient.

fixes COMMS-2097

Change-Id: I38bec1a796f897969f86949e63b52016c07aa5ba
Reviewed-on: https://gerrit.instructure.com/197014
Reviewed-by: Steven Burnett <sburnett@instructure.com>
Tested-by: Jenkins
QA-Review: KC Naegle <knaegle@instructure.com>
Product-Review: Ryan Norton <rnorton@instructure.com>