reconfigure assi_2 upload pane to be accessible
fixes COMMS-2025

The upload pane for assi_2 was not accessible in its previous
implementation because multiple Billboards were rendered within a
single FileDrop. instUI does not support this functionality, so
the upload pane needed to be reconfigured in order to be
accessible. The upload pane now has one smaller FileDrop, and the
uploaded files are rendered to the side & below it instead of
inside of it.

Test Plan:
* As a teacher, create an assignment.
* In Safari, as a student, upload some files to that assignment;
  do not submit.
* Notice the following:
    -When a file is uploading, a single loading indicator appears
     until the file preview appears
    -The first two uploaded files appear to the right of the
     FileDrop. All files uploaded later appear underneath the
     FileDrop in rows of 3.
* Turn on ScreenReader. Tab through the uploaded files; focus
  should go to the 'Delete' icons, and voice over should read
  "Remove <filename> button."

Change-Id: I8811c4853d0b46b8dfa808559853027248a9684c
Reviewed-on: https://gerrit.instructure.com/198691
Tested-by: Jenkins
Reviewed-by: Ryan Norton <rnorton@instructure.com>
QA-Review: Daniel Sasaki <dsasaki@instructure.com>
Product-Review: Michelle Simmons <misimmons@instructure.com>