fix preview error when no submission in a2
if we attempt to load the file preview page in a2 but we do
not have any submission to preview we were throwing an error
page.

this could occur if a student failed to submit an assignment
and a teacher graded them with a 0.

Test Plan:
* as a teacher, create a file upload assignment
* as a teacher, open the gradebook and give a student
  on that assignment a grade
* as a student, go to this assignment in a2
* you should be taken to the file preview page and there
  should be an SVG and a message indicating that there is no
  submission

fixes COMMS-2141

Change-Id: Iebc3bd608e6cc483a4df20071295bff4988a35b8
Reviewed-on: https://gerrit.instructure.com/198478
Tested-by: Jenkins
Reviewed-by: Landon Gilbert-Bland <lbland@instructure.com>
Reviewed-by: Steven Burnett <sburnett@instructure.com>
QA-Review: Michelle Simmons <misimmons@instructure.com>
Product-Review: Steven Burnett <sburnett@instructure.com>