fix bug where we were not loading new submission drafts in A2
fixes COMMS-2161

Before when we were updating a submission draft with new attachments the
view wouldn't update to display those changes. This was due to a bug
where we were returning null instead of the updated submission when
updating state in the view manager. This fixes that.

This is also fixing a bug where after submitting an assignment we would
see the "no submission" svg despite having just submitted. Before we
were required to reload the page to actually view the assignment and its
attachments. This was due to the apollo cache not getting updated
correctly, thus failing to re-render the view. We fixed this by adding
the submission attachments to the response from the createSubmission
mutation. By doing that we are now updating the cache appropriately and
things are just peachy.

TEST PLAN:
* Create a course and add a student to it
* Create an assignment with a file upload submission type
* Masquerade as the student
* Add and remove multiple files to the submission (this should work
  totally fine)
* Submit the assignment and note that the submission should show a
  preview for the files you submitted (or the no preview available message
  svg)
* In Assignments 1 resubmit the assignment multiple times with different
  files and in varying quantities.
* In Assignments 2 you should be able to navigate between your various
  submission attempts and there should be no problems

Change-Id: I93a8e9804eaeffc87ee8510cc1841b6dc863fcb5
Reviewed-on: https://gerrit.instructure.com/198856
Reviewed-by: Landon Gilbert-Bland <lbland@instructure.com>
Tested-by: Jenkins
QA-Review: Michelle Simmons <misimmons@instructure.com>
Product-Review: Steven Burnett <sburnett@instructure.com>