retry speeding up discussions show page w/ prefetch_xhr
closes: CORE-3145

This is attempt 2 of trying to do this. The original commit (g/198713) 
had to be reverted on caturday (in g/198713) because it caused 
CNVS-46307. So this commit essentially reverts the revert and fixes 
the bug in the original

Original commit message:
This uses prefetch_xhr to prefetch the request for the materialized view
of the discussion topic. It still uses the backoff poller it had before
so it can withstand 503 errors like it was designed to.

This will remove one roundtrip from the waterfall needed to render the
page so should speed things up.

Test plan:
* go to /courses/x/discussion_topics/y
* it should look and behave exactly as it did before, but the request to
  get the data for the discussion topic should happen earlier in the
  devtools network panel and should show up as type “fetch” instead of
  “xhr"
* the page should render faster since the data will be ready sooner
  (Note, you may need a setup with multiple app processes like pow
  To see the full extent of the perf improvement)

Test plan to make sure that the bug from the first try of this is fixed:

Prerequisite: Have at least one student and teacher enrollment.

Teacher role

* As a teacher create a discussion. 
  (As the creator you will be automatically subscribed to it)

* From within the discussion, unsubscribe. Refresh your page. 
* Notice the discussion is NOT deleted.

Student role

* Access the discussion. Attempt to subscribe/unsubscribe from it. 
* Notice you DO NOT receive a 401 error.

Change-Id: I8b6bb530e6b34d7abbdc5aee5de3ba9843a78c07
Reviewed-on: https://gerrit.instructure.com/199278
QA-Review: Steven Burnett <sburnett@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>
Tested-by: Jenkins
Reviewed-by: Clay Diffrient <cdiffrient@instructure.com>