get stuff ready to use lodash instead of underscore
closes: CORE-3178

right now, we send 2 copies of underscore/lodash to browsers. that is
wasteful and slow because it causes browsers to download and execute
more JS than they need to. we should use the most recent version of
lodash for everything and have everything just share that one.

Using the old version of lodash.underscore is also preventing us from
upgrading backbone to the latest version (because backbone relies on
using a newer lodash/underscore). and we want to upgrade backbone to
the latest version so we can use the
`https://backbonejs.org/#Model-preinitialize` method which was
specifically made to help people upgrade things to es6 classes and
avoid the "trick babel to allow `this` before `super` hacks" that we
have in our code & need to get rid of before we can drop ie11

but to make upgrading to lodash 4 easier, I made this commit which
avoids using it in a lot of places (so upgrading is a non-issue) and
uses the method alias names that are both present in the old underscore
we use and lodash 4.

some of this was done by:
https://github.com/cpojer/js-codemod#underscore-to-lodash-native

see https://github.com/lodash/lodash/wiki/Migrating for what is
different between underscore and lodash

test plan:
* automated tests should pass

Change-Id: Ieefc0be36d277fc115974aa06c03e46a37b9f426
Reviewed-on: https://gerrit.instructure.com/200550
Reviewed-by: Clay Diffrient <cdiffrient@instructure.com>
Tested-by: Jenkins
QA-Review: Ryan Shaw <ryan@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>