Added filter on start_at and end_at for account courses api.
Changed AccountsController.courses_api to support filtering courses
that start before a value and end after a value. Courses that have
a null value for the field will still be returned. The term's respective
field will also be checked. Either the course or term field passing
the filter will cause the course to be returned.

Test Plan:
  - Jenkins passes