simplify/modernize google analytics & load faster
now that we have streaming working, by loading google analytics
like this in the <head> it will load sooner but not block any of our
scripts from running.

this commit also upgrades us to use the analytics.js instead of the
legacy ga.js library.

it also actually imports it as a named export instead of just relying
on it being on the jQuery global `$` object

note: I actually took the time to set this up with a testing google
analytics account and verified that it got all the page views recorded
as before. Doing so even exposed one place where we were currently
sending PII to google analytics (in the “Act As” page) and I fixed that
https://cl.ly/e869a25266fc

test plan:
* in the console do: Setting.set('google_analytics_key', 'testing123')
  but replace “testing123” with a real valid google analytics account id
* google analytics tracking data should post to your google analytics
  account exactly the same as it did before this commit

Change-Id: I3ca51642c622ce72992159646f9d9338c58b44ba
Reviewed-on: https://gerrit.instructure.com/203581
Tested-by: Jenkins
Reviewed-by: Steven Burnett <sburnett@instructure.com>
QA-Review: Steven Burnett <sburnett@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>