update migration to delete tags without restrictor
the context_module_importer will remove all tags for associating
items with a destroy_all call. this signals the restrictor to mark
them as "manually_deleted", which will then never allow the
state of a tag to be changed by a migration because it thinks
that a user overridden that tags state.

note that this will not fix the problem if the tag was already
created and then removed by the migration. we would have to
do a data fix to remove the bad "manually_deleted" tags on
content.

flag = none

fixes: ADMIN-2716

Test-Plan:
* create a master course and a child course
* in master course, create 2 modules and an assignment in each
* run a sync and check the child course
* in master course, move assignment from second module into first
* in master course, delete the empty module
* in master course, update the moved assignment
* run a sync and make sure everything looks the same in child

Change-Id: I7a7df96a333d72055bdd7852d5abe65727d5b983
Reviewed-on: https://gerrit.instructure.com/204478
QA-Review: Anju Reddy <areddy@instructure.com>
Tested-by: Jenkins
Product-Review: Rex Fleischer <rfleischer@instructure.com>
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
Reviewed-by: James Williams <jamesw@instructure.com>