Re-render post/hide menu on updating submission
Make sure we re-render the "post/hide grades" menu in SpeedGrader when a
submission is graded so the post/hide options are enabled according to
the current state of the submissions (as opposed to their state on
loading the page).

fixes GRADE-2396

Test plan:

Note that this does *not* work for comments at present (even when they
should cause a submission to post) because we specifically only check
for graded submissions when deciding whether a submission is postable.

- Have a course with New Gradebook and Post Policies enabled
- Create an auto-post assignment and open SpeedGrader
  - Observe that the "Hide Grades" option in the eye menu is disabled
  - Issue some grades
  - The "Hide Grades" option should now be enabled and should work as
    expected
- Create a manual-post assignment and open SpeedGrader
  - Observe that the "Post Grades" option in the eye menu is disabled
  - Issue some grades
  - The "Post Grades" option should now be enabled and should work as
    expected

- Have a course with OLD Gradebook
  - Smoke test giving some grades in SpeedGrader
  - Nothing should be affected

Change-Id: I29cd99b32212518d08ede9f160204d6bcae8cd9e
Reviewed-on: https://gerrit.instructure.com/207083
Tested-by: Jenkins
Reviewed-by: Jeremy Neander <jneander@instructure.com>
Reviewed-by: Gary Mei <gmei@instructure.com>
QA-Review: Gary Mei <gmei@instructure.com>
Product-Review: Keith Garner <kgarner@instructure.com>