fix apollo not updating submission drafts in a2
after the recent apollo upgrade we are seeing an issue where
writing to the apollo cache does not take effect and changes
are not visible until the page is refreshed.

we are stringifying data we update into the cache to ensure that
apollo recognizes it as having changed, instead of considering it
to be the same object.

Test Plan:
* as a student in a2, navigate to a file upload assignment
* attempt to add a new file to the submission draft
** you should see the page render with the newly added file
* attempt to remove a file from the submission draft
** you should see the page render with the file removed

flag=assignments_2
fixes COMMS-2399

Change-Id: I6d4e274ac145539e8af7f12dd4ed2e00e5051131
Reviewed-on: https://gerrit.instructure.com/207398
Reviewed-by: Steven Burnett <sburnett@instructure.com>
Tested-by: Jenkins
QA-Review: Steven Burnett <sburnett@instructure.com>
Product-Review: Steven Burnett <sburnett@instructure.com>