spec: [ci-coverage] use dialog reference for queries
This spec has been flaky. The theory is that it might not be acting on
the correct DOM elements, due to DOM pollution. The result is that it
ends up waiting on behavior that was never triggered, and times out.

Change-Id: I107aac0e9fe8185dcdf81137a71ca5a0e75ee7f9
Reviewed-on: https://gerrit.instructure.com/208056
Tested-by: Jenkins
Reviewed-by: Gary Mei <gmei@instructure.com>
QA-Review: Jeremy Neander <jneander@instructure.com>
Product-Review: Jeremy Neander <jneander@instructure.com>