Fixing another npm package that was casuing npm install errors, jsdom.
The errors didn't cause the build to fail, but they made it really hard to track down real errors
when it does!
It was accidentally shrinkwrapped to a version of readable-stream@3.4.0 which needs a newer version of node.
Downgrading it to the version which works with our version of node (by copying that shrinkwrap fragment)

TESTING:
Need to do a Capistrano deploy to staging and see that the error is done
and canvas_i18nliner gem builds. Can't really test well in the dev env
right now.