Go back to using a separate Qualtrics mailiing list per Sync To LMS
Here's why. The logic for this whole thing is 1) take all new students
who need an embedded Qualtrics survey 2) create a Qualtrics mailing list
with them 3) use the mailing list to create a Distribution list 4) the
Distribution list is what creates the survey links where there is one
per student and it's unique, tied to that student.

Ideally, we would have one big Mailing List for the course and one big
Distribution list. However, the API doesn't have the ability to update a
Distribution list, only create one and I fear that if we create a new
Distribution list each time we update the single Mailing list, the
survey links already created will become invalid or people may have
started the survey already and now they are in two Distrubution lists.

SOOO, we're just going to be safe and make sure that each Sync TO LMS
creates a unique Mailing List and a unique Distribution list only for
students who haven't had the survey link generated for them. So multiple
syncs makes it a bit ugly b/c people are spread out between lists, but
you can always just go to the survey itselft and export all the survey
data instead of using MLs and DLs directly.

TESTING: too hard in the dev env. Going to test in staging. However,
this just goes back to the old logic which is already tested, so it's
pretty safe.