Fix SSO login in dev env.
This also relies on the Admin server dev DB creation making the links
point to ssoweb:3002 (using the port number). Currently, with every app
in it's own docker network, we can't use nginx to abstract away the
ports when the containers are calling each other from inside the
container (as opposed to when they redirect in the web browser).