Have all our Docker dev envs on the same network.
This is so that service names resolve as hostnames and we don't
have to reference each other's networks from each app's docker-compose
file. E.g. inside canvasweb you can hit http://ssoweb:3002 and it will
resolve if they are on the same network.

Depends on PRs from each of our apps that works with this docker
container. Right now that's canvas, canvas-js-css, rubycas-server, and join (aka
beyondz-platform).