Support editing files outside the container and having the changes take effect inside.
The purpose of this is to make our dev env feel more like you're just
doing it on your local machine and just using the Docker container as
the web-server to make it show up in your local browser.

This is done by mounting specific volumes. See docker-compose.yml BIG
comment for why we mount specific ones and not the whole app root. Note
that by mounting the config directory, it constantly makes it seem like
there are local changes to config files b/c our configs are not all
setup yet to pull from ENV vars. For now, just .gitignore them again
since they are ignored anyway during a Capistrano deploy but once we
move to Heroku or just cause, we should update all configs to use ENV
vars and check that in so that just building the container doesn't
create changes in src control.

TESTING:
- rebuilt the containers to make sure nothing was masked/hidden by
  mounting the volumes.
- tested making a change to an html.erb file from outside the container
  and seeing the change take effect in the browser