Remove checked in config files for Heroku.
When we move to Heroku, the configs need to be checked in and
read from ENV vars. But for now, just remove these which I started doing
that for b/c when using the local Docker dev env, they keep showing up
as modified when you restart the container. We need the
docker-compose/config files to match the config/*.yml files to avoid that.
Actually, once they are the same we can just get rid of docker-compose/config
and just set the environment variables.