do not display grade for attempt if ungraded in a2
the mid-right grading indicator for students is intended to show
the grade and submission details for the current attempt. we are
currently displaying the latest grade, even if it is for a
previous attempt, when the current attempt is ungraded. we should
leave this blank if the current attempt has no grade to display

Test Plan:
* as a teacher, create a graded assignment of any kind in a2
* as a student, submit to the assignment
* as a teacher, grade the assignment
* as a student, create a new attempt and re-submit the assignment
* for the submitted but ungraded attempt, the grade display on the
  mid-right of the page should show whatever the default for a
  blank grade for that grade type is
** for points this is something like '–/10 Points'
* if you view the previous submission, you should see the grade
  for that attempt
* the submission times for each attempt should correspond to the
  time they were submitted

flag=assignments_2
fixes COMMS-2446

Change-Id: Ie5d48ef9d338a7181864ce0cf844f450324bf633
Reviewed-on: https://gerrit.instructure.com/209615
Reviewed-by: Landon Gilbert-Bland <lbland@instructure.com>
Tested-by: Jenkins
QA-Review: Landon Gilbert-Bland <lbland@instructure.com>
Product-Review: Landon Gilbert-Bland <lbland@instructure.com>