Remove old docker dev env configs that weren't used
This was causing a 45 second delay on restart waiting for 3, 15s
timeouts trying to hit the consul server. We don't use consul for
dynamic settings, so remove that... along with cassandra and an outdated
delayed_jobs config and a selenium config that I never actually figured
out how to use or if it works.

TESTING:
- Rebuilt the docker dev env and logged into http://canvasweb. It didn't
  take 45 seconds for it to become available (with bad gateways until
the timeout elapsed)!