Don't call rake db:reset_encryption_key_hash anymore in dev env.
This was required when we pulling a staging/prod DB into a dev DB and
didn't adjust the value of this:
   select value from settings where name = 'encryption_key_hash';
Now the admin server adjusts that value to match what's in security.yml.
When we did call this rake task, it would make things generally work
(you just had to login again) but it invalidated the access_tokens that
other apps used to call into the Canvas API. Now we have access tokens
generated with the current value of 'encryption_key_hash' that matches
security.yml.

TESTING:
- Ran the admin server scripts to pull prod, turn it into staging DB,
  and then turn that into dev DB. Then ran
docker-compose/scripts/dbrefresh.sh and verified that the kits server
could connect to the Canvas API to pull attendance info.