Don't scroll the whole page to the top constantly
Task: https://app.asana.com/0/1128645836596228/1138467255207468/f

This is a super weird one. I'm not sure what this behavior is supposed
to be, but for some reason there's a timer that scrolls to the top of
the page every 2.5 seconds if you've opened a rubric on the grades page.

This fix just comments out the scroll command and leaves the rest, but
frankly I have no idea what any of this is for, and this is digging into
Canvas code. There's probably a better way to solve this.