Allow clicking on unread badge to mark as read
Closes ADMIN-2918

This makes the blue unread dots in the ReceivedTable into
buttons that can be clicked to mark the share as read.

Note that the unread badge in the left nav will not immediately
reflect the unread-marking, but that badge should update in
less than a minute when the next poll happens.

Test plan:
* Have a user with at least one unread content share
* Navigate to the "Shared Content" page
* The unread shares should still have their blue dot
* Mark one of them as read by clicking on the dot
* The dot should disappear and the share should now
  have an "read_state" of "read"
* Within one minute, the badge on the left nav should
  decrement by one, reflecting one fewer unread share
* Reload the front end
* The share you clicked on should still show as read

Change-Id: Id57e6240fee134b7022765eb355fff9244bfaddf
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/214666
Tested-by: Jenkins
Reviewed-by: Jon Willesen <jonw+gerrit@instructure.com>
QA-Review: Jon Willesen <jonw+gerrit@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>