Fix log file permission errors during Capistrano deploy.
We have logrotate configured to run weekly on the servers, but Rails.logger also
has a built in logrotate functionality. The default is to rotate the logs
when they hit 1MB. Since the server is running as a particular user and does
a File.rename, it's up to the kernel to set the file permissions based on the umask.
See the calls to create_logfile here: https://github.com/ruby/ruby/blob/v2_1_9/lib/logger.rb

This fix alters the behavior of create_logfile and uses 664 permissions so that
users in the group can still write it. HOWEVER, to make that work you need to make sure
new files created in the log directory inherit the group id. Do this by setting the setgid
bit on the directory. E.g. sudo chmod g+s /our/log/dir and make sure the directory itself
is in the proper group. IMPORTANT: do that before releasing this!

TESTING:
- I changed the line in lib/canvas_logger.rb from:
    @logdev = GroupWritableLogDevice.new(log_path)
  to
    @logdev = GroupWritableLogDevice.new(log_path, shift_size: 1024)
- This makes the log roll every 1KB.
- I then set the setgid bit on the log dir.
- I then changed the group permission on the log file and the log directory itself
  to be something other than what the server was running as.
- Then I hit the app and watched the files rotate with the permissions intact.