Add support for loading ENV variables into the Rails app using .env files
We're prepping to move the Portal from an AWS EC2 instance to Heroku.
Everything in Heroku must be checked in to src ctrl and anything server or
environment specific or secret must be set through environment
variables. We need to cut-over all our config files to read from ENV vars instead
of having the values hardcoded in local config files that are not checked in.

This pull request adds the dotenv gem to help with that. Consider that we need
the ENV vars available in these scenarios:
- in the process spawend by Apache + Passenger
- in the process that Capistrano uses to deploy (since it runs rake tasks)
- in the process spawned from the shell where you run rake tasks or load the rails console
- from a cronjob in the crontab environment

We could write a wrapper script to set the ENV vars and make sure the script is run in
all these scenarios. This would mean calling the script from Apache/Passenger,
.bashrc (or .bash_profile), in the Capistrano deploy script, and from crontab (may actually
have to hardcode for that one).

Instead, let's just let Rails load the ENV vars into it's environment using a .env file and the
dotenv gem so we don't have to deal with all that. For more info see
https://github.com/bkeepers/dotenv (especially the "dotenv in production" section):

The way we'll do that is for the Docker dev env, have all the dev values checked into
docker-compose/.env but for staging and prod we'll put those ENV values in a .env.local
file that Capistrano will copy to the app root on deploy.

Here are a couple articles to read:
- https://medium.com/@rdsubhas/ruby-in-production-lessons-learned-36d7ab726d99
- https://www.honeybadger.io/blog/ruby-guide-environment-variables/
- https://coderwall.com/p/djrfra/pass-the-database-pwd-to-rails-with-an-environment-variable
- https://www.phusionpassenger.com/library/indepth/environment_variables.html

I cutover the database connection and domain configs to read from environmnet
variables as a proof of concept in order to be able to test the scenarios below since
I assume that whatever runs would need to access the database!

TESTING:
- do a Capistrano deploy and make sure the ENV var values are set
- run a cronjob and make sure they're set
- run a rake task from the shell and make sure they're set
- load the rails console and make sure they're set
- make sure they're set in the app itself
but in the meantime, we also need the servers as currently deployed able to
provide the ENV vars to the config files.